+++ paper.bib
@@ -71,23 +71,23 @@
   keywords = {libraries, Hessenberg, tridiagonal, bidiagonal, high performance, reduction, Linear algebra}
 }
 @article{cholmod,
-author = {Chen, Yanqing and Davis, Timothy A. and Hager, William W. and Rajamanickam, Sivasankaran},
-title = {Algorithm 887: CHOLMOD, Supernodal Sparse Cholesky Factorization and Update/Downdate},
-year = {2008},
-issue_date = {October 2008},
-publisher = {Association for Computing Machinery},
-address = {New York, NY, USA},
-volume = {35},
-number = {3},
-issn = {0098-3500},
-url = {https://doi.org/10.1145/1391989.1391995},
-doi = {10.1145/1391989.1391995},
-abstract = {CHOLMOD is a set of routines for factorizing sparse symmetric positive definite matrices of the form A or AAT, updating/downdating a sparse Cholesky factorization, solving linear systems, updating/downdating the solution to the triangular system Lx = b, and many other sparse matrix functions for both symmetric and unsymmetric matrices. Its supernodal Cholesky factorization relies on LAPACK and the Level-3 BLAS, and obtains a substantial fraction of the peak performance of the BLAS. Both real and complex matrices are supported. CHOLMOD is written in ANSI/ISO C, with both C and MATLABTM interfaces. It appears in MATLAB 7.2 as x = Ab when A is sparse symmetric positive definite, as well as in several other sparse matrix functions.},
-journal = {ACM Trans. Math. Softw.},
-month = {oct},
-articleno = {22},
-numpages = {14},
-keywords = {sparse matrices, linear equations, Cholesky factorization}
+  author = {Chen, Yanqing and Davis, Timothy A. and Hager, William W. and Rajamanickam, Sivasankaran},
+  title = {Algorithm 887: CHOLMOD, Supernodal Sparse Cholesky Factorization and Update/Downdate},
+  year = {2008},
+  issue_date = {October 2008},
+  publisher = {Association for Computing Machinery},
+  address = {New York, NY, USA},
+  volume = {35},
+  number = {3},
+  issn = {0098-3500},
+  url = {https://doi.org/10.1145/1391989.1391995},
+  doi = {10.1145/1391989.1391995},
+  abstract = {CHOLMOD is a set of routines for factorizing sparse symmetric positive definite matrices of the form A or AAT, updating/downdating a sparse Cholesky factorization, solving linear systems, updating/downdating the solution to the triangular system Lx = b, and many other sparse matrix functions for both symmetric and unsymmetric matrices. Its supernodal Cholesky factorization relies on LAPACK and the Level-3 BLAS, and obtains a substantial fraction of the peak performance of the BLAS. Both real and complex matrices are supported. CHOLMOD is written in ANSI/ISO C, with both C and MATLABTM interfaces. It appears in MATLAB 7.2 as x = Ab when A is sparse symmetric positive definite, as well as in several other sparse matrix functions.},
+  journal = {ACM Trans. Math. Softw.},
+  month = {oct},
+  articleno = {22},
+  numpages = {14},
+  keywords = {sparse matrices, linear equations, Cholesky factorization}
 }
 @book{chandra2001parallel,
   title={Parallel programming in OpenMP},
@@ -105,6 +105,7 @@ keywords = {sparse matrices, linear equations, Cholesky factorization}
   volume = {23},
   number = {4},
   issn = {1937-4771},
+  doi = {10.1016/b978-0-12-803761-4.00011-3},
[…]
