### 2023-11-14 #6059 @sarah-ek
Thanks for the heads-up!
I added the DOI for Intel TBB, and added an affiliation to the metadata.
The Rayon DOI doesn't seem correct. I followed the suggestion of the Rayon devs mentioned here for the citation https://github.com/rayon-rs/rayon/issues/984
I'll see if it compiles now
