+++ faer-bench/src/partial_piv_lu.rs
@@ -1,7 +1,7 @@
 use super::timeit;
 use crate::random;
-use dyn_stack::{PodStack, GlobalPodBuffer, ReborrowMut};
-use faer_core::{Mat, Parallelism};
+use dyn_stack::{GlobalPodBuffer, PodStack, ReborrowMut};
+use faer_core::{unzipped, zipped, Mat, Parallelism};
 use ndarray_linalg::solve::Factorize;
 use std::time::Duration;
 
@@ -64,11 +64,11 @@ pub fn faer<T: faer_core::ComplexField>(
                 }
             }
             let mut lu = Mat::<T>::zeros(n, n);
-            let mut row_fwd = vec![0; n];
-            let mut row_inv = vec![0; n];
+            let mut row_fwd = vec![0u32; n];
+            let mut row_inv = vec![0u32; n];
 
             let mut mem = GlobalPodBuffer::new(
-                faer_lu::partial_pivoting::compute::lu_in_place_req::<T>(
+                faer_lu::partial_pivoting::compute::lu_in_place_req::<u32, T>(
                     n,
                     n,
                     parallelism,
@@ -79,10 +79,8 @@ pub fn faer<T: faer_core::ComplexField>(
             let mut stack = PodStack::new(&mut mem);
 
             let mut block = || {
-                lu.as_mut()
-                    .cwise()
-                    .zip(c.as_ref())
-                    .for_each(|mut dst, src| dst.write(src.read()));
+                zipped!(lu.as_mut(), c.as_ref())
+                    .for_each(|unzipped!(mut dst, src)| dst.write(src.read()));
                 faer_lu::partial_pivoting::compute::lu_in_place(
                     lu.as_mut(),
                     &mut row_fwd,
