+++ faer-bench/src/cholesky.rs
@@ -1,6 +1,6 @@
 use super::timeit;
-use dyn_stack::{PodStack, GlobalPodBuffer, ReborrowMut};
-use faer_core::{Mat, Parallelism};
+use dyn_stack::{GlobalPodBuffer, PodStack, ReborrowMut};
+use faer_core::{unzipped, zipped, Mat, Parallelism};
 use ndarray_linalg::Cholesky;
 use std::time::Duration;
 
@@ -69,10 +69,8 @@ pub fn faer<T: faer_core::ComplexField>(
             let mut stack = PodStack::new(&mut mem);
 
             let time = timeit(|| {
-                chol.as_mut()
-                    .cwise()
-                    .zip(c.as_ref())
-                    .for_each(|mut dst, src| dst.write(src.read()));
+                zipped!(chol.as_mut(), c.as_ref())
+                    .for_each(|unzipped!(mut dst, src)| dst.write(src.read()));
                 faer_cholesky::llt::compute::cholesky_in_place(
                     chol.as_mut(),
                     Default::default(),

[…]
