### 2023-11-17 #6059 @jedbrown
@ashwinvis @mscroggs @JPenuchot :wave: Would you be available to review this submission for JOSS?
