### 2023-11-17 b48565e
inline(always) workaround for target_feature bug
