### 2023-11-26 b298dba
impl sparse qr solve in place
