+++ faer-libs/faer-sparse/src/qr.rs
@@ -56,6 +56,7 @@ pub(crate) fn ghost_col_etree<'m, 'n, I: Index>(
     }
 }
 
+/// workspace: I×(A.nrows() + A.ncols())
 #[inline]
 pub fn col_etree<'out, I: Index>(
     A: SymbolicSparseColMatRef<'_, I>,
@@ -110,7 +111,8 @@ pub(crate) fn ghost_least_common_ancestor<'n, I: Index>(
     *lca as isize
 }
 
-pub(crate) fn ghost_column_counts_aat<'m, 'n, I: Index>(
+/// workspace: I×(5×A.nrows() + A.ncols())
+pub fn ghost_column_counts_aat<'m, 'n, I: Index>(
     col_counts: &mut Array<'m, I>,
     min_row: &mut Array<'n, I::Signed>,
     A: ghost::SymbolicSparseColMatRef<'m, 'n, '_, I>,
@@ -278,6 +280,7 @@ pub fn postorder<I: Index>(post: &mut [I], etree: EliminationTreeRef<'_, I>, sta
 pub mod supernodal {
     use core::iter::zip;
 
+    use dyn_stack::{SizeOverflow, StackReq};
     use faer_core::{
         group_helpers::{SliceGroup, SliceGroupMut},
         householder::apply_block_householder_sequence_transpose_on_the_left_in_place_with_conj,
@@ -303,6 +306,8 @@ pub mod supernodal {
         col_ptrs_for_tau_values: alloc::vec::Vec<I>,
         col_ptrs_for_values: alloc::vec::Vec<I>,
         super_etree: alloc::vec::Vec<I>,
+        max_blocksize: alloc::vec::Vec<I>,
+        nrows: usize,
     }
 
     impl<I: Index> SymbolicSupernodalHouseholder<I> {
@@ -351,9 +356,9 @@ pub mod supernodal {
         child_next: alloc::vec::Vec<I>,
     }
 
-    impl<I> SymbolicSupernodalQr<I> {
+    impl<I: Index> SymbolicSupernodalQr<I> {
         #[inline]
-        pub fn r_transposed(&self) -> &SymbolicSupernodalCholesky<I> {
+        pub fn r_adjoint(&self) -> &SymbolicSupernodalCholesky<I> {
             &self.L
         }
 
@@ -361,6 +366,40 @@ pub mod supernodal {
         pub fn householder(&self) -> &SymbolicSupernodalHouseholder<I> {
             &self.H
         }
+
+        pub fn solve_in_place_req<E: Entity>(
+            &self,
+            rhs_ncols: usize,
+            parallelism: Parallelism,
+        ) -> Result<StackReq, SizeOverflow> {
+            let _ = parallelism;
+            let L_symbolic = self.r_adjoint();
+            let H_symbolic = self.householder();
+            let n_supernodes = L_symbolic.n_supernodes();
+
+            let init_req = faer_core::temp_mat_req::<E>(self.householder().nrows, rhs_ncols)?;
+            let mut loop_req = StackReq::empty();
+            for s in 0..n_supernodes {
+                let s_h_row_begin = H_symbolic.col_ptrs_for_row_indices[s].zx();
+                let s_h_row_full_end = H_symbolic.col_ptrs_for_row_indices[s + 1].zx();
+                let max_blocksize = H_symbolic.max_blocksize[s].zx();
+
+                loop_req = loop_req.try_or(faer_core::householder::apply_block_householder_sequence_transpose_on_the_left_in_place_req::<E>(s_h_row_full_end - s_h_row_begin, max_blocksize, rhs_ncols)?)?;
+            }
+
+            init_req.try_and(loop_req)
+        }
+    }
+
[…]
