### 2023-11-28 23dbb3c
impl simplicial qr
