+++ faer-libs/faer-sparse/src/qr.rs
@@ -1,15 +1,24 @@
 use crate::{
-    cholesky::{ghost_postorder, simplicial::EliminationTreeRef},
+    cholesky::{
+        ghost_postorder,
+        simplicial::EliminationTreeRef,
+        supernodal::{
+            CholeskySymbolicSupernodalParams, SupernodalLltRef, SymbolicSupernodalCholesky,
+        },
+    },
     ghost::{self, Array, Idx, MaybeIdx},
     mem::{self, NONE},
-    Index,
+    try_zeroed, FaerError, Index,
 };
-use dyn_stack::PodStack;
+use core::iter::zip;
+use dyn_stack::{PodStack, SizeOverflow, StackReq};
 use faer_core::{
     constrained::Size,
+    group_helpers::{SliceGroup, SliceGroupMut},
+    householder::apply_block_householder_sequence_transpose_on_the_left_in_place_with_conj,
     permutation::{PermutationRef, SignedIndex},
-    sparse::SymbolicSparseColMatRef,
-    unzipped,
+    sparse::{SparseColMatRef, SymbolicSparseColMatRef},
+    unzipped, zipped, Conj, MatMut, Parallelism,
 };
 use faer_entity::*;
 use reborrow::*;
@@ -111,8 +120,7 @@ pub(crate) fn ghost_least_common_ancestor<'n, I: Index>(
     *lca as isize
 }
 
-/// workspace: I×(5×A.nrows() + A.ncols())
-pub fn ghost_column_counts_aat<'m, 'n, I: Index>(
+pub(crate) fn ghost_column_counts_aat<'m, 'n, I: Index>(
     col_counts: &mut Array<'m, I>,
     min_row: &mut Array<'n, I::Signed>,
     A: ghost::SymbolicSparseColMatRef<'m, 'n, '_, I>,
@@ -189,16 +197,16 @@ pub fn ghost_column_counts_aat<'m, 'n, I: Index>(
     };
 
     for k in M.indices() {
-        let mut k_ = post[k].zx();
-        delta[k_] = I::truncate(if first[k_].idx().is_none() { 1 } else { 0 });
+        let mut pk = post[k].zx();
+        delta[pk] = I::truncate(if first[pk].idx().is_none() { 1 } else { 0 });
         loop {
-            if first[k_].idx().is_some() {
+            if first[pk].idx().is_some() {
                 break;
             }
 
-            first[k_] = MaybeIdx::from_index(k.truncate());
-            if let Some(parent) = etree[k_].idx() {
-                k_ = parent.zx();
+            first[pk] = MaybeIdx::from_index(k.truncate());
+            if let Some(parent) = etree[pk].idx() {
+                pk = parent.zx();
             } else {
                 break;
             }
@@ -206,11 +214,10 @@ pub fn ghost_column_counts_aat<'m, 'n, I: Index>(
     }
 
     for k in M.indices() {
-        let k_ = post[k].zx();
+        let pk = post[k].zx();
 
-        if let Some(parent) = etree[k_].idx() {
-            let parent = parent.zx();
-            decr(&mut delta[parent]);
+        if let Some(parent) = etree[pk].idx() {
+            decr(&mut delta[parent.zx()]);
         }
 
         let head_k = &mut head[k];
@@ -223,10 +230,10 @@ pub fn ghost_column_counts_aat<'m, 'n, I: Index>(
                     .map(|perm| perm.into_arrays().1[i].zx())
                     .unwrap_or(i);
                 let lca =
-                    ghost_least_common_ancestor::<I>(i, k_, first, max_first, prev_leaf, ancestor);
+                    ghost_least_common_ancestor::<I>(i, pk, first, max_first, prev_leaf, ancestor);
 
[…]
