### 2023-11-29 3af1a53
simplicial lu and qr wrapper
