+++ faer-libs/faer-sparse/src/triangular_solve.rs
@@ -1,0 +1,920 @@
+use core::iter::zip;
+use faer_core::{
+    assert, group_helpers::SliceGroup, permutation::Index, sparse::SparseColMatRef, Conj, MatMut,
+    Parallelism,
+};
+use faer_entity::ComplexField;
+use reborrow::*;
+
+pub fn solve_lower_triangular_in_place<I: Index, E: ComplexField>(
+    l: SparseColMatRef<'_, I, E>,
+    conj: Conj,
+    rhs: MatMut<'_, E>,
+    parallelism: Parallelism,
+) {
+    let _ = parallelism;
+    assert!(l.nrows() == l.ncols());
+    assert!(rhs.nrows() == l.nrows());
+
+    let slice_group = SliceGroup::<'_, E>::new;
+
+    let n = l.ncols();
+    let mut x = rhs;
+    for mut x in x.rb_mut().col_chunks_mut(4) {
+        match x.ncols() {
+            1 => {
+                for j in 0..n {
+                    let d = slice_group(l.values_of_col(j)).read(0).faer_inv();
+                    let d = if conj == Conj::Yes { d.faer_conj() } else { d };
+
+                    let xj0 = x.read(j, 0).faer_mul(d);
+                    x.write(j, 0, xj0);
+
+                    for (i, lij) in zip(
+                        l.row_indices_of_col(j),
+                        slice_group(l.values_of_col(j)).into_ref_iter(),
+                    )
+                    .skip(1)
+                    {
+                        let lij = lij.read();
+                        let lij = if conj == Conj::Yes {
+                            lij.faer_conj()
+                        } else {
+                            lij
+                        };
+                        x.write(i, 0, x.read(i, 0).faer_sub(lij.faer_mul(xj0)));
+                    }
+                }
+            }
+            2 => {
+                for j in 0..n {
+                    let d = slice_group(l.values_of_col(j)).read(0).faer_inv();
+                    let d = if conj == Conj::Yes { d.faer_conj() } else { d };
+
+                    let xj0 = x.read(j, 0).faer_mul(d);
+                    x.write(j, 0, xj0);
+                    let xj1 = x.read(j, 1).faer_mul(d);
+                    x.write(j, 1, xj1);
+
+                    for (i, lij) in zip(
+                        l.row_indices_of_col(j),
+                        slice_group(l.values_of_col(j)).into_ref_iter(),
+                    )
+                    .skip(1)
+                    {
+                        let lij = lij.read();
+                        let lij = if conj == Conj::Yes {
+                            lij.faer_conj()
+                        } else {
+                            lij
+                        };
+                        x.write(i, 0, x.read(i, 0).faer_sub(lij.faer_mul(xj0)));
+                        x.write(i, 1, x.read(i, 1).faer_sub(lij.faer_mul(xj1)));
+                    }
+                }
+            }
+            3 => {
+                for j in 0..n {
+                    let d = slice_group(l.values_of_col(j)).read(0).faer_inv();
+                    let d = if conj == Conj::Yes { d.faer_conj() } else { d };
+
+                    let xj0 = x.read(j, 0).faer_mul(d);
+                    x.write(j, 0, xj0);
+                    let xj1 = x.read(j, 1).faer_mul(d);
+                    x.write(j, 1, xj1);
+                    let xj2 = x.read(j, 2).faer_mul(d);
+                    x.write(j, 2, xj2);
+
+                    for (i, lij) in zip(
+                        l.row_indices_of_col(j),
+                        slice_group(l.values_of_col(j)).into_ref_iter(),
+                    )
+                    .skip(1)
+                    {
+                        let lij = lij.read();
+                        let lij = if conj == Conj::Yes {
+                            lij.faer_conj()
+                        } else {
+                            lij
+                        };
[…]
