+++ faer-libs/faer-sparse/src/lu.rs
@@ -1,11 +1,11 @@
 use crate::{
-    cholesky::{simplicial::EliminationTreeRef, supernodal::CholeskySymbolicSupernodalParams},
+    cholesky::simplicial::EliminationTreeRef,
     ghost,
     ghost::{Array, MaybeIdx},
     mem::{
         NONE, {self},
     },
-    try_zeroed, FaerError, Index, SymbolicSparseColMatRef,
+    try_zeroed, FaerError, Index, SymbolicSparseColMatRef, SymbolicSupernodalParams,
 };
 use core::{iter::zip, mem::MaybeUninit};
 use dyn_stack::{PodStack, SizeOverflow, StackReq};
@@ -14,106 +14,107 @@ use faer_core::{
     group_helpers::{SliceGroup, SliceGroupMut},
     mul,
     permutation::{PermutationRef, SignedIndex},
+    solve,
     sparse::SparseColMatRef,
     Conj, MatMut, Parallelism,
 };
 use faer_entity::*;
 use reborrow::*;
 
-pub mod supernodal {
-    use super::*;
-    use faer_core::{assert, solve};
+#[inline(never)]
+fn resize_scalar<E: Entity>(
+    v: &mut GroupFor<E, alloc::vec::Vec<E::Unit>>,
+    n: usize,
+    exact: bool,
+    reserve_only: bool,
+) -> Result<(), FaerError> {
+    let mut failed = false;
+    let reserve = if exact {
+        alloc::vec::Vec::try_reserve_exact
+    } else {
+        alloc::vec::Vec::try_reserve
+    };
 
-    #[inline(never)]
-    fn resize_scalar<E: Entity>(
-        v: &mut GroupFor<E, alloc::vec::Vec<E::Unit>>,
-        n: usize,
-        exact: bool,
-        reserve_only: bool,
-    ) -> Result<(), FaerError> {
-        let mut failed = false;
-        let reserve = if exact {
-            alloc::vec::Vec::try_reserve_exact
-        } else {
-            alloc::vec::Vec::try_reserve
-        };
-
-        E::faer_map(E::faer_as_mut(v), |v| {
-            if !failed {
-                failed = reserve(v, n.saturating_sub(v.len())).is_err();
-                if !reserve_only {
-                    v.resize(Ord::max(n, v.len()), unsafe { core::mem::zeroed() });
-                }
+    E::faer_map(E::faer_as_mut(v), |v| {
+        if !failed {
+            failed = reserve(v, n.saturating_sub(v.len())).is_err();
+            if !reserve_only {
+                v.resize(Ord::max(n, v.len()), unsafe { core::mem::zeroed() });
             }
-        });
-        if failed {
-            Err(FaerError::OutOfMemory)
-        } else {
-            Ok(())
         }
-    }
-
-    #[inline(never)]
-    fn resize_maybe_uninit_scalar<E: Entity>(
-        v: &mut GroupFor<E, alloc::vec::Vec<MaybeUninit<E::Unit>>>,
-        n: usize,
-    ) -> Result<(), FaerError> {
-        let mut failed = false;
-        E::faer_map(E::faer_as_mut(v), |v| {
-            if !failed {
-                failed = v.try_reserve(n.saturating_sub(v.len())).is_err();
-                unsafe { v.set_len(n) };
-            }
-        });
-        if failed {
-            Err(FaerError::OutOfMemory)
-        } else {
-            Ok(())
-        }
-    }
-
-    #[inline(never)]
-    fn resize_index<I: Index>(
-        v: &mut alloc::vec::Vec<I>,
-        n: usize,
-        exact: bool,
-        reserve_only: bool,
-    ) -> Result<(), FaerError> {
-        let reserve = if exact {
-            alloc::vec::Vec::try_reserve_exact
-        } else {
[…]
