+++ faer-libs/faer-sparse/src/cholesky.rs
@@ -287,7 +287,7 @@ pub mod simplicial {
                                 A.row_indices_of_col(k),
                                 SliceGroup::<'_, E>::new(A.values_of_col(k)).into_ref_iter(),
                             ) {
-                                x.write(i, aik.read().faer_conj());
+                                x.write(i, x.read(i).faer_add(aik.read().faer_conj()));
                             }
 
                             let mut d = x.read(k).faer_real();
@@ -451,7 +451,7 @@ pub mod simplicial {
                                 A.row_indices_of_col(k),
                                 SliceGroup::<'_, E>::new(A.values_of_col(k)).into_ref_iter(),
                             ) {
-                                x.write(i, aik.read().faer_conj());
+                                x.write(i, x.read(i).faer_add(aik.read().faer_conj()));
                             }
 
                             let mut d = x.read(k).faer_real();
@@ -783,7 +783,7 @@ pub mod simplicial {
             }
         }
 
-        pub fn dense_solve_in_place_req<E: Entity>(
+        pub fn solve_in_place_req<E: Entity>(
             &self,
             rhs_ncols: usize,
         ) -> Result<StackReq, SizeOverflow> {
@@ -1498,7 +1498,7 @@ pub mod supernodal {
             supernodal::SymbolicSupernodeRef { start, pattern }
         }
 
-        pub fn dense_solve_in_place_req<E: Entity>(
+        pub fn solve_in_place_req<E: Entity>(
             &self,
             rhs_ncols: usize,
         ) -> Result<StackReq, SizeOverflow> {
@@ -2322,11 +2322,12 @@ pub mod supernodal {
                     slice_group(A_lower.values_of_col(j)).into_ref_iter(),
                 ) {
                     let val = val.read();
-                    if i >= s_end {
-                        Ls.write(global_to_local[i].sx(), j_shifted, val);
-                    } else if i >= j {
-                        Ls.write(i - s_start, j_shifted, val);
-                    }
+                    let (ix, iy) = if i >= s_end {
+                        (global_to_local[i].sx(), j_shifted)
+                    } else {
+                        (i - s_start, j_shifted)
+                    };
+                    Ls.write(ix, iy, Ls.read(ix, iy).faer_add(val));
                 }
             }
 
@@ -2512,11 +2513,12 @@ pub mod supernodal {
                     slice_group(A_lower.values_of_col(j)).into_ref_iter(),
                 ) {
                     let val = val.read();
-                    if i >= s_end {
-                        Ls.write(global_to_local[i].sx(), j_shifted, val);
-                    } else if i >= j {
-                        Ls.write(i - s_start, j_shifted, val);
-                    }
+                    let (ix, iy) = if i >= s_end {
+                        (global_to_local[i].sx(), j_shifted)
+                    } else {
+                        (i - s_start, j_shifted)
+                    };
+                    Ls.write(ix, iy, Ls.read(ix, iy).faer_add(val));
                 }
             }
 
@@ -2742,11 +2744,12 @@ pub mod supernodal {
                     slice_group(A_lower.values_of_col(j)).into_ref_iter(),
                 ) {
                     let val = val.read();
-                    if i >= s_end {
-                        Ls.write(global_to_local[i].sx(), j_shifted, val);
-                    } else if i >= j {
-                        Ls.write(i - s_start, j_shifted, val);
-                    }
+                    let (ix, iy) = if i >= s_end {
+                        (global_to_local[i].sx(), j_shifted)
+                    } else {
+                        (i - s_start, j_shifted)
+                    };
[…]
