+++ faer-libs/faer-sparse/src/lu.rs
@@ -1,11 +1,13 @@
 use crate::{
     cholesky::simplicial::EliminationTreeRef,
+    colamd::Control,
     ghost,
-    ghost::{Array, MaybeIdx},
+    ghost::{Array, Idx, MaybeIdx},
     mem::{
         NONE, {self},
     },
-    try_zeroed, FaerError, Index, SymbolicSparseColMatRef, SymbolicSupernodalParams,
+    nomem, try_zeroed, FaerError, Index, SupernodalThreshold, SymbolicSparseColMatRef,
+    SymbolicSupernodalParams,
 };
 use core::{iter::zip, mem::MaybeUninit};
 use dyn_stack::{PodStack, SizeOverflow, StackReq};
@@ -16,7 +18,7 @@ use faer_core::{
     permutation::{PermutationRef, SignedIndex},
     solve,
     sparse::SparseColMatRef,
-    Conj, MatMut, Parallelism,
+    temp_mat_req, temp_mat_uninit, Conj, MatMut, Parallelism,
 };
 use faer_entity::*;
 use reborrow::*;
@@ -123,8 +125,8 @@ pub mod supernodal {
         pub(super) supernode_postorder: alloc::vec::Vec<I>,
         pub(super) supernode_postorder_inv: alloc::vec::Vec<I>,
         pub(super) descendant_count: alloc::vec::Vec<I>,
-        nrows: usize,
-        ncols: usize,
+        pub(super) nrows: usize,
+        pub(super) ncols: usize,
     }
 
     pub struct SupernodalLu<I, E: Entity> {
@@ -143,8 +145,6 @@ pub mod supernodal {
         ut_col_ptr_for_val: alloc::vec::Vec<I>,
         ut_row_ind: alloc::vec::Vec<I>,
         ut_val: GroupFor<E, alloc::vec::Vec<E::Unit>>,
-        // iwork: alloc::vec::Vec<I>,
-        // work: GroupFor<E, alloc::vec::Vec<E::Unit>>,
     }
     unsafe impl<I: Index, E: Entity> Send for SupernodalLu<I, E> {}
     unsafe impl<I: Index, E: Entity> Sync for SupernodalLu<I, E> {}
@@ -188,8 +188,6 @@ pub mod supernodal {
 
                 l_val: E::faer_map(E::UNIT, |()| alloc::vec::Vec::<E::Unit>::new()),
                 ut_val: E::faer_map(E::UNIT, |()| alloc::vec::Vec::<E::Unit>::new()),
-                // iwork: alloc::vec::Vec::new(),
-                // work: E::faer_map(E::UNIT, |()| alloc::vec::Vec::<E::Unit>::new()),
             }
         }
 
@@ -228,7 +226,7 @@ pub mod supernodal {
             faer_core::permutation::permute_rows(temp.rb_mut(), X.rb(), row_perm);
             self.l_solve_in_place_with_conj(conj_lhs, temp.rb_mut(), X.rb_mut(), parallelism);
             self.u_solve_in_place_with_conj(conj_lhs, temp.rb_mut(), X.rb_mut(), parallelism);
-            faer_core::permutation::permute_rows(X.rb_mut(), temp.rb(), col_perm);
+            faer_core::permutation::permute_rows(X.rb_mut(), temp.rb(), col_perm.inverse());
         }
 
         #[track_caller]
@@ -247,7 +245,7 @@ pub mod supernodal {
             assert!(self.nrows() == rhs.nrows());
             let mut X = rhs;
             let mut temp = work;
-            faer_core::permutation::permute_rows(temp.rb_mut(), X.rb(), col_perm.inverse());
[…]

[…]
