### 2023-11-29 #6099 @editorialbot
[REVIEW]: faer: A linear algebra library for the Rust programming language
**Submitting author:** @sarah-ek (<a href="http://orcid.org/0000-0002-5657-0710">Sarah El Kazdadi</a>)
**Repository:** https://github.com/sarah-ek/faer-rs
**Branch with paper.md** (empty if default branch):
**Version:** v0.15.0
