+++ faer-libs/faer-core/src/mul.rs
@@ -1,9 +1,7 @@
 //! Matrix multiplication.
 
-#[cfg(feature = "std")]
-use crate::assert;
 use crate::{
-    c32, c64, group_helpers::*, transmute_unchecked, unzipped, zipped, ComplexField, Conj,
+    assert, c32, c64, group_helpers::*, transmute_unchecked, unzipped, zipped, ComplexField, Conj,
     Conjugate, DivCeil, MatMut, MatRef, Parallelism, SimdGroupFor,
 };
 use core::{iter::zip, marker::PhantomData, mem::MaybeUninit};
@@ -197,9 +195,11 @@ pub mod inner_prod {
         rhs: MatRef<'_, E>,
         conj_rhs: Conj,
     ) -> E {
-        assert!(lhs.nrows() == rhs.nrows());
-        assert!(lhs.ncols() == 1);
-        assert!(rhs.ncols() == 1);
+        assert!(all(
+            lhs.nrows() == rhs.nrows(),
+            lhs.ncols() == 1,
+            rhs.ncols() == 1,
+        ));
         let nrows = lhs.nrows();
         let mut a = lhs;
         let mut b = rhs;
@@ -284,13 +284,14 @@ pub mod matvec_rowmajor {
         let m = a.nrows();
         let n = a.ncols();
 
-        assert!(b.nrows() == n);
-        assert!(b.ncols() == 1);
-        assert!(acc.nrows() == m);
-        assert!(acc.ncols() == 1);
-
-        assert!(a.col_stride() == 1);
-        assert!(b.row_stride() == 1);
+        assert!(all(
+            b.nrows() == n,
+            b.ncols() == 1,
+            acc.nrows() == m,
+            acc.ncols() == 1,
+            a.col_stride() == 1,
+            b.row_stride() == 1,
+        ));
 
         let mut acc = acc;
 
@@ -409,13 +410,14 @@ pub mod matvec_colmajor {
         let m = a.nrows();
         let n = a.ncols();
 
-        assert!(b.nrows() == n);
-        assert!(b.ncols() == 1);
-        assert!(acc.nrows() == m);
-        assert!(acc.ncols() == 1);
-
-        assert!(a.row_stride() == 1);
-        assert!(acc.row_stride() == 1);
+        assert!(all(
+            b.nrows() == n,
+            b.ncols() == 1,
+            acc.nrows() == m,
+            acc.ncols() == 1,
+            a.row_stride() == 1,
+            acc.row_stride() == 1,
+        ));
 
         let mut acc = SliceGroupMut::<'_, E>::new(acc.try_get_contiguous_col_mut(0));
 
@@ -677,13 +679,14 @@ pub mod outer_prod {
         let m = acc.nrows();
         let n = acc.ncols();
 
-        assert!(a.nrows() == m);
-        assert!(a.ncols() == 1);
-        assert!(b.nrows() == n);
-        assert!(b.ncols() == 1);
-
-        assert!(acc.row_stride() == 1);
-        assert!(a.row_stride() == 1);
+        assert!(all(
+            a.nrows() == m,
+            a.ncols() == 1,
+            b.nrows() == n,
+            b.ncols() == 1,
+            acc.row_stride() == 1,
+            a.row_stride() == 1,
+        ));
 
         let mut acc = acc;
 
@@ -867,14 +870,16 @@ impl<const MR_DIV_N: usize, const NR: usize, CB: ConjTy, E: ComplexField> pulp::
         let mr = MR_DIV_N * lane_count;
         let nr = NR;
 
-        assert!(acc.nrows() == mr);
-        assert!(acc.ncols() == nr);
-        assert!(a.nrows() == mr);
[…]
