+++ faer-libs/faer-core/src/lib.rs
@@ -4957,6 +4957,87 @@ const __ROW_INDEX: () = {
             unsafe { <Self as RowIndex<usize>>::get_unchecked(this, col) }
         }
     }
+
+    impl<E: Entity> RowIndex<RangeFull> for RowMut<'_, E> {
+        type Target = Self;
+
+        #[track_caller]
+        #[inline(always)]
+        fn get(this: Self, col: RangeFull) -> Self {
+            let _ = col;
+            this
+        }
+    }
+
+    impl<E: Entity> RowIndex<Range> for RowMut<'_, E> {
+        type Target = Self;
+
+        #[track_caller]
+        #[inline(always)]
+        fn get(this: Self, col: Range) -> Self {
+            this.subcols_mut(col.start, col.end - col.start)
+        }
+    }
+
+    impl<E: Entity> RowIndex<RangeInclusive> for RowMut<'_, E> {
+        type Target = Self;
+
+        #[track_caller]
+        #[inline(always)]
+        fn get(this: Self, col: RangeInclusive) -> Self {
+            assert!(*col.end() != usize::MAX);
+            <Self as RowIndex<Range>>::get(this, *col.start()..*col.end() + 1)
+        }
+    }
+
+    impl<E: Entity> RowIndex<RangeFrom> for RowMut<'_, E> {
+        type Target = Self;
+
+        #[track_caller]
+        #[inline(always)]
+        fn get(this: Self, col: RangeFrom) -> Self {
+            let ncols = this.ncols();
+            <Self as RowIndex<Range>>::get(this, col.start..ncols)
+        }
+    }
+
+    impl<E: Entity> RowIndex<RangeTo> for RowMut<'_, E> {
+        type Target = Self;
+
+        #[track_caller]
+        #[inline(always)]
+        fn get(this: Self, col: RangeTo) -> Self {
+            <Self as RowIndex<Range>>::get(this, 0..col.end)
+        }
+    }
+
+    impl<E: Entity> RowIndex<RangeToInclusive> for RowMut<'_, E> {
+        type Target = Self;
+
+        #[track_caller]
+        #[inline(always)]
+        fn get(this: Self, col: RangeToInclusive) -> Self {
+            assert!(col.end != usize::MAX);
+            <Self as RowIndex<Range>>::get(this, 0..col.end + 1)
+        }
+    }
+
+    impl<'a, E: Entity> RowIndex<usize> for RowMut<'a, E> {
+        type Target = GroupFor<E, &'a mut E::Unit>;
+
+        #[track_caller]
+        #[inline(always)]
+        unsafe fn get_unchecked(this: Self, col: usize) -> Self::Target {
+            unsafe { E::faer_map(this.ptr_inbounds_at_mut(col), |ptr: *mut _| &mut *ptr) }
+        }
+
+        #[track_caller]
+        #[inline(always)]
+        fn get(this: Self, col: usize) -> Self::Target {
+            assert!(col < this.ncols());
+            unsafe { <Self as RowIndex<usize>>::get_unchecked(this, col) }
+        }
+    }
 };
 
 impl<'a, E: Entity> Matrix<DiagRef<'a, E>> {
@@ -5468,6 +5549,52 @@ const __COL_IMPL: () = {
                     __marker: PhantomData,
                 },
             }
+        }
+    }
+
+    impl<E: SimpleEntity> core::ops::Index<usize> for ColRef<'_, E> {
+        type Output = E;
+
+        #[inline]
+        #[track_caller]
+        fn index(&self, row: usize) -> &E {
+            self.get(row)
+        }
+    }
+
[…]
