### 2023-12-02 67adf05
high level sparse api
