+++ faer-libs/faer/src/lib.rs
@@ -2,7 +2,7 @@
 //! for algebraic operations on medium/large matrices, as well as matrix decompositions.
 //!
 //! Most of the high-level functionality in this library is provided through associated functions in
-//! its vocabulary types: [`Mat`]/[`MatRef`]/[`MatMut`], as well as the [`Faer`] extension trait.
+//! its vocabulary types: [`Mat`]/[`MatRef`]/[`MatMut`], as well as the [`FaerMat`] extension trait.
 //! The parent crates (`faer-core`, `faer-cholesky`, `faer-lu`, etc.), on the other hand, offer a
 //! lower-level of abstraction in exchange for more control over memory allocations and
 //! multithreading behavior.
@@ -59,18 +59,18 @@
 //! `faer` provides a variety of matrix factorizations, each with its own advantages and drawbacks:
 //!
 //! ## Cholesky decomposition
-//! [`Faer::cholesky`] decomposes a self-adjoint positive definite matrix $A$ such that
+//! [`FaerMat::cholesky`] decomposes a self-adjoint positive definite matrix $A$ such that
 //! $$A = LL^H,$$
 //! where $L$ is a lower triangular matrix. This decomposition is highly efficient and has good
 //! stability properties.
 //! ## Bunch-Kaufman decomposition
-//! [`Faer::lblt`] decomposes a self-adjoint (possibly indefinite) matrix $A$ such that
+//! [`FaerMat::lblt`] decomposes a self-adjoint (possibly indefinite) matrix $A$ such that
 //! $$P A P^\top = LBL^H,$$
 //! where $P$ is a permutation matrix, $L$ is a lower triangular matrix, and $B$ is a block
 //! diagonal matrix, with $1 \times 1$ or $2 \times 2$ diagonal blocks.
 //! This decomposition is efficient and has good stability properties.
 //! ## LU decomposition with partial pivoting
-//! [`Faer::partial_piv_lu`] decomposes a square invertible matrix $A$ into a lower triangular
+//! [`FaerMat::partial_piv_lu`] decomposes a square invertible matrix $A$ into a lower triangular
 //! matrix $L$, a unit upper triangular matrix $U$, and a permutation matrix $P$, such that
 //! $$PA = LU.$$
 //! It is used by default for computing the determinant, and is generally the recommended method
@@ -78,20 +78,20 @@
 //! recommend using a [`Solver`] instead of computing the inverse explicitly).
 //!
 //! ## LU decomposition with full pivoting
-//! [`Faer::full_piv_lu`] Decomposes a generic rectangular matrix $A$ into a lower triangular matrix
-//! $L$, a unit upper triangular matrix $U$, and permutation matrices $P$ and $Q$, such that
+//! [`FaerMat::full_piv_lu`] Decomposes a generic rectangular matrix $A$ into a lower triangular
+//! matrix $L$, a unit upper triangular matrix $U$, and permutation matrices $P$ and $Q$, such that
 //! $$PAQ^\top = LU.$$
 //! It can be more stable than the LU decomposition with partial pivoting, in exchange for being
 //! more computationally expensive.
 //!
 //! ## QR decomposition
-//! The QR decomposition ([`Faer::qr`]) decomposes a matrix $A$ into the product
+//! The QR decomposition ([`FaerMat::qr`]) decomposes a matrix $A$ into the product
 //! $$A = QR,$$
 //! where $Q$ is a unitary matrix, and $R$ is an upper trapezoidal matrix. It is often used for
 //! solving least squares problems.
[…]
