+++ faer-libs/faer-core/Cargo.toml
@@ -58,7 +58,7 @@ criterion = "0.5"
 rand = "0.8.5"
 nalgebra = "0.32.3"
 assert_approx_eq = "1.1.0"
-dbgf = "0.1.0"
+dbgf = "0.1.1"
 
 [[bench]]
 name = "bench"
