+++ faer-libs/faer-qr/Cargo.toml
@@ -42,7 +42,7 @@ rand = "0.8.5"
 nalgebra = "0.32.3"
 assert_approx_eq = "1.1.0"
 rayon = "1.8"
-dbgf = "0.1.0"
+dbgf = "0.1.1"
 matrixcompare = "0.3"
 
 [[bench]]
