+++ faer-libs/faer-sparse/Cargo.toml
@@ -24,7 +24,7 @@ rayon = { workspace = true }
 bytemuck = { workspace = true }
 pulp = { workspace = true, default-features = false}
 coe-rs = { workspace = true }
-dbgf = "0.1.0"
+dbgf = "0.1.1"
 
 [features]
 default = ["std", "rayon"]
@@ -55,7 +55,7 @@ perf-warn = ["log"]
 criterion.workspace = true
 rand = "0.8.5"
 assert_approx_eq = "1.1.0"
-dbgf = "0.1.0"
+dbgf = "0.1.1"
 paste = "1.0.14"
 matrix-market-rs = "0.1.3"
 amd = "0.2.2"
