+++ faer-libs/faer-evd/Cargo.toml
@@ -32,7 +32,7 @@ criterion = "0.5"
 rand = "0.8.5"
 nalgebra = "0.32.3"
 assert_approx_eq = "1.1.0"
-dbgf = "0.1.0"
+dbgf = "0.1.1"
 
 [features]
 default = ["std", "rayon"]
