### 2023-12-08 8979ac2
Release 0.16.0
faer@0.16.0
faer-cholesky@0.16.0
faer-core@0.16.0
faer-evd@0.16.0
faer-lu@0.16.0
faer-qr@0.16.0
faer-sparse@0.16.0
faer-svd@0.16.0

Generated by cargo-workspaces
