+++ faer-entity/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "faer-entity"
-version = "0.15.0"
+version = "0.16.0"
 edition = "2021"
 authors = ["sarah <>"]
 description = "Basic linear algebra routines"
