### 2023-12-10 1ad55b9
Started refactor
