+++ faer-libs/faer-core/src/complex_native/c64.rs
@@ -1,0 +1,604 @@
+
+// 64-bit complex floating point type. See the module-level documentation for more details.
+#[allow(non_camel_case_types)]
+#[derive(Copy, Clone, PartialEq)]
+#[repr(C)]
+pub struct c64 {
+    pub re: f64,
+    pub im: f64,
+}
+
+impl c64 {
+    #[inline(always)]
+    pub fn new(re: f64, im: f64) -> Self {
+        Self { re, im }
+    }
+
+    #[inline(always)]
+    pub fn to_num_complex(self) -> Complex<f64> {
+        self.into()
+    }
+
+    #[inline(always)]
+    pub fn abs(self) -> f64 {
+        self.faer_abs()
+    }
+}
+impl core::ops::Neg for c64 {
+    type Output = c64;
+
+    #[inline(always)]
+    fn neg(self) -> Self::Output {
+        Self::new(-self.re, -self.im)
+    }
+}
+
+impl core::ops::Add<f64> for c64 {
+    type Output = c64;
+
+    #[inline(always)]
+    fn add(self, rhs: f64) -> Self::Output {
+        Self::new(self.re + rhs, self.im)
+    }
+}
+impl core::ops::Add<c64> for f64 {
+    type Output = c64;
+
+    #[inline(always)]
+    fn add(self, rhs: c64) -> Self::Output {
+        Self::Output::new(self + rhs.re, rhs.im)
+    }
+}
+impl core::ops::Add for c64 {
+    type Output = c64;
+
+    #[inline(always)]
+    fn add(self, rhs: Self) -> Self::Output {
+        Self::new(self.re + rhs.re, self.im + rhs.im)
+    }
+}
+
+impl core::ops::Sub<f64> for c64 {
+    type Output = c64;
+
+    #[inline(always)]
+    fn sub(self, rhs: f64) -> Self::Output {
+        Self::new(self.re - rhs, self.im)
+    }
+}
+impl core::ops::Sub<c64> for f64 {
+    type Output = c64;
+
+    #[inline(always)]
+    fn sub(self, rhs: c64) -> Self::Output {
+        Self::Output::new(self - rhs.re, -rhs.im)
+    }
+}
+impl core::ops::Sub for c64 {
+    type Output = c64;
+
+    #[inline(always)]
+    fn sub(self, rhs: Self) -> Self::Output {
+        Self::new(self.re - rhs.re, self.im - rhs.im)
+    }
+}
+
+impl core::ops::Mul<f64> for c64 {
+    type Output = c64;
+
+    #[inline(always)]
+    fn mul(self, rhs: f64) -> Self::Output {
+        Self::new(self.re * rhs, self.im * rhs)
+    }
+}
+impl core::ops::Mul<c64> for f64 {
+    type Output = c64;
+
+    #[inline(always)]
+    fn mul(self, rhs: c64) -> Self::Output {
+        Self::Output::new(self * rhs.re, self * rhs.im)
+    }
+}
+impl core::ops::Mul for c64 {
+    type Output = c64;
+
+    #[inline(always)]
+    fn mul(self, rhs: Self) -> Self::Output {
+        Self::new(
+            self.re * rhs.re - self.im * rhs.im,
+            self.re * rhs.im + self.im * rhs.re,
+        )
+    }
+}
+
+impl core::ops::Div<f64> for c64 {
+    type Output = c64;
+
+    #[inline(always)]
+    fn div(self, rhs: f64) -> Self::Output {
+        Self::new(self.re / rhs, self.im / rhs)
+    }
+}
+impl core::ops::Div<c64> for f64 {
[…]
