### 2023-12-12 #6099 @mhoemmen
# Notes on the above Checklist

## Software paper

### Summary

The summary is concise and explains the library well to me.  For example, "dense linear algebra library" is explained as "matrix decompositions and solving linear systems."  ("Linear algebra library" means different things to different people, so having the explanation is helpful.)

It would be interesting to know what inspired the name "faer."

### A statement of need

The statement of need is compelling.  The following quote conveys the motivation well.

> Aside from faer, the Rust ecosystem lacks high performance matrix factorization libraries that aren’t C library wrappers, which presents a distribution challenge and can impede generic programming.

The "impede generic programming" phrase is enough justification for me.  I've encountered C++ libraries that exist only for this reason, to make it possible to factor sparse matrices that aren't supported by popular C libraries.
