### 2023-12-13 a3278bf
Implemented several traits
