+++ faer-libs/faer-core/src/complex_native/c32_impl.rs
@@ -1,11 +1,12 @@
 use crate::complex_native::c32conj_impl::c32conj;
 use faer_entity::*;
-use num_traits::{One, Zero};
+use num_traits::{float::FloatCore, One, Zero};
 use pulp::{cast, Simd};
 
 #[cfg(feature = "std")]
 use num_complex::ComplexFloat;
 
+#[cfg(feature = "std")]
 macro_rules! impl_from_num_complex {
     ($( $method:ident ( self $( , $arg:ident : $ty:ty )* ) -> $ret:ty ; )*) => {
         $(
@@ -40,12 +41,14 @@ impl c32 {
     }
 
     /// Create a complex number from a phase.
+    #[cfg(feature = "std")]
     #[inline(always)]
     pub fn cis(phase: f32) -> Self {
         Self::new(phase.cos(), phase.sin())
     }
 
     /// Create a complex number from polar coordinates.
+    #[cfg(feature = "std")]
     #[inline(always)]
     pub fn from_polar(r: f32, theta: f32) -> Self {
         Self::new(r * theta.cos(), r * theta.sin())
@@ -75,6 +78,7 @@ impl c32 {
         self.faer_conj()
     }
 
+    #[cfg(feature = "std")]
     impl_from_num_complex!(
         is_nan(self) -> bool;
         is_infinite(self) -> bool;
