### 2023-12-16 c3043f2
Added ops between native and native conj
