+++ faer-libs/faer-core/src/complex_native/c32_impl.rs
@@ -204,6 +204,24 @@ impl core::ops::Add for c32 {
     }
 }
 
+impl core::ops::Add<c32conj> for c32 {
+    type Output = c32;
+
+    #[inline(always)]
+    fn add(self, rhs: c32conj) -> Self::Output {
+        Self::new(self.re + rhs.re, self.im - rhs.neg_im)
+    }
+}
+
+impl core::ops::Add<c32> for c32conj {
+    type Output = c32;
+
+    #[inline(always)]
+    fn add(self, rhs: c32) -> Self::Output {
+        Self::Output::new(self.re + rhs.re, rhs.im - self.neg_im)
+    }
+}
+
 impl core::ops::Sub<f32> for c32 {
     type Output = c32;
 
@@ -228,6 +246,24 @@ impl core::ops::Sub for c32 {
     #[inline(always)]
     fn sub(self, rhs: Self) -> Self::Output {
         Self::new(self.re - rhs.re, self.im - rhs.im)
+    }
+}
+
+impl core::ops::Sub<c32conj> for c32 {
+    type Output = c32;
+
+    #[inline(always)]
+    fn sub(self, rhs: c32conj) -> Self::Output {
+        Self::new(self.re - rhs.re, self.im + rhs.neg_im)
+    }
+}
+
+impl core::ops::Sub<c32> for c32conj {
+    type Output = c32;
+
+    #[inline(always)]
+    fn sub(self, rhs: c32) -> Self::Output {
+        Self::Output::new(self.re - rhs.re, -self.neg_im - rhs.im)
     }
 }
 
@@ -261,6 +297,30 @@ impl core::ops::Mul for c32 {
     }
 }
 
+impl core::ops::Mul<c32conj> for c32 {
+    type Output = c32;
+
+    #[inline(always)]
+    fn mul(self, rhs: c32conj) -> Self::Output {
+        Self::new(
+            self.re * rhs.re + self.im * rhs.neg_im,
+            self.im * rhs.re - self.re * rhs.neg_im,
+        )
+    }
+}
+
+impl core::ops::Mul<c32> for c32conj {
+    type Output = c32;
+
+    #[inline(always)]
+    fn mul(self, rhs: c32) -> Self::Output {
+        Self::Output::new(
+            self.re * rhs.re + self.neg_im * rhs.im,
+            rhs.im * self.re - rhs.re * self.neg_im,
+        )
+    }
+}
+
 impl core::ops::Div<f32> for c32 {
     type Output = c32;
 
@@ -286,6 +346,27 @@ impl core::ops::Div for c32 {
     #[allow(clippy::suspicious_arithmetic_impl)]
     #[inline(always)]
     fn div(self, rhs: Self) -> Self::Output {
+        self * rhs.faer_inv()
+    }
+}
+
+impl core::ops::Div<c32conj> for c32 {
+    type Output = c32;
+
+    #[allow(clippy::suspicious_arithmetic_impl)]
+    #[inline(always)]
+    fn div(self, rhs: c32conj) -> Self::Output {
+        // TODO: Can be done a bit better
+        self * rhs.canonicalize().faer_inv()
+    }
+}
+
+impl core::ops::Div<c32> for c32conj {
+    type Output = c32;
+
+    #[allow(clippy::suspicious_arithmetic_impl)]
+    #[inline(always)]
+    fn div(self, rhs: c32) -> Self::Output {
         self * rhs.faer_inv()
     }
 }
@@ -317,6 +398,24 @@ impl core::ops::Rem for c32 {
     }
 }
 
+impl core::ops::Rem<c32conj> for c32 {
+    type Output = c32;
+
[…]
