+++ faer-libs/faer-core/src/complex_native/c64_impl.rs
@@ -1,6 +1,8 @@
 use crate::complex_native::c64conj_impl::c64conj;
 use faer_entity::*;
-use num_traits::{float::FloatCore, One, Zero};
+#[cfg(not(feature = "std"))]
+use num_traits::float::FloatCore;
+use num_traits::{One, Zero};
 use pulp::{cast, Simd};
 
 #[cfg(feature = "std")]
