+++ faer-libs/faer-core/src/lib.rs
@@ -6326,6 +6326,15 @@ const __COL_IMPL: () = {
             norm_l2((*self).rb().as_2d())
         }
 
+        /// Returns the sum of `self`.
+        #[inline]
+        pub fn sum(&self) -> E
+        where
+            E: ComplexField,
+        {
+            sum((*self).rb().as_2d())
+        }
+
         /// Returns a view over the matrix.
         #[inline]
         pub fn as_ref(&self) -> ColRef<'_, E> {
@@ -6758,6 +6767,15 @@ const __COL_IMPL: () = {
             norm_l2((*self).rb().as_2d())
         }
 
+        /// Returns the sum of `self`.
+        #[inline]
+        pub fn sum(&self) -> E
+        where
+            E: ComplexField,
+        {
+            sum((*self).rb().as_2d())
+        }
+
         /// Returns a view over the matrix.
         #[inline]
         pub fn as_ref(&self) -> ColRef<'_, E> {
@@ -7118,6 +7136,15 @@ const __ROW_IMPL: () = {
             E: ComplexField,
         {
             norm_l2((*self).rb().as_2d())
+        }
+
+        /// Returns the sum of `self`.
+        #[inline]
+        pub fn sum(&self) -> E
+        where
+            E: ComplexField,
+        {
+            sum((*self).rb().as_2d())
         }
 
         /// Returns a view over the matrix.
@@ -7526,6 +7553,15 @@ const __ROW_IMPL: () = {
             E: ComplexField,
         {
             norm_l2((*self).rb().as_2d())
+        }
+
+        /// Returns the sum of `self`.
+        #[inline]
+        pub fn sum(&self) -> E
+        where
+            E: ComplexField,
+        {
+            sum((*self).rb().as_2d())
         }
 
         /// Returns a view over the matrix.
@@ -8424,6 +8460,15 @@ const __MAT_IMPL: () = {
             E: ComplexField,
         {
             norm_l2((*self).rb())
+        }
+
+        /// Returns the sum of `self`.
+        #[inline]
+        pub fn sum(&self) -> E
+        where
+            E: ComplexField,
+        {
+            sum((*self).rb())
         }
 
         /// Returns a view over the matrix.
@@ -9593,6 +9638,15 @@ const __MAT_IMPL: () = {
             norm_l2((*self).rb())
         }
 
+        /// Returns the sum of `self`.
+        #[inline]
+        pub fn sum(&self) -> E
+        where
+            E: ComplexField,
+        {
+            sum((*self).rb())
+        }
+
         /// Returns a view over the matrix.
         #[inline]
         pub fn as_ref(&self) -> MatRef<'_, E> {
@@ -10626,6 +10680,15 @@ impl<E: Entity> Col<E> {
     {
         norm_l2((*self).as_ref().as_2d())
     }
+
+    /// Returns the sum of `self`.
+    #[inline]
+    pub fn sum(&self) -> E
+    where
+        E: ComplexField,
+    {
+        sum((*self).as_ref().as_2d())
+    }
 }
 
 impl<E: Entity> Row<E> {
@@ -11109,6 +11172,15 @@ impl<E: Entity> Row<E> {
         E: ComplexField,
     {
         norm_l2((*self).as_ref().as_2d())
+    }
+    
+    /// Returns the sum of `self`.
+    #[inline]
+    pub fn sum(&self) -> E
+    where
+        E: ComplexField,
+    {
+        sum((*self).as_ref().as_2d())
     }
 }
 
@@ -11744,6 +11816,15 @@ impl<E: Entity> Mat<E> {
         E: ComplexField,
     {
         norm_l2((*self).as_ref())
+    }
+
[…]
