### 2024-01-17 #6099 @jedbrown
@mhoemmen :wave: How is your review going? I see you had a question about allocation tucked in the end of your comment. Faer supports `no_std`, which is an environment without allocation or a standard library. In that sense, it is low-level. I think you'll find ergonomics of the high-level interfaces are intermediate, albeit safe.
