+++ faer-libs/faer-core/src/lib.rs
@@ -13577,6 +13577,51 @@ fn norm_l2_with_simd_and_offset_prologue<E: ComplexField, S: pulp::Simd>(
 }
 
 #[inline(always)]
+fn sum_with_simd_and_offset_prologue<E: ComplexField, S: pulp::Simd>(
+    simd: S,
+    data: SliceGroup<'_, E>,
+    offset: pulp::Offset<SimdMaskFor<E, S>>,
+) -> SimdGroupFor<E, S> {
+    use group_helpers::*;
+    
+    let simd = SimdFor::<E, S>::new(simd);
+
+    let zero = simd.splat(E::faer_zero());
+
+    let mut acc0 = zero;
+    let mut acc1 = zero;
+    let mut acc2 = zero;
+    let mut acc3 = zero;
+    let (head, body, tail) = simd.as_aligned_simd(data, offset);
+    let (body4, body1) = body.as_arrays::<4>();
+    let head = head.read_or(zero);
+    acc0 = simd.add(acc0, head);
+
+    for [x0, x1, x2, x3] in body4.into_ref_iter().map(RefGroup::unzip) {
+        let x0 = x0.get();
+        let x1 = x1.get();
+        let x2 = x2.get();
+        let x3 = x3.get();
+        acc0 = simd.add(acc0, x0);
+        acc1 = simd.add(acc1, x1);
+        acc2 = simd.add(acc2, x2);
+        acc3 = simd.add(acc3, x3);
+    }
+
+    for x0 in body1.into_ref_iter() {
+        let x0 = x0.get();
+        acc0 = simd.add(acc0, x0);
+    }
+
+    let tail = tail.read_or(zero);
+    acc3 = simd.add(acc3, tail);
+    
+    acc0 = simd.add(acc0, acc1);
+    acc2 = simd.add(acc2, acc3);
+    simd.add(acc0, acc2)
+}
+
+#[inline(always)]
 fn norm_max_contiguous<E: RealField>(data: MatRef<'_, E>) -> E {
     struct Impl<'a, E: RealField> {
         data: MatRef<'a, E>,
@@ -13719,6 +13764,57 @@ fn norm_l2_with_simd_and_offset_pairwise_rows<E: ComplexField, S: Simd>(
 }
 
 #[inline(always)]
+fn sum_with_simd_and_offset_pairwise_rows<E: ComplexField, S: Simd>(
+    simd: S,
+    data: SliceGroup<'_, E>,
+    offset: pulp::Offset<SimdMaskFor<E, S>>,
+    last_offset: pulp::Offset<SimdMaskFor<E, S>>,
+) -> SimdGroupFor<E, S> {
+    struct Impl<'a, E: ComplexField, S: Simd> {
+        simd: S,
+        data: SliceGroup<'a, E>,
+        offset: pulp::Offset<SimdMaskFor<E, S>>,
+        last_offset: pulp::Offset<SimdMaskFor<E, S>>,
+    }
+
+    impl<E: ComplexField, S: Simd> pulp::NullaryFnOnce for Impl<'_, E, S> {
+        type Output = SimdGroupFor<E, S>;
+
+        #[inline(always)]
+        fn call(self) -> Self::Output {
+            let Self {
+                simd,
+                data,
+                offset,
+                last_offset,
+            } = self;
+
+            if data.len() == NORM_L2_THRESHOLD {
+                sum_with_simd_and_offset_prologue(simd, data, offset)
+            } else if data.len() < NORM_L2_THRESHOLD {
+                sum_with_simd_and_offset_prologue(simd, data, last_offset)
+            } else {
+                let split_point = ((data.len() + 1) / 2).next_power_of_two();
+                let (head, tail) = data.split_at(split_point);
[…]
