### 2024-02-03 99a2c21
Num traits for complex native (#92)
* Started refactor

* Implemented several traits

* Fixed nostd build

* Added ops between native and native conj

* Fixed warnings

* Addressed comments
