### 2024-02-03 6ca9fb6
Sum impl (#93)
* Init impl with failing test

* Implemented tree-sum
