+++ faer-libs/faer-core/src/lib.rs
@@ -4891,6 +4891,15 @@ const __COL_IMPL: () = {
             norm_l2((*self).rb().as_2d())
         }
 
+        /// Returns the sum of `self`.
+        #[inline]
+        pub fn sum(&self) -> E
+        where
+            E: ComplexField,
+        {
+            sum((*self).rb().as_2d())
+        }
+
         /// Returns a view over the matrix.
         #[inline]
         pub fn as_ref(&self) -> ColRef<'_, E> {
@@ -5323,6 +5332,15 @@ const __COL_IMPL: () = {
             norm_l2((*self).rb().as_2d())
         }
 
+        /// Returns the sum of `self`.
+        #[inline]
+        pub fn sum(&self) -> E
+        where
+            E: ComplexField,
+        {
+            sum((*self).rb().as_2d())
+        }
+
         /// Returns a view over the matrix.
         #[inline]
         pub fn as_ref(&self) -> ColRef<'_, E> {
@@ -5683,6 +5701,15 @@ const __ROW_IMPL: () = {
             E: ComplexField,
         {
             norm_l2((*self).rb().as_2d())
+        }
+
+        /// Returns the sum of `self`.
+        #[inline]
+        pub fn sum(&self) -> E
+        where
+            E: ComplexField,
+        {
+            sum((*self).rb().as_2d())
         }
 
         /// Returns a view over the matrix.
@@ -6091,6 +6118,15 @@ const __ROW_IMPL: () = {
             E: ComplexField,
         {
             norm_l2((*self).rb().as_2d())
+        }
+
+        /// Returns the sum of `self`.
+        #[inline]
+        pub fn sum(&self) -> E
+        where
+            E: ComplexField,
+        {
+            sum((*self).rb().as_2d())
         }
 
         /// Returns a view over the matrix.
@@ -6989,6 +7025,15 @@ const __MAT_IMPL: () = {
             E: ComplexField,
         {
             norm_l2((*self).rb())
+        }
+
+        /// Returns the sum of `self`.
+        #[inline]
+        pub fn sum(&self) -> E
+        where
+            E: ComplexField,
+        {
+            sum((*self).rb())
         }
 
         /// Returns a view over the matrix.
@@ -8158,6 +8203,15 @@ const __MAT_IMPL: () = {
             norm_l2((*self).rb())
         }
 
+        /// Returns the sum of `self`.
+        #[inline]
+        pub fn sum(&self) -> E
+        where
+            E: ComplexField,
+        {
+            sum((*self).rb())
+        }
+
         /// Returns a view over the matrix.
         #[inline]
         pub fn as_ref(&self) -> MatRef<'_, E> {
@@ -9191,6 +9245,15 @@ impl<E: Entity> Col<E> {
     {
         norm_l2((*self).as_ref().as_2d())
     }
+
+    /// Returns the sum of `self`.
+    #[inline]
+    pub fn sum(&self) -> E
+    where
+        E: ComplexField,
+    {
+        sum((*self).as_ref().as_2d())
+    }
 }
 
 impl<E: Entity> Row<E> {
@@ -9674,6 +9737,15 @@ impl<E: Entity> Row<E> {
         E: ComplexField,
     {
         norm_l2((*self).as_ref().as_2d())
+    }
+    
+    /// Returns the sum of `self`.
+    #[inline]
+    pub fn sum(&self) -> E
+    where
+        E: ComplexField,
+    {
+        sum((*self).as_ref().as_2d())
     }
 }
 
[…]
