### 2024-02-03 f473819
fix no_std and sum impl alignment variability
