+++ faer-libs/faer-core/src/complex_native/c64_impl.rs
@@ -123,7 +123,7 @@ impl c64 {
 
     #[inline(always)]
     pub fn l1_norm(&self) -> f64 {
-        self.re.abs() + self.im.abs()
+        self.re.faer_abs() + self.im.faer_abs()
     }
 
     #[inline(always)]
@@ -593,7 +593,7 @@ unsafe impl bytemuck::Pod for c64 {}
 impl core::fmt::Debug for c64 {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         self.re.fmt(f)?;
-        let im_abs = self.im.abs();
+        let im_abs = self.im.faer_abs();
         if self.im.is_sign_positive() {
             f.write_str(" + ")?;
             im_abs.fmt(f)?;
