+++ faer-libs/faer-core/src/lib.rs
@@ -95,7 +95,6 @@ use core::{
 };
 use dyn_stack::{PodStack, SizeOverflow, StackReq};
 use group_helpers::SliceGroup;
-use inner::*;
 use num_complex::Complex;
 use pulp::Simd;
 use reborrow::*;
@@ -709,27 +708,27 @@ pub mod inner {
 
     #[derive(Debug)]
     #[doc(hidden)]
-    pub struct SparseColMatRefInner<'a, I, E: Entity> {
+    pub struct SparseColMatRef<'a, I, E: Entity> {
         pub(crate) symbolic: sparse::SymbolicSparseColMatRef<'a, I>,
         pub(crate) values: SliceGroup<'a, E>,
     }
 
     #[derive(Debug)]
     #[doc(hidden)]
-    pub struct SparseRowMatRefInner<'a, I, E: Entity> {
+    pub struct SparseRowMatRef<'a, I, E: Entity> {
         pub(crate) symbolic: sparse::SymbolicSparseRowMatRef<'a, I>,
         pub(crate) values: SliceGroup<'a, E>,
     }
 
-    impl<I, E: Entity> Copy for SparseRowMatRefInner<'_, I, E> {}
-    impl<I, E: Entity> Clone for SparseRowMatRefInner<'_, I, E> {
+    impl<I, E: Entity> Copy for SparseRowMatRef<'_, I, E> {}
+    impl<I, E: Entity> Clone for SparseRowMatRef<'_, I, E> {
         #[inline]
         fn clone(&self) -> Self {
             *self
         }
     }
-    impl<I, E: Entity> Copy for SparseColMatRefInner<'_, I, E> {}
-    impl<I, E: Entity> Clone for SparseColMatRefInner<'_, I, E> {
+    impl<I, E: Entity> Copy for SparseColMatRef<'_, I, E> {}
+    impl<I, E: Entity> Clone for SparseColMatRef<'_, I, E> {
         #[inline]
         fn clone(&self) -> Self {
             *self
@@ -2727,10 +2726,10 @@ pub mod sparse {
     }
 
     /// Sparse matrix in column format, either compressed or uncompressed.
-    pub type SparseRowMatRef<'a, I, E> = Matrix<inner::SparseRowMatRefInner<'a, I, E>>;
+    pub type SparseRowMatRef<'a, I, E> = Matrix<inner::SparseRowMatRef<'a, I, E>>;
 
     /// Sparse matrix in column format, either compressed or uncompressed.
-    pub type SparseColMatRef<'a, I, E> = Matrix<inner::SparseColMatRefInner<'a, I, E>>;
+    pub type SparseColMatRef<'a, I, E> = Matrix<inner::SparseColMatRef<'a, I, E>>;
 
     impl<'a, I: Index, E: Entity> SparseRowMatRef<'a, I, E> {
         /// Creates a new sparse matrix view.
@@ -2748,7 +2747,7 @@ pub mod sparse {
             let values = SliceGroup::new(values);
             assert!(symbolic.col_indices().len() == values.len());
             Self {
-                inner: inner::SparseRowMatRefInner { symbolic, values },
+                inner: inner::SparseRowMatRef { symbolic, values },
             }
         }
 
@@ -2761,7 +2760,7 @@ pub mod sparse {
         #[inline]
         pub fn transpose(self) -> SparseColMatRef<'a, I, E> {
             SparseColMatRef {
-                inner: inner::SparseColMatRefInner {
+                inner: inner::SparseColMatRef {
                     symbolic: SymbolicSparseColMatRef {
                         nrows: self.inner.symbolic.ncols,
                         ncols: self.inner.symbolic.nrows,
@@ -2780,7 +2779,7 @@ pub mod sparse {
[…]
