### 2024-02-03 97fbe57
elide bound checks in sparse triangular solve
