+++ faer-libs/faer-sparse/src/triangular_solve.rs
@@ -4,7 +4,6 @@ use faer_core::{
     Parallelism,
 };
 use faer_entity::ComplexField;
-use reborrow::*;
 
 pub fn solve_lower_triangular_in_place<I: Index, E: ComplexField>(
     l: SparseColMatRef<'_, I, E>,
@@ -18,127 +17,149 @@ pub fn solve_lower_triangular_in_place<I: Index, E: ComplexField>(
 
     let slice_group = SliceGroup::<'_, E>::new;
 
-    let n = l.ncols();
-    let mut x = rhs;
-    for mut x in x.rb_mut().col_chunks_mut(4) {
-        match x.ncols() {
-            1 => {
-                for j in 0..n {
-                    let d = slice_group(l.values_of_col(j)).read(0).faer_inv();
-                    let d = if conj == Conj::Yes { d.faer_conj() } else { d };
+    faer_core::constrained::Size::with2(
+        rhs.nrows(),
+        rhs.ncols(),
+        #[inline(always)]
+        |N, K| {
+            let mut x = faer_core::constrained::MatMut::new(rhs, N, K);
+            let l = faer_core::constrained::sparse::SparseColMatRef::new(l, N, N);
 
-                    let xj0 = x.read(j, 0).faer_mul(d);
-                    x.write(j, 0, xj0);
+            let mut k = 0usize;
+            while k < *K {
+                let bs = Ord::min(*K - k, 4);
+                match bs {
+                    1 => {
+                        let k0 = K.check(k);
 
-                    for (i, lij) in zip(
-                        l.row_indices_of_col(j),
-                        slice_group(l.values_of_col(j)).into_ref_iter(),
-                    )
-                    .skip(1)
-                    {
-                        let lij = lij.read();
-                        let lij = if conj == Conj::Yes {
-                            lij.faer_conj()
-                        } else {
-                            lij
-                        };
-                        x.write(i, 0, x.read(i, 0).faer_sub(lij.faer_mul(xj0)));
+                        for j in N.indices() {
+                            let d = slice_group(l.values_of_col(j)).read(0).faer_inv();
+                            let d = if conj == Conj::Yes { d.faer_conj() } else { d };
+
+                            let xj0 = x.read(j, k0).faer_mul(d);
+                            x.write(j, k0, xj0);
+
+                            for (i, lij) in zip(
+                                l.row_indices_of_col(j),
+                                slice_group(l.values_of_col(j)).into_ref_iter(),
+                            )
+                            .skip(1)
+                            {
+                                let lij = lij.read();
+                                let lij = if conj == Conj::Yes {
+                                    lij.faer_conj()
+                                } else {
+                                    lij
+                                };
+                                x.write(i, k0, x.read(i, k0).faer_sub(lij.faer_mul(xj0)));
+                            }
+                        }
                     }
-                }
-            }
-            2 => {
-                for j in 0..n {
-                    let d = slice_group(l.values_of_col(j)).read(0).faer_inv();
-                    let d = if conj == Conj::Yes { d.faer_conj() } else { d };
+                    2 => {
+                        let k0 = K.check(k);
+                        let k1 = K.check(k + 1);
+                        for j in N.indices() {
+                            let d = slice_group(l.values_of_col(j)).read(0).faer_inv();
+                            let d = if conj == Conj::Yes { d.faer_conj() } else { d };
 
-                    let xj0 = x.read(j, 0).faer_mul(d);
-                    x.write(j, 0, xj0);
-                    let xj1 = x.read(j, 1).faer_mul(d);
-                    x.write(j, 1, xj1);
+                            let xj0 = x.read(j, k0).faer_mul(d);
+                            x.write(j, k0, xj0);
[…]
