### 2024-02-05 47f1f2b
impl Sparse[Col|Row]MatMut, Sparse[Col|Row]Mat, with constructors from triplets
