+++ faer-libs/faer-core/src/sparse.rs
@@ -1,0 +1,3377 @@
+//! Sparse matrix data structures.
+//!
+//! Most sparse matrix algorithms accept matrices in sparse column-oriented format.
+//! This format represents each column of the matrix by storing the row indices of its non-zero
+//! elements, as well as their values.
+//!
+//! The indices and the values are each stored in a contiguous slice (or group of slices for
+//! arbitrary values). In order to specify where each column starts and ends, a slice of size
+//! `ncols + 1` stores the start of each column, with the last element being equal to the total
+//! number of non-zeros (or the capacity in uncompressed mode).
+//!
+//! # Example
+//!
+//! Consider the 4-by-5 matrix:
+//! ```notcode
+//! 10.0  0.0  12.0  -1.0  13.0
+//!  0.0  0.0  25.0  -2.0   0.0
+//!  1.0  0.0   0.0   0.0   0.0
+//!  4.0  0.0   0.0   0.0   5.0
+//! ```
+//!
+//! The matrix is stored as follows:
+//! ```notcode
+//! column pointers:  0 |  3 |  3 |  5 |  7 |  9
+//!
+//! row indices:    0 |    2 |    3 |    0 |    1 |    0 |    1 |    0 |    3
+//! values     : 10.0 |  1.0 |  4.0 | 12.0 | 25.0 | -1.0 | -2.0 | 13.0 |  5.0
+//! ```
+
+use super::*;
+use crate::{assert, debug_assert, group_helpers::VecGroup};
+use core::{cell::Cell, iter::zip, ops::Range, slice::SliceIndex};
+use group_helpers::SliceGroup;
+use permutation::{Index, SignedIndex};
+
+mod ghost {
+    pub use crate::constrained::{group_helpers::*, permutation::*, sparse::*, *};
+}
+
+mod mem {
+    #[inline]
+    pub fn fill_zero<I: bytemuck::Zeroable>(slice: &mut [I]) {
+        let len = slice.len();
+        unsafe { core::ptr::write_bytes(slice.as_mut_ptr(), 0u8, len) }
+    }
+}
+
+#[inline(always)]
+#[track_caller]
+#[doc(hidden)]
+pub unsafe fn __get_unchecked<I, R: Clone + SliceIndex<[I]>>(slice: &[I], i: R) -> &R::Output {
+    #[cfg(debug_assertions)]
+    {
+        let _ = &slice[i.clone()];
+    }
+    unsafe { slice.get_unchecked(i) }
+}
+#[inline(always)]
+#[track_caller]
+#[doc(hidden)]
+pub unsafe fn __get_unchecked_mut<I, R: Clone + SliceIndex<[I]>>(
+    slice: &mut [I],
+    i: R,
+) -> &mut R::Output {
+    #[cfg(debug_assertions)]
+    {
+        let _ = &slice[i.clone()];
+    }
+    unsafe { slice.get_unchecked_mut(i) }
+}
+
+#[inline(always)]
+#[doc(hidden)]
+pub fn windows2<I>(slice: &[I]) -> impl DoubleEndedIterator<Item = &[I; 2]> {
+    slice
+        .windows(2)
+        .map(|window| unsafe { &*(window.as_ptr() as *const [I; 2]) })
+}
+
+#[inline]
+#[doc(hidden)]
+pub const fn repeat_byte(byte: u8) -> usize {
+    union Union {
+        bytes: [u8; 32],
+        value: usize,
+    }
+
+    let data = Union { bytes: [byte; 32] };
+    unsafe { data.value }
+}
+
+/// Symbolic view structure of sparse matrix in column format, either compressed or uncompressed.
+///
+/// Requires:
[…]
