+++ faer-libs/faer-core/src/sparse.rs
@@ -30,6 +30,7 @@
 use super::*;
 use crate::{assert, debug_assert, group_helpers::VecGroup};
 use core::{cell::Cell, iter::zip, ops::Range, slice::SliceIndex};
+use dyn_stack::GlobalPodBuffer;
 use group_helpers::SliceGroup;
 use permutation::{Index, SignedIndex};
 
@@ -315,6 +316,18 @@ impl<I: Index> SymbolicSparseRowMat<I> {
         }
     }
 
+    /// Copies the current matrix into a newly allocated matrix.
+    #[inline]
+    pub fn to_owned(&self) -> Result<SymbolicSparseRowMat<I>, FaerError> {
+        self.as_ref().to_owned()
+    }
+
+    /// Copies the current matrix into a newly allocated matrix, with column-major order.
+    #[inline]
+    pub fn to_col_major(&self) -> Result<SymbolicSparseColMat<I>, FaerError> {
+        self.as_ref().to_col_major()
+    }
+
     /// Returns the number of symbolic non-zeros in the matrix.
     ///
     /// The value is guaranteed to be less than `I::Signed::MAX`.
@@ -504,6 +517,18 @@ impl<I: Index> SymbolicSparseColMat<I> {
         }
     }
 
+    /// Copies the current matrix into a newly allocated matrix.
+    #[inline]
+    pub fn to_owned(&self) -> Result<SymbolicSparseColMat<I>, FaerError> {
+        self.as_ref().to_owned()
+    }
+
+    /// Copies the current matrix into a newly allocated matrix, with row-major order.
+    #[inline]
+    pub fn to_row_major(&self) -> Result<SymbolicSparseRowMat<I>, FaerError> {
+        self.as_ref().to_row_major()
+    }
+
     /// Returns the number of symbolic non-zeros in the matrix.
     ///
     /// The value is guaranteed to be less than `I::Signed::MAX`.
@@ -678,6 +703,20 @@ impl<'a, I: Index> SymbolicSparseRowMatRef<'a, I> {
             col_nnz: self.row_nnz,
             row_ind: self.col_ind,
         }
+    }
+
+    /// Copies the current matrix into a newly allocated matrix.
+    #[inline]
+    pub fn to_owned(&self) -> Result<SymbolicSparseRowMat<I>, FaerError> {
+        self.transpose()
+            .to_owned()
+            .map(SymbolicSparseColMat::into_transpose)
+    }
+
+    /// Copies the current matrix into a newly allocated matrix, with column-major order.
+    #[inline]
+    pub fn to_col_major(&self) -> Result<SymbolicSparseColMat<I>, FaerError> {
+        self.transpose().to_row_major().map(|m| m.into_transpose())
     }
 
     /// Returns the number of symbolic non-zeros in the matrix.
@@ -871,6 +910,39 @@ impl<'a, I: Index> SymbolicSparseColMatRef<'a, I> {
         }
     }
 
+    /// Copies the current matrix into a newly allocated matrix.
+    #[inline]
+    pub fn to_owned(&self) -> Result<SymbolicSparseColMat<I>, FaerError> {
+        Ok(SymbolicSparseColMat {
+            nrows: self.nrows,
+            ncols: self.ncols,
+            col_ptr: try_collect(self.col_ptr.iter().copied())?,
+            col_nnz: self
+                .col_nnz
+                .map(|nnz| try_collect(nnz.iter().copied()))
+                .transpose()?,
+            row_ind: try_collect(self.row_ind.iter().copied())?,
+        })
+    }
+
+    /// Copies the current matrix into a newly allocated matrix, with row-major order.
+    #[inline]
[…]
