+++ faer-libs/faer-core/src/sort.rs
@@ -1,0 +1,1225 @@
+use crate::debug_assert;
+use faer_entity::{from_copy, into_copy, Entity, GroupCopyFor, GroupFor};
+
+pub unsafe trait Ptr: Sized + Copy + Ord {
+    type Item: Copy;
+
+    fn get_ptr(ptr: *mut Self::Item) -> Self;
+
+    fn null() -> Self;
+
+    unsafe fn offset_from(self, origin: Self) -> isize;
+
+    unsafe fn add(self, offset: usize) -> Self;
+    unsafe fn sub(self, offset: usize) -> Self;
+
+    unsafe fn read(self) -> Self::Item;
+    unsafe fn write(self, item: Self::Item);
+
+    unsafe fn copy_nonoverlapping(src: Self, dst: Self, len: usize);
+    unsafe fn reverse(ptr: Self, len: usize);
+
+    unsafe fn swap(a: Self, b: Self) {
+        let a_item = a.read();
+        let b_item = b.read();
+
+        a.write(b_item);
+        b.write(a_item);
+    }
+
+    unsafe fn swap_idx(self, i: usize, j: usize) {
+        Self::swap(self.add(i), self.add(j));
+    }
+}
+
+unsafe impl<T: Copy> Ptr for *mut T {
+    type Item = T;
+
+    #[inline]
+    fn get_ptr(ptr: *mut Self::Item) -> Self {
+        ptr
+    }
+
+    #[inline]
+    fn null() -> Self {
+        core::ptr::null_mut()
+    }
+
+    #[inline]
+    unsafe fn offset_from(self, origin: Self) -> isize {
+        self.offset_from(origin)
+    }
+
+    #[inline]
+    unsafe fn add(self, offset: usize) -> Self {
+        self.add(offset)
+    }
+
+    #[inline]
+    unsafe fn sub(self, offset: usize) -> Self {
+        self.sub(offset)
+    }
+
+    #[inline]
+    unsafe fn read(self) -> Self::Item {
+        *self
+    }
+
+    #[inline]
+    unsafe fn write(self, item: Self::Item) {
+        *self = item;
+    }
+
+    #[inline]
+    unsafe fn copy_nonoverlapping(src: Self, dst: Self, len: usize) {
+        core::ptr::copy_nonoverlapping(src, dst, len);
+    }
+
+    #[inline]
+    unsafe fn reverse(ptr: Self, len: usize) {
+        core::slice::from_raw_parts_mut(ptr, len).reverse()
+    }
+}
+
+unsafe impl<P: Ptr, Q: Ptr> Ptr for (P, Q) {
+    type Item = (P::Item, Q::Item);
+
+    #[inline]
+    fn get_ptr(ptr: *mut Self::Item) -> Self {
+        unsafe {
+            (
+                P::get_ptr(core::ptr::addr_of_mut!((*ptr).0)),
+                Q::get_ptr(core::ptr::addr_of_mut!((*ptr).1)),
+            )
+        }
+    }
+
+    #[inline]
+    fn null() -> Self {
+        (P::null(), Q::null())
+    }
+
+    #[inline]
+    unsafe fn offset_from(self, origin: Self) -> isize {
+        self.0.offset_from(origin.0)
+    }
+
+    #[inline]
+    unsafe fn add(self, offset: usize) -> Self {
+        (self.0.add(offset), self.1.add(offset))
+    }
+
+    #[inline]
+    unsafe fn sub(self, offset: usize) -> Self {
+        (self.0.sub(offset), self.1.sub(offset))
+    }
+
+    #[inline]
+    unsafe fn read(self) -> Self::Item {
+        (self.0.read(), self.1.read())
+    }
+
+    #[inline]
[…]
