+++ faer-libs/faer-core/src/complex_native/c32_impl.rs
@@ -1013,6 +1013,11 @@ unsafe impl Entity for c32 {
     }
 
     #[inline(always)]
+    fn faer_as_ptr<T>(group: *mut GroupFor<Self, T>) -> GroupFor<Self, *mut T> {
+        group
+    }
+
+    #[inline(always)]
     fn faer_map_impl<T, U>(
         group: GroupFor<Self, T>,
         f: &mut impl FnMut(T) -> U,
