+++ faer-libs/faer-core/src/complex_native/c32conj_impl.rs
@@ -80,6 +80,11 @@ unsafe impl Entity for c32conj {
     }
 
     #[inline(always)]
+    fn faer_as_ptr<T>(group: *mut GroupFor<Self, T>) -> GroupFor<Self, *mut T> {
+        group
+    }
+
+    #[inline(always)]
     fn faer_map_impl<T, U>(
         group: GroupFor<Self, T>,
         f: &mut impl FnMut(T) -> U,
