### 2024-02-07 36e15b9
document sorting requirements
