+++ faer-libs/faer-evd/src/hessenberg.rs
@@ -7,12 +7,14 @@ use faer_core::{
         apply_block_householder_on_the_right_in_place_with_conj, make_householder_in_place_v2,
         upgrade_householder_factor,
     },
-    mul::{inner_prod::inner_prod_with_conj, matmul},
+    mul::{inner_prod::inner_prod_with_conj, matmul, triangular::BlockStructure},
     parallelism_degree, temp_mat_req, temp_mat_uninit, temp_mat_zeroed, unzipped, zipped,
     ComplexField, Conj, Entity, MatMut, MatRef, Parallelism, SimdCtx,
 };
 use faer_entity::*;
 use reborrow::*;
+
+const BLOCKING_THRESHOLD: usize = 256;
 
 pub fn make_hessenberg_in_place_req<E: Entity>(
     n: usize,
@@ -20,16 +22,23 @@ pub fn make_hessenberg_in_place_req<E: Entity>(
     parallelism: Parallelism,
 ) -> Result<StackReq, SizeOverflow> {
     let _ = parallelism;
-    StackReq::try_any_of([
+    if n > BLOCKING_THRESHOLD {
         StackReq::try_all_of([
-            temp_mat_req::<E>(n, 1)?,
-            temp_mat_req::<E>(n, 1)?,
-            temp_mat_req::<E>(n, 1)?,
-            temp_mat_req::<E>(n, parallelism_degree(parallelism))?,
-            temp_mat_req::<E>(n, parallelism_degree(parallelism))?,
-        ])?,
-        apply_block_householder_on_the_right_in_place_req::<E>(n, householder_blocksize, n)?,
-    ])
+            temp_mat_req::<E>(n, householder_blocksize)?,
+            temp_mat_req::<E>(n, householder_blocksize)?,
+        ])
+    } else {
+        StackReq::try_any_of([
+            StackReq::try_all_of([
+                temp_mat_req::<E>(n, 1)?,
+                temp_mat_req::<E>(n, 1)?,
+                temp_mat_req::<E>(n, 1)?,
+                temp_mat_req::<E>(n, parallelism_degree(parallelism))?,
+                temp_mat_req::<E>(n, parallelism_degree(parallelism))?,
+            ])?,
+            apply_block_householder_on_the_right_in_place_req::<E>(n, householder_blocksize, n)?,
+        ])
+    }
 }
 
 struct HessenbergFusedUpdate<'a, E: ComplexField> {
@@ -406,6 +415,27 @@ pub fn make_hessenberg_in_place<E: ComplexField>(
     assert!(a.row_stride() == 1);
 
     let n = a.nrows();
+
+    if n > BLOCKING_THRESHOLD {
+        let bs = householder.ncols();
+        let (z, stack) = temp_mat_uninit::<E>(n, bs, stack);
+
+        make_hessenberg_in_place_qgvdg_blocked(a, z, householder, parallelism, stack);
+    } else {
+        make_hessenberg_in_place_basic(a, householder, parallelism, stack);
+    }
+}
+
+fn make_hessenberg_in_place_basic<E: ComplexField>(
+    a: MatMut<'_, E>,
+    householder: MatMut<'_, E>,
+    parallelism: Parallelism,
+    stack: PodStack<'_>,
+) {
+    assert!(a.nrows() == a.ncols());
+    assert!(a.row_stride() == 1);
+
+    let n = a.nrows();
     if n < 2 {
         return;
     }
@@ -623,6 +653,315 @@ pub fn make_hessenberg_in_place<E: ComplexField>(
     }
 }
 
+fn make_hessenberg_in_place_qgvdg_unblocked<E: ComplexField>(
[…]
