### 2024-02-09 8c3c1f7
bidiag qr algorithm impl
