+++ faer-libs/faer-core/src/jacobi.rs
@@ -112,6 +112,87 @@ impl<E: RealField> JacobiRotation<E> {
         self.apply_on_the_left_in_place_arch(E::Simd::default(), x, y);
     }
 
+    #[inline(never)]
+    fn apply_on_the_left_in_place_fallback(&self, x: MatMut<'_, E>, y: MatMut<'_, E>) {
+        let Self { c, s } = *self;
+        zipped!(x, y).for_each(move |unzipped!(mut x, mut y)| {
+            let x_ = x.read();
+            let y_ = y.read();
+            x.write(c.faer_mul(x_).faer_add(s.faer_mul(y_)));
+            y.write(s.faer_neg().faer_mul(x_).faer_add(c.faer_mul(y_)));
+        });
+    }
+
+    #[inline(always)]
+    pub fn apply_on_the_right_in_place_with_simd_and_offset<S: pulp::Simd>(
+        &self,
+        simd: S,
+        offset: pulp::Offset<E::SimdMask<S>>,
+        x: MatMut<'_, E>,
+        y: MatMut<'_, E>,
+    ) {
+        self.transpose()
+            .apply_on_the_left_in_place_with_simd_and_offset(
+                simd,
+                offset,
+                x.transpose_mut(),
+                y.transpose_mut(),
+            );
+    }
+
+    #[inline(always)]
+    pub fn apply_on_the_left_in_place_with_simd_and_offset<S: pulp::Simd>(
+        &self,
+        simd: S,
+        offset: pulp::Offset<E::SimdMask<S>>,
+        x: MatMut<'_, E>,
+        y: MatMut<'_, E>,
+    ) {
+        let Self { c, s } = *self;
+        assert!(all(x.nrows() == 1, y.nrows() == 1, x.ncols() == y.ncols()));
+
+        if c == E::faer_one() && s == E::faer_zero() {
+            return;
+        }
+
+        if x.col_stride() != 1 || y.col_stride() != 1 {
+            self.apply_on_the_left_in_place_fallback(x, y);
+            return;
+        }
+
+        let simd = SimdFor::<E, S>::new(simd);
+
+        let x = SliceGroupMut::<'_, E>::new(x.transpose_mut().try_get_contiguous_col_mut(0));
+        let y = SliceGroupMut::<'_, E>::new(y.transpose_mut().try_get_contiguous_col_mut(0));
+
+        let c = simd.splat(c);
+        let s = simd.splat(s);
+
+        let (x_head, x_body, x_tail) = simd.as_aligned_simd_mut(x, offset);
+        let (y_head, y_body, y_tail) = simd.as_aligned_simd_mut(y, offset);
+
+        #[inline(always)]
+        fn process<E: RealField, S: pulp::Simd>(
+            simd: SimdFor<E, S>,
+            mut x: impl Write<Output = SimdGroupFor<E, S>>,
+            mut y: impl Write<Output = SimdGroupFor<E, S>>,
+            c: SimdGroupFor<E, S>,
+            s: SimdGroupFor<E, S>,
+        ) {
+            let zero = simd.splat(E::faer_zero());
+            let x_ = x.read_or(zero);
+            let y_ = y.read_or(zero);
+            x.write(simd.mul_add_e(c, x_, simd.mul(s, y_)));
+            y.write(simd.mul_add_e(c, y_, simd.neg(simd.mul(s, x_))));
+        }
+
+        process(simd, x_head, y_head, c, s);
+        for (x, y) in x_body.into_mut_iter().zip(y_body.into_mut_iter()) {
+            process(simd, x, y, c, s);
+        }
+        process(simd, x_tail, y_tail, c, s);
+    }
+
     #[inline]
[…]
