+++ faer-libs/faer-svd/src/bidiag_real_svd.rs
@@ -17,10 +17,25 @@ use coe::Coerce;
 use core::{iter::zip, mem::swap};
 use dyn_stack::{PodStack, SizeOverflow, StackReq};
 use faer_core::{
-    assert, jacobi::JacobiRotation, join_raw, temp_mat_req, temp_mat_uninit, temp_mat_zeroed,
-    unzipped, zipped, ComplexField, Conj, Entity, MatMut, MatRef, Parallelism, RealField,
+    assert, group_helpers::SimdFor, jacobi::JacobiRotation, join_raw, temp_mat_req,
+    temp_mat_uninit, temp_mat_zeroed, unzipped, zipped, ComplexField, Conj, Entity, MatMut, MatRef,
+    Parallelism, RealField,
 };
 use reborrow::*;
+
+#[allow(dead_code)]
+fn bidiag_to_mat<E: RealField>(diag: &[E], subdiag: &[E]) -> faer_core::Mat<E> {
+    let mut mat = faer_core::Mat::<E>::zeros(diag.len() + 1, diag.len());
+
+    for (i, d) in diag.iter().enumerate() {
+        mat.write(i, i, *d);
+    }
+    for (i, d) in subdiag.iter().enumerate() {
+        mat.write(i + 1, i, *d);
+    }
+
+    mat
+}
 
 fn norm<E: RealField>(v: MatRef<'_, E>) -> E {
     faer_core::mul::inner_prod::inner_prod_with_conj(v, Conj::No, v, Conj::No).faer_sqrt()
@@ -1137,6 +1152,348 @@ fn deflation44<E: RealField>(
     Some(rot)
 }
 
+fn bidiag_svd_qr_algorithm_impl<E: RealField>(
+    diag: &mut [E],
+    subdiag: &mut [E],
+    mut u: Option<MatMut<'_, E>>,
+    mut v: Option<MatMut<'_, E>>,
+    epsilon: E,
+    consider_zero_threshold: E,
+) {
+    let n = diag.len();
+    let max_iter = 30usize.saturating_mul(n).saturating_mul(n);
+
+    let epsilon = epsilon.faer_scale_real(E::faer_from_f64(128.0));
+
+    if let Some(mut u) = u.rb_mut() {
+        u.fill_zero();
+        u.diagonal_mut().column_vector_mut().fill(E::faer_one());
+    }
+    if let Some(mut v) = v.rb_mut() {
+        v.fill_zero();
+        v.diagonal_mut().column_vector_mut().fill(E::faer_one());
+    }
+
+    u = u.map(|u| u.submatrix_mut(0, 0, n, n));
+    v = v.map(|v| v.submatrix_mut(0, 0, n, n));
+
+    let mut max_val = E::faer_zero();
+
+    for x in &*diag {
+        let val = x.faer_abs();
+        if val > max_val {
+            max_val = val;
+        }
+    }
+    for x in &*subdiag {
+        let val = x.faer_abs();
+        if val > max_val {
+            max_val = val;
+        }
+    }
+
+    let max_val = E::faer_one();
+
+    if max_val == E::faer_zero() {
+        return;
+    }
+
+    for x in &mut *diag {
+        *x = (*x).faer_div(max_val);
+    }
+    for x in &mut *subdiag {
+        *x = (*x).faer_div(max_val);
+    }
+
+    struct Impl<'a, E: Entity> {
+        epsilon: E,
+        consider_zero_threshold: E,
+        max_iter: usize,
+        diag: &'a mut [E],
+        subdiag: &'a mut [E],
+        u: Option<MatMut<'a, E>>,
+        v: Option<MatMut<'a, E>>,
+    }
+
[…]
