+++ faer-libs/faer-svd/benches/bench.rs
@@ -111,6 +111,7 @@ fn bidiag_svd(c: &mut Criterion) {
                         Some(u.as_mut()),
                         Some(v.as_mut()),
                         4,
+                        128,
                         f64::EPSILON,
                         f64::MIN_POSITIVE,
                         Parallelism::None,
@@ -147,6 +148,7 @@ fn bidiag_svd(c: &mut Criterion) {
                         Some(u.as_mut()),
                         Some(v.as_mut()),
                         4,
+                        128,
                         f64::EPSILON,
                         f64::MIN_POSITIVE,
                         Parallelism::Rayon(0),
