### 2024-02-11 35b4375
impl more sparse matrix operations
