+++ faer-libs/faer-sparse/src/lib.rs
@@ -48,14 +48,15 @@ impl SupernodalThreshold {
     pub const AUTO: Self = Self(1.0);
 }
 
+use faer_core::sparse::util::{
+    ghost_permute_hermitian_unsorted, ghost_permute_hermitian_unsorted_symbolic,
+};
 pub use faer_core::{
     permutation::{Index, SignedIndex},
-    sparse::{adjoint, adjoint_symbolic, permute_hermitian_unsorted, transpose},
+    sparse::util::{
+        adjoint, adjoint_symbolic, ghost_adjoint_symbolic, permute_hermitian_unsorted, transpose,
+    },
     FaerError,
-};
-
-use faer_core::sparse::{
-    ghost_permute_hermitian_unsorted, ghost_permute_hermitian_unsorted_symbolic,
 };
 
 #[allow(unused_macros)]
