+++ faer-libs/faer-core/src/sparse.rs
@@ -28,7 +28,7 @@
 //! ```
 
 use super::*;
-use crate::{assert, debug_assert, group_helpers::VecGroup};
+use crate::{assert, group_helpers::VecGroup};
 use core::{cell::Cell, iter::zip, ops::Range, slice::SliceIndex};
 use dyn_stack::GlobalPodBuffer;
 use group_helpers::SliceGroup;
@@ -114,7 +114,6 @@ pub const fn repeat_byte(byte: u8) -> usize {
 ///
 /// Passing such a matrix to an algorithm that does not explicitly permit this is unspecified
 /// (though not undefined) behavior.
-#[derive(Debug)]
 pub struct SymbolicSparseColMatRef<'a, I> {
     nrows: usize,
     ncols: usize,
@@ -147,7 +146,6 @@ pub struct SymbolicSparseColMatRef<'a, I> {
 ///
 /// Passing such a matrix to an algorithm that does not explicitly permit this is unspecified
 /// (though not undefined) behavior.
-#[derive(Debug)]
 pub struct SymbolicSparseRowMatRef<'a, I> {
     nrows: usize,
     ncols: usize,
@@ -171,7 +169,7 @@ pub struct SymbolicSparseRowMatRef<'a, I> {
 /// * `nnz_per_col[j] <= col_ptrs[j+1] - col_ptrs[j]`
 /// * if `nnz_per_col` is `Some(_)`, elements of `row_indices[col_ptrs[j]..][..nnz_per_col[j]]` are
 ///   less than `nrows`
-#[derive(Debug, Clone)]
+#[derive(Clone)]
 pub struct SymbolicSparseColMat<I> {
     nrows: usize,
     ncols: usize,
@@ -195,7 +193,7 @@ pub struct SymbolicSparseColMat<I> {
 /// * `nnz_per_row[i] <= row_ptrs[i+1] - row_ptrs[i]`
 /// * if `nnz_per_row` is `Some(_)`, elements of `col_indices[row_ptrs[i]..][..nnz_per_row[i]]` are
 ///   less than `ncols`
-#[derive(Debug, Clone)]
+#[derive(Clone)]
 pub struct SymbolicSparseRowMat<I> {
     nrows: usize,
     ncols: usize,
@@ -1184,7 +1182,7 @@ impl<'a, I: Index> SymbolicSparseColMatRef<'a, I> {
         let mut mem = GlobalPodBuffer::try_new(StackReq::new::<I>(self.nrows))
             .map_err(|_| FaerError::OutOfMemory)?;
 
-        adjoint_symbolic(&mut col_ptr, &mut row_ind, *self, PodStack::new(&mut mem));
+        util::adjoint_symbolic(&mut col_ptr, &mut row_ind, *self, PodStack::new(&mut mem));
 
         let transpose = unsafe {
             SymbolicSparseColMat::new_unchecked(self.ncols, self.nrows, col_ptr, None, row_ind)
@@ -1293,6 +1291,127 @@ impl<'a, I: Index> SymbolicSparseColMatRef<'a, I> {
             .unwrap_or(__get_unchecked(self.col_ptr, j + 1).zx());
 
         start..end
+    }
+}
+
+impl<I: Index> core::fmt::Debug for SymbolicSparseColMatRef<'_, I> {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        let mat = *self;
+        let mut iter = (0..mat.ncols()).into_iter().flat_map(move |j| {
+            struct Wrapper(usize, usize);
+            impl core::fmt::Debug for Wrapper {
+                fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+                    let row = self.0;
[…]
