+++ faer-libs/faer-core/src/matrix_ops.rs
@@ -79,13 +79,13 @@ impl MatrixKind for Dense {
 }
 impl<I: Index> MatrixKind for SparseColMat<I> {
     type Ref<'a, E: Entity> = sparse::SparseColMatRef<'a, I, E>;
-    type Mut<'a, E: Entity> = ();
-    type Own<E: Entity> = ();
+    type Mut<'a, E: Entity> = sparse::SparseColMatMut<'a, I, E>;
+    type Own<E: Entity> = sparse::SparseColMat<I, E>;
 }
 impl<I: Index> MatrixKind for SparseRowMat<I> {
     type Ref<'a, E: Entity> = sparse::SparseRowMatRef<'a, I, E>;
-    type Mut<'a, E: Entity> = ();
-    type Own<E: Entity> = ();
+    type Mut<'a, E: Entity> = sparse::SparseRowMatMut<'a, I, E>;
+    type Own<E: Entity> = sparse::SparseRowMat<I, E>;
 }
 impl MatrixKind for Scale {
     type Ref<'a, E: Entity> = &'a MatScale<E>;
@@ -340,6 +340,46 @@ impl<I: Index, E: Entity> GenericMatrix for inner::SparseRowMatRef<'_, I, E> {
     #[inline(always)]
     fn as_ref(this: &Matrix<Self>) -> <Self::Kind as MatrixKind>::Ref<'_, Self::Elem> {
         *this
+    }
+}
+
+impl<I: Index, E: Entity> GenericMatrix for inner::SparseColMatMut<'_, I, E> {
+    type Kind = SparseColMat<I>;
+    type Elem = E;
+
+    #[inline(always)]
+    fn as_ref(this: &Matrix<Self>) -> <Self::Kind as MatrixKind>::Ref<'_, Self::Elem> {
+        this.rb()
+    }
+}
+
+impl<I: Index, E: Entity> GenericMatrix for inner::SparseRowMatMut<'_, I, E> {
+    type Kind = SparseRowMat<I>;
+    type Elem = E;
+
+    #[inline(always)]
+    fn as_ref(this: &Matrix<Self>) -> <Self::Kind as MatrixKind>::Ref<'_, Self::Elem> {
+        this.rb()
+    }
+}
+
+impl<I: Index, E: Entity> GenericMatrix for inner::SparseColMat<I, E> {
+    type Kind = SparseColMat<I>;
+    type Elem = E;
+
+    #[inline(always)]
+    fn as_ref(this: &Matrix<Self>) -> <Self::Kind as MatrixKind>::Ref<'_, Self::Elem> {
+        this.as_ref()
+    }
+}
+
+impl<I: Index, E: Entity> GenericMatrix for inner::SparseRowMat<I, E> {
+    type Kind = SparseRowMat<I>;
+    type Elem = E;
+
+    #[inline(always)]
+    fn as_ref(this: &Matrix<Self>) -> <Self::Kind as MatrixKind>::Ref<'_, Self::Elem> {
+        this.as_ref()
     }
 }
 
@@ -1251,6 +1291,40 @@ impl MatEq<Dense> for Dense {
     }
 }
 
+impl<I: Index> MatEq<SparseColMat<I>> for SparseColMat<I> {
+    fn mat_eq<E: ComplexField, LhsE: Conjugate<Canonical = E>, RhsE: Conjugate<Canonical = E>>(
+        lhs: KindRef<'_, LhsE, Self>,
+        rhs: KindRef<'_, RhsE, Self>,
+    ) -> bool {
+        if lhs.nrows() != rhs.nrows() || lhs.ncols() != rhs.ncols() {
+            return false;
+        }
+
+        let n = lhs.ncols();
+        let mut equal = true;
+        for j in 0..n {
[…]

[…]
