### 2024-02-11 ff226aa
reexport FaerError and Index
