+++ faer-libs/Cargo.toml
@@ -29,6 +29,7 @@ bytemuck = { version = "1", default-features = false }
 
 rand = { version = "0.8", default-features = false }
 rayon = "1"
+serde = { version = "1", features = ["derive"] }
 assert2 = "0.3"
 equator = "0.1.8"
 log = { version = "0.4", default-features = false }
