+++ faer-libs/faer-core/Cargo.toml
@@ -14,7 +14,7 @@ rust-version = "1.67.0"
 faer-entity = { workspace = true, default-features = false }
 
 seq-macro = "0.3"
-paste =  "1.0"
+paste = "1.0"
 
 coe-rs = { workspace = true }
 reborrow = { workspace = true }
@@ -29,12 +29,13 @@ bytemuck = { workspace = true }
 
 rand = { workspace = true, optional = true, default-features = false }
 rayon = { workspace = true, optional = true }
+serde = { workspace = true, optional = true }
 equator = { workspace = true }
 log = { workspace = true, optional = true, default-features = false }
 matrixcompare-core = { version = "0.1.0", optional = true }
 
 [features]
-default = ["std", "rayon"]
+default = ["std", "rayon", "serde"]
 std = [
   "faer-entity/std",
   "gemm/std",
@@ -43,17 +44,10 @@ std = [
   "num-traits/std",
   "num-complex/std",
 ]
-rayon = [
-  "std",
-  "gemm/rayon",
-  "dep:rayon",
-]
-nightly = [
-  "faer-entity/nightly",
-  "gemm/nightly",
-  "pulp/nightly",
-]
+rayon = ["std", "gemm/rayon", "dep:rayon"]
+nightly = ["faer-entity/nightly", "gemm/nightly", "pulp/nightly"]
 perf-warn = ["log"]
+serde = ["dep:serde"]
 
 [dev-dependencies]
 criterion = "0.5"
@@ -61,6 +55,7 @@ rand = "0.8.5"
 nalgebra = "0.32.3"
 assert_approx_eq = "1.1.0"
 dbgf = "0.1.1"
+serde_test = "1.0.176"
 
 [[bench]]
 name = "bench"
