### 2024-02-11 4ce8f94
npy deserialization, cholesky error location
