### 2024-02-11 0b8f338
minmize the amount of allocations performed, and also store both rows and colums
