+++ faer-libs/faer-core/src/serde_impl.rs
@@ -1,41 +1,15 @@
 //! Serde implementations for Mat
 
+use core::marker::PhantomData;
+
 use faer_entity::Entity;
-use serde::{Deserialize, Serialize, Serializer};
+use serde::{
+    de::{DeserializeSeed, SeqAccess, Visitor},
+    ser::{SerializeSeq, SerializeStruct},
+    Deserialize, Serialize, Serializer,
+};
 
 use crate::Mat;
-
-#[derive(Debug, Serialize, Deserialize)]
-#[cfg_attr(test, derive(PartialEq))]
-struct SerdeMat<E: Entity> {
-    col_length: usize,
-    data: Vec<E>,
-}
-
-impl<E: Entity> Into<Mat<E>> for SerdeMat<E> {
-    fn into(self) -> Mat<E> {
-        let row_length = self.data.len() / self.col_length;
-        Mat::from_fn(row_length, self.col_length, |i, j| {
-            self.data[j + (i * self.col_length)]
-        })
-    }
-}
-
-impl<E: Entity> Into<SerdeMat<E>> for &Mat<E> {
-    fn into(self) -> SerdeMat<E> {
-        let mut data = Vec::with_capacity(self.nrows() * self.ncols());
-        for i in 0..self.nrows() {
-            for j in 0..self.ncols() {
-                data.push(self.read(i, j));
-            }
-        }
-
-        SerdeMat {
-            col_length: self.ncols(),
-            data,
-        }
-    }
-}
 
 impl<E: Entity> Serialize for Mat<E>
 where
@@ -45,8 +19,31 @@ where
     where
         S: Serializer,
     {
-        let mat: SerdeMat<E> = self.into();
-        mat.serialize(s)
+        struct MatSequenceSerializer<'a, E: Entity>(&'a Mat<E>);
+
+        impl<'a, E: Entity> Serialize for MatSequenceSerializer<'a, E>
+        where
+            E: Serialize,
+        {
+            fn serialize<S>(&self, s: S) -> Result<S::Ok, S::Error>
+            where
+                S: Serializer,
+            {
+                let mut seq = s.serialize_seq(Some(self.0.nrows() * self.0.ncols()))?;
+                for i in 0..self.0.nrows() {
+                    for j in 0..self.0.ncols() {
+                        seq.serialize_element(&self.0.read(i, j))?;
+                    }
+                }
+                seq.end()
+            }
+        }
+
+        let mut structure = s.serialize_struct("Mat", 3)?;
+        structure.serialize_field("nrows", &self.nrows())?;
+        structure.serialize_field("ncols", &self.ncols())?;
+        structure.serialize_field("data", &MatSequenceSerializer(self))?;
+        structure.end()
     }
 }
 
@@ -58,13 +55,154 @@ where
     where
         D: serde::Deserializer<'a>,
     {
-        let mat: SerdeMat<E> = SerdeMat::<E>::deserialize(d)?;
-        if mat.data.len() % mat.col_length != 0 {
-            return Err(serde::de::Error::custom(
-                "serialized matrix is now valid as col_length isnt a divisor of the length of the data",
-            ));
+        #[derive(Deserialize)]
+        #[serde(field_identifier, rename_all = "lowercase")]
+        enum Field {
+            Nrows,
+            Ncols,
+            Data,
         }
-        Ok(mat.into())
+        const FIELDS: &'static [&'static str] = &["nrows", "ncols", "data"];
+        struct MatVisitor<E: Entity>(PhantomData<E>);
+        impl<'a, E: Entity + Deserialize<'a>> Visitor<'a> for MatVisitor<E> {
+            type Value = Mat<E>;
+
+            fn expecting(&self, formatter: &mut alloc::fmt::Formatter) -> alloc::fmt::Result {
[…]
