### 2024-02-12 00a664f
add serde feature (#98)
* add serde feature

* minmize the amount of allocations performed, and also store both rows and colums

* also erros when too large
