+++ faer-libs/faer-core/src/serde_impl.rs
@@ -1,0 +1,406 @@
+//! Serde implementations for Mat
+
+use core::marker::PhantomData;
+
+use faer_entity::Entity;
+use serde::{
+    de::{DeserializeSeed, SeqAccess, Visitor},
+    ser::{SerializeSeq, SerializeStruct},
+    Deserialize, Serialize, Serializer,
+};
+
+use crate::Mat;
+
+impl<E: Entity> Serialize for Mat<E>
+where
+    E: Serialize,
+{
+    fn serialize<S>(&self, s: S) -> Result<<S as Serializer>::Ok, <S as Serializer>::Error>
+    where
+        S: Serializer,
+    {
+        struct MatSequenceSerializer<'a, E: Entity>(&'a Mat<E>);
+
+        impl<'a, E: Entity> Serialize for MatSequenceSerializer<'a, E>
+        where
+            E: Serialize,
+        {
+            fn serialize<S>(&self, s: S) -> Result<S::Ok, S::Error>
+            where
+                S: Serializer,
+            {
+                let mut seq = s.serialize_seq(Some(self.0.nrows() * self.0.ncols()))?;
+                for i in 0..self.0.nrows() {
+                    for j in 0..self.0.ncols() {
+                        seq.serialize_element(&self.0.read(i, j))?;
+                    }
+                }
+                seq.end()
+            }
+        }
+
+        let mut structure = s.serialize_struct("Mat", 3)?;
+        structure.serialize_field("nrows", &self.nrows())?;
+        structure.serialize_field("ncols", &self.ncols())?;
+        structure.serialize_field("data", &MatSequenceSerializer(self))?;
+        structure.end()
+    }
+}
+
+impl<'a, E: Entity> Deserialize<'a> for Mat<E>
+where
+    E: Deserialize<'a>,
+{
+    fn deserialize<D>(d: D) -> Result<Self, <D as serde::Deserializer<'a>>::Error>
+    where
+        D: serde::Deserializer<'a>,
+    {
+        #[derive(Deserialize)]
+        #[serde(field_identifier, rename_all = "lowercase")]
+        enum Field {
+            Nrows,
+            Ncols,
+            Data,
+        }
+        const FIELDS: &'static [&'static str] = &["nrows", "ncols", "data"];
+        struct MatVisitor<E: Entity>(PhantomData<E>);
+        impl<'a, E: Entity + Deserialize<'a>> Visitor<'a> for MatVisitor<E> {
+            type Value = Mat<E>;
+
+            fn expecting(&self, formatter: &mut alloc::fmt::Formatter) -> alloc::fmt::Result {
+                formatter.write_str("a faer matrix")
+            }
+
+            fn visit_map<A>(self, mut map: A) -> Result<Self::Value, A::Error>
+            where
+                A: serde::de::MapAccess<'a>,
+            {
+                enum MatrixOrVec<E: Entity> {
+                    Matrix(Mat<E>),
+                    Vec(Vec<E>),
+                }
+                impl<E: Entity> MatrixOrVec<E> {
+                    fn into_mat(self, nrows: usize, ncols: usize) -> Mat<E> {
+                        match self {
+                            MatrixOrVec::Matrix(m) => m,
+                            MatrixOrVec::Vec(v) => {
+                                Mat::from_fn(nrows, ncols, |i, j| v[i * ncols + j])
+                            }
+                        }
+                    }
+                }
+                struct MatrixOrVecDeserializer<'a, E: Entity + Deserialize<'a>> {
+                    marker: PhantomData<&'a E>,
+                    nrows: Option<usize>,
+                    ncols: Option<usize>,
+                }
+                impl<'a, E: Entity + Deserialize<'a>> MatrixOrVecDeserializer<'a, E> {
[…]
