+++ faer-libs/faer-core/src/lib.rs
@@ -267,14 +267,43 @@ pub trait AsColMut<E: Entity> {
 }
 
 /// Trait for types that can be converted to a matrix view.
+///
+/// This trait is implemented for types of the matrix family, like [`Mat`],
+/// [`MatRef`], and [`MatMut`], but not for types like [`Col`], [`Row`], or
+/// their families. For a more general trait, see [`As2D`].
 pub trait AsMatRef<E: Entity> {
     /// Convert to a matrix view.
     fn as_mat_ref(&self) -> MatRef<'_, E>;
 }
 /// Trait for types that can be converted to a mutable matrix view.
+///
+/// This trait is implemented for types of the matrix family, like [`Mat`],
+/// [`MatRef`], and [`MatMut`], but not for types like [`Col`], [`Row`], or
+/// their families. For a more general trait, see [`As2D`].
 pub trait AsMatMut<E: Entity> {
     /// Convert to a mutable matrix view.
     fn as_mat_mut(&mut self) -> MatMut<'_, E>;
+}
+
+/// Trait for types that can be converted to a 2D matrix view.
+///
+/// This trait is implemented for any type that can be represented as a
+/// 2D matrix view, like [`Mat`], [`Row`], [`Col`], and their respective
+/// references and mutable references. For a trait specific to the matrix
+/// family, see [`AsMatRef`] and [`AsMatMut`].
+pub trait As2D<E: Entity> {
+    /// Convert to a 2D matrix view.
+    fn as_2d_ref(&self) -> MatRef<'_, E>;
+}
+/// Trait for types that can be converted to a mutable 2D matrix view.
+///
+/// This trait is implemented for any type that can be represented as a
+/// 2D matrix view, like [`Mat`], [`Row`], [`Col`], and their respective
+/// references and mutable references. For a trait specific to the matrix
+/// family, see [`AsMatRef`] and [`AsMatMut`].
+pub trait As2DMut<E: Entity> {
+    /// Convert to a mutable 2D matrix view.
+    fn as_2d_mut(&mut self) -> MatMut<'_, E>;
 }
 
 // AS COL
@@ -477,6 +506,228 @@ const _: () = {
         #[inline]
         fn as_mat_mut(&mut self) -> MatMut<'_, E> {
             (**self).as_mut()
+        }
+    }
+};
+
+// AS 2D
+const _: () = {
+    // Matrix family
+    impl<E: Entity> As2D<E> for &'_ MatRef<'_, E> {
+        #[inline]
+        fn as_2d_ref(&self) -> MatRef<'_, E> {
+            **self
+        }
+    }
+
+    impl<E: Entity> As2D<E> for MatRef<'_, E> {
+        #[inline]
+        fn as_2d_ref(&self) -> MatRef<'_, E> {
+            *self
+        }
+    }
+
+    impl<E: Entity> As2D<E> for &'_ MatMut<'_, E> {
+        #[inline]
+        fn as_2d_ref(&self) -> MatRef<'_, E> {
+            (**self).rb()
+        }
+    }
+
+    impl<E: Entity> As2D<E> for MatMut<'_, E> {
+        #[inline]
+        fn as_2d_ref(&self) -> MatRef<'_, E> {
+            (*self).rb()
+        }
+    }
+
+    impl<E: Entity> As2D<E> for &'_ Mat<E> {
+        #[inline]
+        fn as_2d_ref(&self) -> MatRef<'_, E> {
[…]
