### 2024-02-17 e6f395c
more sensible handling of null cases
