### 2024-02-17 d93995d
typo in docstring, added basic concat test
