### 2024-02-17 9a74c50
silly error in col! and row! handling empty case
