### 2024-02-17 d33392c
Implemented kroneckor product
