### 2024-02-17 d52ae4f
extra convenience macros (#100)
* added col! and row! macros

* more sensible handling of null cases

* added concat function, analogous to

* typo in docstring

* typo in docstring

* typo in docstring, added basic concat test

* silly error in col! and row! handling empty case

---------

Co-authored-by: DeliciousHair <matt.shepit@4c4f58.com>
