+++ faer-libs/faer-core/src/lib.rs
@@ -9655,6 +9655,68 @@ macro_rules! mat {
     };
 }
 
+/// Creates a [`Col`] containing the arguments.
+///
+/// ```
+/// use faer_core::col;
+///
+/// let col_vec = col![3.0, 5.0, 7.0, 9.0];
+///
+/// assert_eq!(col_vec.read(0), 3.0);
+/// assert_eq!(col_vec.read(1), 5.0);
+/// assert_eq!(col_vec.read(2), 7.0);
+/// assert_eq!(col_vec.read(3), 9.0);
+/// ```
+#[macro_export]
+macro_rules! col {
+    () => {
+        // TODO:
+        // is f64 the appropriate type for this situation?
+        $crate::Col::<f64>::new();
+    };
+
+    ($($v:expr),+ $(,)?) => {{
+        let data = &[$($v),+];
+        let n = data.len();
+
+        #[allow(unused_unsafe)]
+        unsafe {
+            $crate::Col::<_>::from_fn(n, |i| $crate::ref_to_ptr(&data[i]).read())
+        }
+    }};
+}
+
+/// Creates a [`Row`] containing the arguments.
+///
+/// ```
+/// use faer_core::row;
+///
+/// let row_vec = row![3.0, 5.0, 7.0, 9.0];
+///
+/// assert_eq!(row_vec.read(0), 3.0);
+/// assert_eq!(row_vec.read(1), 5.0);
+/// assert_eq!(row_vec.read(2), 7.0);
+/// assert_eq!(row_vec.read(3), 9.0);
+/// ```
+#[macro_export]
+macro_rules! row {
+    () => {
+        // TODO:
+        // is f64 the appropriate type for this situation?
+        $crate::Row::<f64>::new();
+    };
+
+    ($($v:expr),+ $(,)?) => {{
+        let data = &[$($v),+];
+        let n = data.len();
+
+        #[allow(unused_unsafe)]
+        unsafe {
+            $crate::Row::<_>::from_fn(n, |i| $crate::ref_to_ptr(&data[i]).read())
+        }
+    }};
+}
+
 /// Parallelism strategy that can be passed to most of the routines in the library.
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub enum Parallelism {
@@ -12545,6 +12607,75 @@ pub mod row {
     }
 }
 
+/// Convenience function to concatonate a nested list of matrices into a single
+/// big ['Mat']. Concatonation pattern follows the numpy.block convention that
+/// each sub-list must have an equal number of columns (net) but the boundaries
+/// do not need to align. In other words, this sort of thing:
+/// ```notcode
+///   AAAbb
+///   AAAbb
+///   cDDDD
+/// ```
+/// is perfectly acceptable.
+///
+/// TODO:
+/// function should be wrapped in a (public) macro and tests added.
+fn concat<E: ComplexField>(blocks: &[&[MatRef<'_, E>]]) -> Mat<E> {
+    #[inline(always)]
+    fn count_total_columns<E: ComplexField>(block_row: &[MatRef<'_, E>]) -> usize {
+        let mut out: usize = 0;
+        for elem in block_row.iter() {
+            out += elem.ncols();
+        }
+        out
+    }
+
+    #[inline(always)]
+    fn count_rows<E: ComplexField>(block_row: &[MatRef<'_, E>]) -> usize {
+        let mut out: usize = 0;
+        for (i, e) in block_row.iter().enumerate() {
+            if i.eq(&0) {
+                out = e.nrows();
+            } else {
[…]
