### 2024-02-17 21b7f14
Implemented As2D and As2DMut trait (#102)
* Implemented As2D and As2DMut trait

* Updated grammar
