### 2024-02-17 e896f52
Implemented Kroneckor Product (#103)
* Implemented As2D and As2DMut trait

* Updated grammar

* Implemented kroneckor product
