+++ faer-libs/faer-core/src/lib.rs
@@ -4531,6 +4531,18 @@ const _: () = {
             sum((*self).rb().as_2d())
         }
 
+        /// Kroneckor product of `self` and `rhs`.
+        ///
+        /// This is an allocating operation; see [`kron`] for the
+        /// allocation-free version or more info in general.
+        #[inline]
+        pub fn kron(&self, rhs: impl As2D<E>) -> Mat<E>
+        where
+            E: ComplexField,
+        {
+            self.as_2d_ref().kron(rhs)
+        }
+
         /// Returns a view over the matrix.
         #[inline]
         pub fn as_ref(&self) -> ColRef<'_, E> {
@@ -4974,6 +4986,18 @@ const _: () = {
             sum((*self).rb().as_2d())
         }
 
+        /// Kroneckor product of `self` and `rhs`.
+        ///
+        /// This is an allocating operation; see [`kron`] for the
+        /// allocation-free version or more info in general.
+        #[inline]
+        pub fn kron(&self, rhs: impl As2D<E>) -> Mat<E>
+        where
+            E: ComplexField,
+        {
+            self.as_2d_ref().kron(rhs)
+        }
+
         /// Returns a view over the matrix.
         #[inline]
         pub fn as_ref(&self) -> ColRef<'_, E> {
@@ -5346,6 +5370,18 @@ const _: () = {
             E: ComplexField,
         {
             sum((*self).rb().as_2d())
+        }
+
+        /// Kroneckor product of `self` and `rhs`.
+        ///
+        /// This is an allocating operation; see [`kron`] for the
+        /// allocation-free version or more info in general.
+        #[inline]
+        pub fn kron(&self, rhs: impl As2D<E>) -> Mat<E>
+        where
+            E: ComplexField,
+        {
+            self.as_2d_ref().kron(rhs)
         }
 
         /// Returns a view over the matrix.
@@ -5765,6 +5801,18 @@ const _: () = {
             E: ComplexField,
         {
             sum((*self).rb().as_2d())
+        }
+
+        /// Kroneckor product of `self` and `rhs`.
+        ///
+        /// This is an allocating operation; see [`kron`] for the
+        /// allocation-free version or more info in general.
+        #[inline]
+        pub fn kron(&self, rhs: impl As2D<E>) -> Mat<E>
+        where
+            E: ComplexField,
+        {
+            self.as_2d_ref().kron(rhs)
         }
 
         /// Returns a view over the matrix.
@@ -6633,6 +6681,27 @@ const _: () = {
             E: ComplexField,
         {
             sum((*self).rb())
+        }
+
+        /// Kroneckor product of `self` and `rhs`.
+        ///
+        /// This is an allocating operation; see [`kron`] for the
+        /// allocation-free version or more info in general.
+        #[inline]
+        pub fn kron(&self, rhs: impl As2D<E>) -> Mat<E>
+        where
+            E: ComplexField,
+        {
+            let lhs = (*self).rb();
+            let rhs = rhs.as_2d_ref();
+            let mut dst = Mat::new();
+            dst.resize_with(
+                lhs.nrows() * rhs.nrows(),
+                lhs.ncols() * rhs.ncols(),
+                |_, _| E::zeroed(),
+            );
+            kron(dst.as_mut(), lhs, rhs);
+            dst
[…]
