### 2024-02-17 c8610c8
qol changes, implement concat! macro
