+++ faer-libs/faer-core/src/lib.rs
@@ -4536,6 +4536,7 @@ const _: () = {
         /// This is an allocating operation; see [`kron`] for the
         /// allocation-free version or more info in general.
         #[inline]
+        #[track_caller]
         pub fn kron(&self, rhs: impl As2D<E>) -> Mat<E>
         where
             E: ComplexField,
@@ -4991,6 +4992,7 @@ const _: () = {
         /// This is an allocating operation; see [`kron`] for the
         /// allocation-free version or more info in general.
         #[inline]
+        #[track_caller]
         pub fn kron(&self, rhs: impl As2D<E>) -> Mat<E>
         where
             E: ComplexField,
@@ -5377,6 +5379,7 @@ const _: () = {
         /// This is an allocating operation; see [`kron`] for the
         /// allocation-free version or more info in general.
         #[inline]
+        #[track_caller]
         pub fn kron(&self, rhs: impl As2D<E>) -> Mat<E>
         where
             E: ComplexField,
@@ -5808,6 +5811,7 @@ const _: () = {
         /// This is an allocating operation; see [`kron`] for the
         /// allocation-free version or more info in general.
         #[inline]
+        #[track_caller]
         pub fn kron(&self, rhs: impl As2D<E>) -> Mat<E>
         where
             E: ComplexField,
@@ -6688,6 +6692,7 @@ const _: () = {
         /// This is an allocating operation; see [`kron`] for the
         /// allocation-free version or more info in general.
         #[inline]
+        #[track_caller]
         pub fn kron(&self, rhs: impl As2D<E>) -> Mat<E>
         where
             E: ComplexField,
@@ -7636,6 +7641,7 @@ const _: () = {
         /// This is an allocating operation; see [`kron`] for the
         /// allocation-free version or more info in general.
         #[inline]
+        #[track_caller]
         pub fn kron(&self, rhs: impl As2D<E>) -> Mat<E>
         where
             E: ComplexField,
@@ -8652,6 +8658,7 @@ impl<E: Entity> Col<E> {
     /// This is an allocating operation; see [`kron`] for the
     /// allocation-free version or more info in general.
     #[inline]
+    #[track_caller]
     pub fn kron(&self, rhs: impl As2D<E>) -> Mat<E>
     where
         E: ComplexField,
@@ -9160,6 +9167,7 @@ impl<E: Entity> Row<E> {
     /// This is an allocating operation; see [`kron`] for the
     /// allocation-free version or more info in general.
     #[inline]
+    #[track_caller]
     pub fn kron(&self, rhs: impl As2D<E>) -> Mat<E>
     where
         E: ComplexField,
@@ -9820,6 +9828,7 @@ impl<E: Entity> Mat<E> {
     /// This is an allocating operation; see [`kron`] for the
     /// allocation-free version or more info in general.
     #[inline]
+    #[track_caller]
     pub fn kron(&self, rhs: impl As2D<E>) -> Mat<E>
     where
         E: ComplexField,
@@ -10023,6 +10032,28 @@ macro_rules! mat {
     };
 }
 
+/// Concatenates the matrices in each row horizontally,
[…]
