### 2024-02-18 47ff0a5
fill up docs, reexport macros and sparse stuff
