+++ faer-libs/faer/src/lib.rs
@@ -164,6 +164,8 @@
 //!   parallelism by default.
 //! - `matrixcompare`: enabled by default. Enables macros for approximate equality checks on
 //!   matrices.
+//! - `serde`: Enables serialization and deserialization of [`Mat`].
+//! - `npy`: Enables conversions to/from numpy's matrix file format.
 //! - `perf-warn`: Produces performance warnings when matrix operations are called with suboptimal
 //! data layout.
 //! - `polars`: Enables basic interoperability with the `polars` crate.
@@ -207,8 +209,8 @@ pub mod prelude {
 }
 
 pub use faer_core::{
-    complex_native, get_global_parallelism, mat, scale, set_global_parallelism, unzipped, zipped,
-    Col, ColMut, ColRef, Mat, MatMut, MatRef, Parallelism, Row, RowMut, RowRef, Side,
+    col, complex_native, get_global_parallelism, mat, row, scale, set_global_parallelism, unzipped,
+    zipped, Col, ColMut, ColRef, Mat, MatMut, MatRef, Parallelism, Row, RowMut, RowRef, Side,
 };
 #[cfg(feature = "std")]
 #[cfg_attr(docsrs, doc(cfg(feature = "std")))]
@@ -2111,7 +2113,9 @@ pub mod sparse {
     pub use faer_core::{
         permutation::Index,
         sparse::{
-            SparseColMatRef, SparseRowMatRef, SymbolicSparseColMatRef, SymbolicSparseRowMatRef,
+            SparseColMat, SparseColMatMut, SparseColMatRef, SparseRowMat, SparseRowMatMut,
+            SparseRowMatRef, SymbolicSparseColMat, SymbolicSparseColMatRef, SymbolicSparseRowMat,
+            SymbolicSparseRowMatRef,
         },
     };
     pub use faer_sparse::{lu::LuError, FaerError};
@@ -2708,32 +2712,42 @@ pub mod sparse {
         /// Assuming `self` is a lower triangular matrix, solves the equation `self * X = rhs`, and
         /// stores the result in `rhs`.
         ///
-        /// The diagonal element is assumed to be the first stored element in each column, if the
+        /// # Note
+        /// The matrix indices need not be sorted, but
+        /// the diagonal element is assumed to be the first stored element in each column, if the
         /// matrix is column-major, or the last stored element in each row, if it is row-major.
         fn sp_solve_lower_triangular_in_place(&self, rhs: impl AsMatMut<E>);
         /// Assuming `self` is an upper triangular matrix, solves the equation `self * X = rhs`, and
         /// stores the result in `rhs`.
         ///
-        /// The diagonal element is assumed to be the last stored element in each column, if the
+        /// # Note
+        /// The matrix indices need not be sorted, but
+        /// the diagonal element is assumed to be the last stored element in each column, if the
         /// matrix is column-major, or the first stored element in each row, if it is row-major.
         fn sp_solve_upper_triangular_in_place(&self, rhs: impl AsMatMut<E>);
         /// Assuming `self` is a unit lower triangular matrix, solves the equation `self * X = rhs`,
         /// and stores the result in `rhs`.
         ///
-        /// The diagonal element is assumed to be the first stored element in each column, if the
+        /// # Note
+        /// The matrix indices need not be sorted, but
+        /// the diagonal element is assumed to be the first stored element in each column, if the
         /// matrix is column-major, or the last stored element in each row, if it is row-major.
         fn sp_solve_unit_lower_triangular_in_place(&self, rhs: impl AsMatMut<E>);
         /// Assuming `self` is a unit upper triangular matrix, solves the equation `self * X = rhs`,
[…]
