+++ faer-libs/Cargo.toml
@@ -13,13 +13,13 @@ members = [
 resolver = "2"
 
 [workspace.dependencies]
-faer-entity = { version = "0.16.0", default-features = false, path = "../faer-entity" }
+faer-entity = { version = "0.17.0", default-features = false, path = "../faer-entity" }
 
-gemm = { version = "0.17.0", default-features = false }
+gemm = { version = "0.17.1", default-features = false }
 
 coe-rs = "0.1"
 reborrow = "0.5"
-pulp = { version = "0.18.4", default-features = false }
+pulp = { version = "0.18.8", default-features = false }
 dyn-stack = { version = "0.10", default-features = false }
 
 num-traits = { version = "0.2", default-features = false }
@@ -31,7 +31,7 @@ rand = { version = "0.8", default-features = false }
 rayon = "1"
 serde = { version = "1", features = ["derive"] }
 assert2 = "0.3"
-equator = "0.1.8"
+equator = "0.1.9"
 log = { version = "0.4", default-features = false }
 
 criterion = { git = "https://github.com/sarah-ek/criterion.rs" }
