+++ faer-entity/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "faer-entity"
-version = "0.16.0"
+version = "0.17.0"
 edition = "2021"
 authors = ["sarah <>"]
 description = "Basic linear algebra routines"
@@ -14,7 +14,7 @@ rust-version = "1.67.0"
 [dependencies]
 coe-rs = "0.1"
 reborrow = "0.5"
-pulp = { version = "0.18.4", default-features = false }
+pulp = { version = "0.18.8", default-features = false }
 
 num-traits = { version = "0.2", default-features = false }
 num-complex = { version = "0.4", default-features = false }
