+++ faer-libs/faer-sparse/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "faer-sparse"
-version = "0.16.0"
+version = "0.17.0"
 edition = "2021"
 authors = ["sarah <>"]
 description = "Basic linear algebra routines"
@@ -12,10 +12,10 @@ keywords = ["math", "matrix", "linear-algebra"]
 [dependencies]
 faer-entity = { workspace = true, default-features = false }
 
-faer-core = { version = "0.16.0", default-features = false, path = "../faer-core" }
-faer-cholesky = { version = "0.16.0", default-features = false, path = "../faer-cholesky" }
-faer-qr = { version = "0.16.0", default-features = false, path = "../faer-qr" }
-faer-lu = { version = "0.16.0", default-features = false, path = "../faer-lu" }
+faer-core = { version = "0.17.0", default-features = false, path = "../faer-core" }
+faer-cholesky = { version = "0.17.0", default-features = false, path = "../faer-cholesky" }
+faer-qr = { version = "0.17.0", default-features = false, path = "../faer-qr" }
+faer-lu = { version = "0.17.0", default-features = false, path = "../faer-lu" }
 
 reborrow = { workspace = true }
 dyn-stack = { workspace = true }
