+++ faer-libs/faer-svd/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "faer-svd"
-version = "0.17.0"
+version = "0.17.1"
 edition = "2021"
 authors = ["sarah <>"]
 description = "Basic linear algebra routines"
@@ -12,8 +12,8 @@ keywords = ["math", "matrix", "linear-algebra"]
 [dependencies]
 faer-entity = { workspace = true, default-features = false }
 
-faer-core = { version = "0.17.0", default-features = false, path = "../faer-core" }
-faer-qr = { version = "0.17.0", default-features = false, path = "../faer-qr" }
+faer-core = { version = "0.17.1", default-features = false, path = "../faer-core" }
+faer-qr = { version = "0.17.1", default-features = false, path = "../faer-qr" }
 
 coe-rs = { workspace = true }
 reborrow = { workspace = true }
