+++ faer-libs/faer-core/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "faer-core"
-version = "0.17.0"
+version = "0.17.1"
 edition = "2021"
 authors = ["sarah <>"]
 description = "Basic linear algebra routines"
