+++ faer-libs/faer-lu/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "faer-lu"
-version = "0.17.0"
+version = "0.17.1"
 edition = "2021"
 authors = ["sarah <>"]
 description = "Basic linear algebra routines"
@@ -14,7 +14,7 @@ paste = "1.0.14"
 
 faer-entity = { workspace = true, default-features = false }
 
-faer-core = { version = "0.17.0", default-features = false, path = "../faer-core" }
+faer-core = { version = "0.17.1", default-features = false, path = "../faer-core" }
 
 coe-rs = { workspace = true }
 reborrow = { workspace = true }
