+++ faer-libs/Cargo.toml
@@ -31,7 +31,7 @@ rand = { version = "0.8", default-features = false }
 rayon = "1"
 serde = { version = "1", features = ["derive"] }
 assert2 = "0.3"
-equator = "0.1.9"
+equator = "0.1.10"
 log = { version = "0.4", default-features = false }
 
 criterion = { git = "https://github.com/sarah-ek/criterion.rs" }
