+++ src/linalg/triangular_solve.rs
@@ -1,0 +1,804 @@
+//! Triangular solve module.
+
+use crate::{
+    assert, debug_assert, unzipped, utils::thread::join_raw, zipped, ComplexField, Conj, Conjugate,
+    MatMut, MatRef, Parallelism,
+};
+use faer_entity::SimdCtx;
+use reborrow::*;
+
+#[inline(always)]
+fn identity<E: Clone>(x: E) -> E {
+    x.clone()
+}
+
+#[inline(always)]
+fn conj<E: ComplexField>(x: E) -> E {
+    x.faer_conj()
+}
+
+#[inline(always)]
+unsafe fn solve_unit_lower_triangular_in_place_base_case_generic_unchecked<E: ComplexField>(
+    tril: MatRef<'_, E>,
+    rhs: MatMut<'_, E>,
+    maybe_conj_lhs: impl Fn(E) -> E,
+) {
+    let n = tril.nrows();
+    match n {
+        0 | 1 => (),
+        2 => {
+            let nl10_div_l11 = maybe_conj_lhs(tril.read_unchecked(1, 0)).faer_neg();
+
+            let (_, x0, _, x1) = rhs.split_at_mut(1, 0);
+            let x0 = x0.subrows_mut(0, 1);
+            let x1 = x1.subrows_mut(0, 1);
+
+            zipped!(x0, x1).for_each(|unzipped!(x0, mut x1)| {
+                x1.write(x1.read().faer_add(nl10_div_l11.faer_mul(x0.read())));
+            });
+        }
+        3 => {
+            let nl10_div_l11 = maybe_conj_lhs(tril.read_unchecked(1, 0)).faer_neg();
+            let nl20_div_l22 = maybe_conj_lhs(tril.read_unchecked(2, 0)).faer_neg();
+            let nl21_div_l22 = maybe_conj_lhs(tril.read_unchecked(2, 1)).faer_neg();
+
+            let (_, x0, _, x1_2) = rhs.split_at_mut(1, 0);
+            let (_, x1, _, x2) = x1_2.split_at_mut(1, 0);
+            let x0 = x0.subrows_mut(0, 1);
+            let x1 = x1.subrows_mut(0, 1);
+            let x2 = x2.subrows_mut(0, 1);
+
+            zipped!(x0, x1, x2).for_each(|unzipped!(mut x0, mut x1, mut x2)| {
+                let y0 = x0.read();
+                let mut y1 = x1.read();
+                let mut y2 = x2.read();
+                y1 = y1.faer_add(nl10_div_l11.faer_mul(y0));
+                y2 = y2
+                    .faer_add(nl20_div_l22.faer_mul(y0))
+                    .faer_add(nl21_div_l22.faer_mul(y1));
+                x0.write(y0);
+                x1.write(y1);
+                x2.write(y2);
+            });
+        }
+        4 => {
+            let nl10_div_l11 = maybe_conj_lhs(tril.read_unchecked(1, 0)).faer_neg();
+            let nl20_div_l22 = maybe_conj_lhs(tril.read_unchecked(2, 0)).faer_neg();
+            let nl21_div_l22 = maybe_conj_lhs(tril.read_unchecked(2, 1)).faer_neg();
+            let nl30_div_l33 = maybe_conj_lhs(tril.read_unchecked(3, 0)).faer_neg();
+            let nl31_div_l33 = maybe_conj_lhs(tril.read_unchecked(3, 1)).faer_neg();
+            let nl32_div_l33 = maybe_conj_lhs(tril.read_unchecked(3, 2)).faer_neg();
+
+            let (_, x0, _, x1_2_3) = rhs.split_at_mut(1, 0);
[…]
