### 2024-02-18 eed262e
Improved kron performance (#105)
