### 2024-02-19 ddba518
svd: add pseudo_inverse computation
