+++ paper.bib
@@ -1,0 +1,114 @@
+@article{BLIS1,
+  author      = {Field G. {Van~Zee} and Robert A. {van~de~Geijn}},
+  title       = {{BLIS}: A Framework for Rapidly Instantiating {BLAS} Functionality},
+  journal     = {ACM Transactions on Mathematical Software},
+  volume      = {41},
+  number      = {3},
+  pages       = {14:1--14:33},
+  month       = {June},
+  year        = {2015},
+  issue_date  = {June 2015},
+  url         = {https://doi.acm.org/10.1145/2764454},
+}
+@inproceedings{10.1145/2503210.2503219,
+  author = {Wang, Qian and Zhang, Xianyi and Zhang, Yunquan and Yi, Qing},
+  title = {AUGEM: Automatically Generate High Performance Dense Linear Algebra Kernels on X86 CPUs},
+  year = {2013},
+  isbn = {9781450323789},
+  publisher = {Association for Computing Machinery},
+  address = {New York, NY, USA},
+  url = {https://doi.org/10.1145/2503210.2503219},
+  doi = {10.1145/2503210.2503219},
+  abstract = {Basic Liner algebra subprograms (BLAS) is a fundamental library in scientific computing. In this paper, we present a template-based optimization framework, AUGEM, which can automatically generate fully optimized assembly code for several dense linear algebra (DLA) kernels, such as GEMM, GEMV, AXPY and DOT, on varying multi-core CPUs without requiring any manual interference from developers. In particular, based on domain-specific knowledge about algorithms of the DLA kernels, we use a collection of parameterized code templates to formulate a number of commonly occurring instruction sequences within the optimized low-level C code of these DLA kernels. Then, our framework uses a specialized low-level C optimizer to identify instruction sequences that match the pre-defined code templates and thereby translates them into extremely efficient SSE/AVX instructions. The DLA kernels generated by our template-based approach surpass the implementations of Intel MKL and AMD ACML BLAS libraries, on both Intel Sandy Bridge and AMD Piledriver processors.},
+  booktitle = {Proceedings of the International Conference on High Performance Computing, Networking, Storage and Analysis},
+  articleno = {25},
+  numpages = {12},
+  keywords = {DLA code optimization, auto-tuning, code generation},
+  location = {Denver, Colorado},
+  series = {SC '13}
+}
+@BOOK{lapack99,
+  AUTHOR = {Anderson, E. and Bai, Z. and Bischof, C. and
+  Blackford, S. and Demmel, J. and Dongarra, J. and
+  Du Croz, J. and Greenbaum, A. and Hammarling, S. and
+  McKenney, A. and Sorensen, D.},
+  TITLE = {{LAPACK} Users' Guide},
+  EDITION = {Third},
+  PUBLISHER = {Society for Industrial and Applied Mathematics},
+  YEAR = {1999},
+  ADDRESS = {Philadelphia, PA},
+  ISBN = {0-89871-447-8 (paperback)} 
+} 
+@MISC{eigenweb,
+  author = {Ga\"{e}l Guennebaud and Beno\^{i}t Jacob and others},
+  title = {Eigen v3},
+  howpublished = {http://eigen.tuxfamily.org},
+  year = {2010}
+}
+@MISC{rayon,
+  author = {{Rayon developers}},
+  title = {Rayon},
+  howpublished = {https://github.com/rayon-rs/rayon},
+  year = {2015}
+}
+@article{10.1145/2382585.2382587,
+  author = {{Van~Zee}, Field G. and {van~de~Geijn}, Robert A. and Quintana-Ort\'{\i}, Gregorio and Elizondo, G. Joseph},
[…]
