+++ .github/workflows/code-quality.yml
@@ -24,7 +24,6 @@ jobs:
 
       - name: Formatting check
         run:
-          cd faer-libs &&
           cargo fmt --all -- --check &&
           cd ../faer-entity &&
           cargo fmt --all -- --check
@@ -33,7 +32,6 @@ jobs:
 
       - name: Linting check
         run:
-          cd faer-libs &&
           cargo clippy --all-targets &&
           cd ../faer-entity &&
           cargo clippy --all-targets
