+++ faer-bench/src/double_f64.rs
@@ -1,0 +1,855 @@
+use bytemuck::{Pod, Zeroable};
+use faer_core::pulp::Simd;
+use pulp::Scalar;
+
+#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
+#[repr(C)]
+pub struct Double<T>(pub T, pub T);
+pub type DoubleF64 = Double<f64>;
+
+unsafe impl<T: Zeroable> Zeroable for Double<T> {}
+unsafe impl<T: Pod> Pod for Double<T> {}
+
+impl<I: Iterator> Iterator for Double<I> {
+    type Item = Double<I::Item>;
+
+    #[inline(always)]
+    fn next(&mut self) -> Option<Self::Item> {
+        let x0 = self.0.next()?;
+        let x1 = self.1.next()?;
+        Some(Double(x0, x1))
+    }
+}
+
+#[inline(always)]
+fn quick_two_sum<S: Simd>(simd: S, a: S::f64s, b: S::f64s) -> (S::f64s, S::f64s) {
+    let s = simd.f64s_add(a, b);
+    let err = simd.f64s_sub(b, simd.f64s_sub(s, a));
+    (s, err)
+}
+
+#[inline(always)]
+fn two_sum<S: Simd>(simd: S, a: S::f64s, b: S::f64s) -> (S::f64s, S::f64s) {
+    let s = simd.f64s_add(a, b);
+    let bb = simd.f64s_sub(s, a);
+
+    // (a - (s - bb)) + (b - bb)
+    let err = simd.f64s_add(simd.f64s_sub(a, simd.f64s_sub(s, bb)), simd.f64s_sub(b, bb));
+    (s, err)
+}
+
+#[inline(always)]
+#[allow(dead_code)]
+fn quick_two_diff<S: Simd>(simd: S, a: S::f64s, b: S::f64s) -> (S::f64s, S::f64s) {
+    let s = simd.f64s_sub(a, b);
+    let err = simd.f64s_sub(simd.f64s_sub(a, s), b);
+    (s, err)
+}
+
+#[inline(always)]
+fn two_diff<S: Simd>(simd: S, a: S::f64s, b: S::f64s) -> (S::f64s, S::f64s) {
+    let s = simd.f64s_sub(a, b);
+    let bb = simd.f64s_sub(s, a);
+
+    // (a - (s - bb)) - (b + bb)
+    let err = simd.f64s_sub(simd.f64s_sub(a, simd.f64s_sub(s, bb)), simd.f64s_add(b, bb));
+    (s, err)
+}
+
+#[inline(always)]
+fn two_prod<S: Simd>(simd: S, a: S::f64s, b: S::f64s) -> (S::f64s, S::f64s) {
+    let p = simd.f64s_mul(a, b);
+    let err = simd.f64s_mul_add(a, b, simd.f64s_neg(p));
+
+    (p, err)
+}
+
+pub mod simd {
+    use super::*;
+
+    #[inline(always)]
+    pub fn simd_add<S: Simd>(simd: S, a: Double<S::f64s>, b: Double<S::f64s>) -> Double<S::f64s> {
+        let (s, e) = two_sum(simd, a.0, b.0);
+        let e = simd.f64s_add(e, simd.f64s_add(a.1, b.1));
+        let (s, e) = quick_two_sum(simd, s, e);
+        Double(s, e)
+    }
+
+    #[inline(always)]
+    pub fn simd_sub<S: Simd>(simd: S, a: Double<S::f64s>, b: Double<S::f64s>) -> Double<S::f64s> {
+        let (s, e) = two_diff(simd, a.0, b.0);
+        let e = simd.f64s_add(e, a.1);
[…]
