+++ src/linalg/qr/no_pivoting/compute.rs
@@ -328,7 +328,7 @@ pub fn qr_in_place<E: ComplexField>(
     ));
 
     #[cfg(feature = "perf-warn")]
-    if matrix.row_stride().unsigned_abs() != 1 && faer_core::__perf_warn!(QR_WARN) {
+    if matrix.row_stride().unsigned_abs() != 1 && crate::__perf_warn!(QR_WARN) {
         if matrix.col_stride().unsigned_abs() == 1 {
             log::warn!(target: "faer_perf", "QR prefers column-major matrix. Found row-major matrix.");
         } else {
