+++ src/linalg/qr/col_pivoting/compute.rs
@@ -745,7 +745,7 @@ pub fn qr_in_place<'out, I: Index, E: ComplexField>(
             assert!(all(col_perm.len() == n, col_perm_inv.len() == n));
 
             #[cfg(feature = "perf-warn")]
-            if matrix.row_stride().unsigned_abs() != 1 && faer_core::__perf_warn!(QR_WARN) {
+            if matrix.row_stride().unsigned_abs() != 1 && crate::__perf_warn!(QR_WARN) {
                 if matrix.col_stride().unsigned_abs() == 1 {
                     log::warn!(target: "faer_perf", "QR with column pivoting prefers column-major matrix. Found row-major matrix.");
                 } else {
