+++ src/linalg/cholesky/ldlt_diagonal/compute.rs
@@ -379,7 +379,7 @@ pub fn raw_cholesky_in_place<E: ComplexField>(
 ) -> LdltInfo {
     assert!(matrix.ncols() == matrix.nrows());
     #[cfg(feature = "perf-warn")]
-    if matrix.row_stride().unsigned_abs() != 1 && faer_core::__perf_warn!(CHOLESKY_WARN) {
+    if matrix.row_stride().unsigned_abs() != 1 && crate::__perf_warn!(CHOLESKY_WARN) {
         if matrix.col_stride().unsigned_abs() == 1 {
             log::warn!(target: "faer_perf", "LDLT prefers column-major matrix. Found row-major matrix.");
         } else {
