+++ src/linalg/solvers.rs
@@ -2211,202 +2211,6 @@ where
     }
 }
 
-/// De-serialization from common matrix file formats.
-#[cfg(feature = "std")]
-#[cfg_attr(docsrs, doc(cfg(feature = "std")))]
-pub mod io {
-    #[allow(unused_imports)]
-    use super::*;
-    #[allow(unused_imports)]
-    use complex_native::{c32, c64};
-
-    #[cfg(feature = "npy")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "npy")))]
-    pub struct Npy<'a> {
-        aligned_bytes: &'a [u8],
-        nrows: usize,
-        ncols: usize,
-        prefix_len: usize,
-        dtype: NpyDType,
-        fortran_order: bool,
-    }
-
-    #[cfg(feature = "npy")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "npy")))]
-    #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-    pub enum NpyDType {
-        F32,
-        F64,
-        C32,
-        C64,
-        Other,
-    }
-
-    #[cfg(feature = "npy")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "npy")))]
-    pub trait FromNpy: modules::core::SimpleEntity {
-        const DTYPE: NpyDType;
-    }
-
-    #[cfg(feature = "npy")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "npy")))]
-    impl FromNpy for f32 {
-        const DTYPE: NpyDType = NpyDType::F32;
-    }
-    #[cfg(feature = "npy")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "npy")))]
-    impl FromNpy for f64 {
-        const DTYPE: NpyDType = NpyDType::F64;
-    }
-    #[cfg(feature = "npy")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "npy")))]
-    impl FromNpy for c32 {
-        const DTYPE: NpyDType = NpyDType::C32;
-    }
-    #[cfg(feature = "npy")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "npy")))]
-    impl FromNpy for c64 {
-        const DTYPE: NpyDType = NpyDType::C64;
-    }
-
-    #[cfg(feature = "npy")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "npy")))]
-    impl<'a> Npy<'a> {
-        fn parse_npyz(
-            data: &[u8],
-            npyz: npyz::NpyFile<&[u8]>,
-        ) -> Result<(NpyDType, usize, usize, usize, bool), std::io::Error> {
-            let ver_major = data[6] - b'\x00';
-            let length = if ver_major <= 1 {
-                2usize
-            } else if ver_major <= 3 {
-                4usize
-            } else {
-                return Err(std::io::Error::new(
-                    std::io::ErrorKind::Other,
-                    "unsupported version",
-                ));
-            };
-            let header_len = if length == 2 {
-                u16::from_le_bytes(data[8..10].try_into().unwrap()) as usize
-            } else {
-                u32::from_le_bytes(data[8..12].try_into().unwrap()) as usize
-            };
-            let dtype = || -> NpyDType {
-                match npyz.dtype() {
-                    npyz::DType::Plain(str) => {
-                        let is_complex = match str.type_char() {
-                            npyz::TypeChar::Float => false,
-                            npyz::TypeChar::Complex => true,
-                            _ => return NpyDType::Other,
-                        };
-
-                        let byte_size = str.size_field();
-                        if byte_size == 8 && is_complex {
-                            NpyDType::C32
-                        } else if byte_size == 16 && is_complex {
[…]
