+++ src/linalg/svd/mod.rs
@@ -738,7 +738,7 @@ pub fn compute_svd_custom_epsilon<E: ComplexField>(
     #[cfg(feature = "perf-warn")]
     match (u.rb(), v.rb()) {
         (Some(matrix), _) | (_, Some(matrix)) => {
-            if matrix.row_stride().unsigned_abs() != 1 && faer_core::__perf_warn!(QR_WARN) {
+            if matrix.row_stride().unsigned_abs() != 1 && crate::__perf_warn!(QR_WARN) {
                 if matrix.col_stride().unsigned_abs() == 1 {
                     log::warn!(target: "faer_perf", "SVD prefers column-major singular vector matrices. Found row-major matrix.");
                 } else {
