+++ Cargo.toml
@@ -19,9 +19,12 @@ num-complex = { version = "0.4.5", default-features = false }
 num-traits = { version = "0.2.18", default-features = false }
 
 matrixcompare-core = { version = "0.1.0", optional = true }
+matrixcompare = { version = "0.3", optional = true }
+
 rayon = { version = "1.8.1", optional = true }
 serde = { version = "1", optional = true,  features = ["derive"] }
 log = { version = "0.4", optional = true, default-features = false }
+npyz = { version = "0.8", optional = true }
 
 [features]
 default = ["std", "rayon", "serde"]
@@ -36,6 +39,8 @@ rayon = ["std", "gemm/rayon", "dep:rayon"]
 nightly = ["faer-entity/nightly", "gemm/nightly"]
 perf-warn = ["log"]
 serde = ["dep:serde"]
+matrixcompare = ["dep:matrixcompare"]
+npy = ["std", "dep:npyz"]
 
 [dev-dependencies]
 amd = "0.2.2"
@@ -46,3 +51,7 @@ rand = "0.8.5"
 
 [profile.dev]
 opt-level = 3
+
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs", "--html-in-header", "katex-header.html"]

[…]
