### 2024-02-19 2859ef7
fix some clippy warnings
