+++ .github/workflows/run-tests.yml
@@ -35,7 +35,7 @@ jobs:
       - name: Verify 1.67.0
         run:
           cargo check &&
-          cd ../faer-entity &&
+          cd ./faer-entity &&
           cargo check
 
   testing:
