+++ .github/workflows/code-quality.yml
@@ -25,7 +25,7 @@ jobs:
       - name: Formatting check
         run:
           cargo fmt --all -- --check &&
-          cd ../faer-entity &&
+          cd ./faer-entity &&
           cargo fmt --all -- --check
         # want to get all quality issues
         continue-on-error: true
@@ -33,6 +33,6 @@ jobs:
       - name: Linting check
         run:
           cargo clippy --all-targets &&
-          cd ../faer-entity &&
+          cd ./faer-entity &&
           cargo clippy --all-targets
         continue-on-error: true
