+++ src/row/rowown.rs
@@ -11,6 +11,12 @@ use crate::{
 };
 use core::mem::ManuallyDrop;
 
+/// Heap allocated resizable row vector.
+///
+/// # Note
+///
+/// The memory layout of `Col` is guaranteed to be row-major, meaning that it has a column stride
+/// of `1`.
 #[repr(C)]
 pub struct Row<E: Entity> {
     inner: VecOwnImpl<E>,
@@ -510,7 +516,7 @@ impl<E: Entity> Row<E> {
 
     /// Kroneckor product of `self` and `rhs`.
     ///
-    /// This is an allocating operation; see [`kron`] for the
+    /// This is an allocating operation; see [`faer::linalg::kron`](crate::linalg::kron) for the
     /// allocation-free version or more info in general.
     #[inline]
     #[track_caller]
