+++ src/linalg/solvers.rs
@@ -1,6 +1,6 @@
 use crate::{
-    assert, col::*, linalg::matmul::triangular::BlockStructure, mat::*, perm::PermRef, unzipped,
-    zipped, Side, *,
+    assert, col::*, diag::DiagRef, linalg::matmul::triangular::BlockStructure, mat::*,
+    perm::PermRef, unzipped, zipped, Side, *,
 };
 use dyn_stack::*;
 use reborrow::*;
@@ -10,6 +10,7 @@ pub use crate::{
     sparse::linalg::solvers::{SpSolver, SpSolverCore, SpSolverLstsq, SpSolverLstsqCore},
 };
 
+/// Object-safe base for [`Solver`]
 pub trait SolverCore<E: Entity>: SpSolverCore<E> {
     /// Reconstructs the original matrix using the decomposition.
     fn reconstruct(&self) -> Mat<E>;
@@ -19,9 +20,12 @@ pub trait SolverCore<E: Entity>: SpSolverCore<E> {
     /// Panics if the matrix is not square.
     fn inverse(&self) -> Mat<E>;
 }
+/// Object-safe base for [`SolverLstsq`]
 pub trait SolverLstsqCore<E: Entity>: SolverCore<E> + SpSolverLstsqCore<E> {}
 
+/// Solver that can compute solution of a linear system.
 pub trait Solver<E: ComplexField>: SolverCore<E> + SpSolver<E> {}
+/// Dense solver that can compute the least squares solution of an overdetermined linear system.
 pub trait SolverLstsq<E: Entity>: SolverLstsqCore<E> + SpSolverLstsq<E> {}
 
 const _: () = {
@@ -102,6 +106,12 @@ pub struct Eigendecomposition<E: Entity> {
 }
 
 impl<E: ComplexField> Cholesky<E> {
+    /// Returns the Cholesky factorization of the input
+    /// matrix, or an error if the matrix is not positive definite.
+    ///
+    /// The factorization is such that $A = LL^H$, where $L$ is lower triangular.
+    ///
+    /// The matrix is interpreted as Hermitian, but only the provided side is accessed.
     #[track_caller]
     pub fn try_new<ViewE: Conjugate<Canonical = E>>(
         matrix: MatRef<'_, ViewE>,
@@ -151,6 +161,7 @@ impl<E: ComplexField> Cholesky<E> {
         self.factors.nrows()
     }
 
+    /// Returns the factor $L$ of the Cholesky decomposition.
     pub fn compute_l(&self) -> Mat<E> {
         let mut factor = self.factors.to_owned();
         zipped!(factor.as_mut())
@@ -247,6 +258,9 @@ impl<E: ComplexField> SolverCore<E> for Cholesky<E> {
 }
 
 impl<E: ComplexField> Lblt<E> {
+    /// Returns the Bunch-Kaufman factorization of the input matrix.
+    ///
+    /// The matrix is interpreted as Hermitian, but only the provided side is accessed.
     #[track_caller]
     pub fn new<ViewE: Conjugate<Canonical = E>>(matrix: MatRef<'_, ViewE>, side: Side) -> Self {
         assert!(matrix.nrows() == matrix.ncols());
@@ -442,6 +456,10 @@ impl<E: ComplexField> SolverCore<E> for Lblt<E> {
 }
 
 impl<E: ComplexField> PartialPivLu<E> {
+    /// Returns the LU decomposition of the input matrix with partial (row) pivoting.
+    ///
+    /// The factorization is such that $PA = LU$, where $L$ is lower triangular, $U$ is unit
+    /// upper triangular, and $P$ is the permutation arising from the pivoting.
     #[track_caller]
     pub fn new<ViewE: Conjugate<Canonical = E>>(matrix: MatRef<'_, ViewE>) -> Self {
[…]
