+++ src/linalg/lu/full_pivoting/mod.rs
@@ -6,7 +6,11 @@
 //! The full pivoting LU decomposition is more numerically stable than the one with partial
 //! pivoting, but is more expensive to compute.
 
+/// Computing the decomposition.
 pub mod compute;
+/// Reconstructing the inverse of the original matrix from the decomposition.
 pub mod inverse;
+/// Reconstructing the inverse of the original matrix from the decomposition.
 pub mod reconstruct;
+/// Solving a linear system usin the decomposition.
 pub mod solve;
