### 2024-02-20 748f5ac
add back (de)serialization and sparse matrix ops
