+++ src/linalg/mat_ops.rs
@@ -172,6 +172,123 @@ macro_rules! impl_sparse_mul {
     };
 }
 
+macro_rules! impl_partial_eq_sparse {
+    ($lhs: ty, $rhs: ty) => {
+        impl<I: Index, LhsE: Conjugate, RhsE: Conjugate<Canonical = LhsE::Canonical>>
+            PartialEq<$rhs> for $lhs
+        {
+            fn eq(&self, other: &$rhs) -> bool {
+                self.as_ref().eq(&other.as_ref())
+            }
+        }
+    };
+}
+
+macro_rules! impl_add_sub_sparse {
+    ($lhs: ty, $rhs: ty, $out: ty) => {
+        impl<
+                I: Index,
+                E: ComplexField,
+                LhsE: Conjugate<Canonical = E>,
+                RhsE: Conjugate<Canonical = E>,
+            > Add<$rhs> for $lhs
+        {
+            type Output = $out;
+            #[track_caller]
+            fn add(self, other: $rhs) -> Self::Output {
+                self.as_ref().add(other.as_ref())
+            }
+        }
+
+        impl<
+                I: Index,
+                E: ComplexField,
+                LhsE: Conjugate<Canonical = E>,
+                RhsE: Conjugate<Canonical = E>,
+            > Sub<$rhs> for $lhs
+        {
+            type Output = $out;
+            #[track_caller]
+            fn sub(self, other: $rhs) -> Self::Output {
+                self.as_ref().sub(other.as_ref())
+            }
+        }
+    };
+}
+
+macro_rules! impl_add_sub_assign_sparse {
+    ($lhs: ty, $rhs: ty) => {
+        impl<I: Index, LhsE: ComplexField, RhsE: Conjugate<Canonical = LhsE>> AddAssign<$rhs>
+            for $lhs
+        {
+            #[track_caller]
+            fn add_assign(&mut self, other: $rhs) {
+                self.as_mut().add_assign(other.as_ref())
+            }
+        }
+
+        impl<I: Index, LhsE: ComplexField, RhsE: Conjugate<Canonical = LhsE>> SubAssign<$rhs>
+            for $lhs
+        {
+            #[track_caller]
+            fn sub_assign(&mut self, other: $rhs) {
+                self.as_mut().sub_assign(other.as_ref())
+            }
+        }
+    };
+}
+
+macro_rules! impl_neg_sparse {
+    ($mat: ty, $out: ty) => {
+        impl<I: Index, E: Conjugate> Neg for $mat
+        where
+            E::Canonical: ComplexField,
+        {
+            type Output = $out;
+            #[track_caller]
+            fn neg(self) -> Self::Output {
+                self.as_ref().neg()
+            }
+        }
+    };
+}
+
+macro_rules! impl_scalar_mul_sparse {
+    ($lhs: ty, $rhs: ty, $out: ty) => {
+        impl<
+                I: Index,
[…]
