+++ Cargo.toml
@@ -48,6 +48,7 @@ assert_approx_eq = "1.1.0"
 matrix-market-rs = "0.1.3"
 matrixcompare = "0.3.0"
 rand = "0.8.5"
+serde_test = "1.0.176"
 
 [profile.dev]
 opt-level = 3
