+++ src/col/mod.rs
@@ -4,15 +4,15 @@ use faer_entity::*;
 use reborrow::*;
 
 #[repr(C)]
-struct VecImpl<E: Entity> {
-    ptr: GroupCopyFor<E, NonNull<E::Unit>>,
-    len: usize,
-    stride: isize,
+pub(crate) struct VecImpl<E: Entity> {
+    pub(crate) ptr: GroupCopyFor<E, NonNull<E::Unit>>,
+    pub(crate) len: usize,
+    pub(crate) stride: isize,
 }
 #[repr(C)]
-struct VecOwnImpl<E: Entity> {
-    ptr: GroupCopyFor<E, NonNull<E::Unit>>,
-    len: usize,
+pub(crate) struct VecOwnImpl<E: Entity> {
+    pub(crate) ptr: GroupCopyFor<E, NonNull<E::Unit>>,
+    pub(crate) len: usize,
 }
 
 impl<E: Entity> Copy for VecImpl<E> {}
@@ -53,385 +53,13 @@ pub trait AsColMut<E: Entity> {
     fn as_col_mut(&mut self) -> ColMut<'_, E>;
 }
 
-// AS COL
-const _: () = {
-    impl<E: Entity> AsColRef<E> for ColRef<'_, E> {
-        #[inline]
-        fn as_col_ref(&self) -> ColRef<'_, E> {
-            *self
-        }
-    }
-    impl<E: Entity> AsColRef<E> for &'_ ColRef<'_, E> {
-        #[inline]
-        fn as_col_ref(&self) -> ColRef<'_, E> {
-            **self
-        }
-    }
-    impl<E: Entity> AsColRef<E> for ColMut<'_, E> {
-        #[inline]
-        fn as_col_ref(&self) -> ColRef<'_, E> {
-            (*self).rb()
-        }
-    }
-    impl<E: Entity> AsColRef<E> for &'_ ColMut<'_, E> {
-        #[inline]
-        fn as_col_ref(&self) -> ColRef<'_, E> {
-            (**self).rb()
-        }
-    }
-    impl<E: Entity> AsColRef<E> for Col<E> {
-        #[inline]
-        fn as_col_ref(&self) -> ColRef<'_, E> {
-            (*self).as_ref()
-        }
-    }
-    impl<E: Entity> AsColRef<E> for &'_ Col<E> {
-        #[inline]
-        fn as_col_ref(&self) -> ColRef<'_, E> {
-            (**self).as_ref()
-        }
-    }
-
-    impl<E: Entity> AsColMut<E> for ColMut<'_, E> {
-        #[inline]
-        fn as_col_mut(&mut self) -> ColMut<'_, E> {
-            (*self).rb_mut()
-        }
-    }
-
-    impl<E: Entity> AsColMut<E> for &'_ mut ColMut<'_, E> {
-        #[inline]
-        fn as_col_mut(&mut self) -> ColMut<'_, E> {
-            (**self).rb_mut()
-        }
-    }
-
-    impl<E: Entity> AsColMut<E> for Col<E> {
-        #[inline]
-        fn as_col_mut(&mut self) -> ColMut<'_, E> {
-            (*self).as_mut()
-        }
-    }
-
-    impl<E: Entity> AsColMut<E> for &'_ mut Col<E> {
-        #[inline]
-        fn as_col_mut(&mut self) -> ColMut<'_, E> {
-            (**self).as_mut()
-        }
-    }
-};
-
-// COL INDEX
-const _: () = {
-    // RangeFull
-    // Range
-    // RangeInclusive
-    // RangeTo
-    // RangeToInclusive
-    // usize
-
-    use core::ops::RangeFull;
-    type Range = core::ops::Range<usize>;
[…]
