+++ faer-bench/src/double_f64.rs
@@ -1,11 +1,17 @@
 use bytemuck::{Pod, Zeroable};
-use faer_core::pulp::Simd;
-use pulp::Scalar;
+use faer_entity::{
+    pulp::{Scalar, Simd},
+    *,
+};
 
+// https://web.mit.edu/tabbott/Public/quaddouble-debian/qd-2.3.4-old/docs/qd.pdf
+// https://gitlab.com/hodge_star/mantis
+
+/// Value representing the implicit sum of two floating point terms, such that the absolute
+/// value of the second term is less half a ULP of the first term.
 #[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
 #[repr(C)]
 pub struct Double<T>(pub T, pub T);
-pub type DoubleF64 = Double<f64>;
 
 unsafe impl<T: Zeroable> Zeroable for Double<T> {}
 unsafe impl<T: Pod> Pod for Double<T> {}
@@ -19,6 +25,47 @@ impl<I: Iterator> Iterator for Double<I> {
         let x1 = self.1.next()?;
         Some(Double(x0, x1))
     }
+}
+
+#[inline(always)]
+#[allow(dead_code)]
+fn two_sum_e<S: Simd>(simd: S, a: S::f64s, b: S::f64s) -> (S::f64s, S::f64s) {
+    let sign_bit = simd.f64s_splat(-0.0);
+    let cmp = simd.u64s_greater_than(
+        pulp::cast(simd.f64s_or(a, sign_bit)),
+        pulp::cast(simd.f64s_or(b, sign_bit)),
+    );
+    let (a, b) = (
+        simd.m64s_select_f64s(cmp, a, b),
+        simd.m64s_select_f64s(cmp, b, a),
+    );
+
+    quick_two_sum(simd, a, b)
+}
+
+#[inline(always)]
+#[allow(dead_code)]
+fn quick_two_diff<S: Simd>(simd: S, a: S::f64s, b: S::f64s) -> (S::f64s, S::f64s) {
+    let s = simd.f64s_sub(a, b);
+    let err = simd.f64s_sub(simd.f64s_sub(a, s), b);
+    (s, err)
+}
+
+#[inline(always)]
+#[allow(dead_code)]
+fn two_diff<S: Simd>(simd: S, a: S::f64s, b: S::f64s) -> (S::f64s, S::f64s) {
+    let s = simd.f64s_sub(a, b);
+    let bb = simd.f64s_sub(s, a);
+
+    // (a - (s - bb)) - (b + bb)
+    let err = simd.f64s_sub(simd.f64s_sub(a, simd.f64s_sub(s, bb)), simd.f64s_add(b, bb));
+    (s, err)
+}
+
+#[inline(always)]
+#[allow(dead_code)]
+fn two_diff_e<S: Simd>(simd: S, a: S::f64s, b: S::f64s) -> (S::f64s, S::f64s) {
+    two_sum_e(simd, a, simd.f64s_neg(b))
 }
 
 #[inline(always)]
@@ -39,24 +86,6 @@ fn two_sum<S: Simd>(simd: S, a: S::f64s, b: S::f64s) -> (S::f64s, S::f64s) {
 }
 
 #[inline(always)]
-#[allow(dead_code)]
-fn quick_two_diff<S: Simd>(simd: S, a: S::f64s, b: S::f64s) -> (S::f64s, S::f64s) {
-    let s = simd.f64s_sub(a, b);
-    let err = simd.f64s_sub(simd.f64s_sub(a, s), b);
-    (s, err)
-}
-
-#[inline(always)]
[…]
