### 2024-02-21 df623b6
Monocrate refactor (#107)
* [wip] big refactor

* restore accidentally deleted files

* fix ci

* restore accidentally deleted files

* fix all-features build

* fix some clippy warnings

* clippy fix

* fix ci

* document stuff

* add missing import

* add back (de)serialization and sparse matrix ops

* add examples

* reorganize stuff

* fix bench build

---------

Co-authored-by: sarah <>
