+++ src/linalg/reductions/norm_l2.rs
@@ -1,0 +1,344 @@
+use super::LINEAR_IMPL_THRESHOLD;
+use crate::{
+    complex_native::*,
+    mat::MatRef,
+    utils::{simd::*, slice::*},
+};
+use faer_entity::*;
+use pulp::{Read, Simd};
+
+#[inline(always)]
+fn norm_l2_with_simd_and_offset_prologue<E: ComplexField, S: pulp::Simd>(
+    simd: S,
+    data: SliceGroup<'_, E>,
+    offset: pulp::Offset<SimdMaskFor<E, S>>,
+) -> (
+    SimdGroupFor<E::Real, S>,
+    SimdGroupFor<E::Real, S>,
+    SimdGroupFor<E::Real, S>,
+) {
+    let simd_real = SimdFor::<E::Real, S>::new(simd);
+    let simd = SimdFor::<E, S>::new(simd);
+    let half_big = simd_real.splat(E::Real::faer_min_positive_sqrt_inv());
+    let half_small = simd_real.splat(E::Real::faer_min_positive_sqrt());
+    let zero = simd.splat(E::faer_zero());
+    let zero_real = simd_real.splat(E::Real::faer_zero());
+
+    let (head, body, tail) = simd.as_aligned_simd(data, offset);
+    let (body2, body1) = body.as_arrays::<2>();
+
+    let mut acc0 = simd.abs2(head.read_or(zero));
+    let mut acc1 = zero_real;
+
+    let mut acc_small0 = simd.abs2(simd.scale_real(half_small, head.read_or(zero)));
+    let mut acc_small1 = zero_real;
+
+    let mut acc_big0 = simd.abs2(simd.scale_real(half_big, head.read_or(zero)));
+    let mut acc_big1 = zero_real;
+
+    for [x0, x1] in body2.into_ref_iter().map(RefGroup::unzip) {
+        let x0 = x0.get();
+        let x1 = x1.get();
+        acc0 = simd.abs2_add_e(x0, acc0);
+        acc1 = simd.abs2_add_e(x1, acc1);
+
+        acc_small0 = simd.abs2_add_e(simd.scale_real(half_small, x0), acc_small0);
+        acc_small1 = simd.abs2_add_e(simd.scale_real(half_small, x1), acc_small1);
+
+        acc_big0 = simd.abs2_add_e(simd.scale_real(half_big, x0), acc_big0);
+        acc_big1 = simd.abs2_add_e(simd.scale_real(half_big, x1), acc_big1);
+    }
+
+    for x0 in body1.into_ref_iter() {
+        let x0 = x0.get();
+        acc0 = simd.abs2_add_e(x0, acc0);
+        acc_small0 = simd.abs2_add_e(simd.scale_real(half_small, x0), acc_small0);
+        acc_big0 = simd.abs2_add_e(simd.scale_real(half_big, x0), acc_big0);
+    }
+
+    acc0 = simd.abs2_add_e(tail.read_or(zero), acc0);
+    acc_small0 = simd.abs2_add_e(simd.scale_real(half_small, tail.read_or(zero)), acc_small0);
+    acc_big0 = simd.abs2_add_e(simd.scale_real(half_big, tail.read_or(zero)), acc_big0);
+
+    acc0 = simd_real.add(acc0, acc1);
+    acc_small0 = simd_real.add(acc_small0, acc_small1);
+    acc_big0 = simd_real.add(acc_big0, acc_big1);
+
+    (acc_small0, acc0, acc_big0)
+}
+
+#[inline(always)]
+fn norm_l2_with_simd_and_offset_pairwise_rows<E: ComplexField, S: Simd>(
[…]

[…]
