+++ CITATION.cff
@@ -1,0 +1,21 @@
+# This CITATION.cff file was generated with cffinit.
+# Visit https://bit.ly/cffinit to generate yours today!
+
+cff-version: 1.2.0
+title: faer-rs
+message: >-
+  If you use this software, please cite it using the
+  metadata from this file.
+type: software
+authors:
+  - given-names: Sarah
+    family-names: El Kazdadi
+    email: sarah.elkazdadi@gmail.com
+    orcid: 'https://orcid.org/0000-0002-5657-0710'
+repository-code: 'https://github.com/sarah-ek/faer-rs'
+url: 'https://faer-rs.github.io/'
+abstract: A high performance linear algebra library for Rust
+keywords:
+  - math
+  - linear algebra
+license: MIT
