### 2024-02-25 24ac564
accept columns as well as matrices in solve functions
