### 2024-02-25 12ebf9c
impl more norm functions
