+++ src/row/rowown.rs
@@ -493,6 +493,15 @@ impl<E: Entity> Row<E> {
     {
         self.as_ref().as_2d().norm_max()
     }
+    /// Returns the L1 norm of `self`.
+    #[inline]
+    pub fn norm_l1(&self) -> E::Real
+    where
+        E: ComplexField,
+    {
+        self.as_ref().as_2d().norm_l1()
+    }
+
     /// Returns the L2 norm of `self`.
     #[inline]
     pub fn norm_l2(&self) -> E::Real
@@ -500,6 +509,15 @@ impl<E: Entity> Row<E> {
         E: ComplexField,
     {
         self.as_ref().as_2d().norm_l2()
+    }
+
+    /// Returns the squared L2 norm of `self`.
+    #[inline]
+    pub fn squared_norm_l2(&self) -> E::Real
+    where
+        E: ComplexField,
+    {
+        self.as_ref().as_2d().squared_norm_l2()
     }
 
     /// Returns the sum of `self`.
