+++ src/mat/matmut.rs
@@ -883,6 +883,16 @@ impl<'a, E: Entity> MatMut<'a, E> {
     {
         self.rb().norm_max()
     }
+
+    /// Returns the L1 norm of `self`.
+    #[inline]
+    pub fn norm_l1(&self) -> E::Real
+    where
+        E: ComplexField,
+    {
+        self.rb().norm_l1()
+    }
+
     /// Returns the L2 norm of `self`.
     #[inline]
     pub fn norm_l2(&self) -> E::Real
@@ -890,6 +900,15 @@ impl<'a, E: Entity> MatMut<'a, E> {
         E: ComplexField,
     {
         self.rb().norm_l2()
+    }
+
+    /// Returns the squared L2 norm of `self`.
+    #[inline]
+    pub fn squared_norm_l2(&self) -> E::Real
+    where
+        E: ComplexField,
+    {
+        self.rb().squared_norm_l2()
     }
 
     /// Returns the sum of `self`.
