### 2024-02-25 c6bf603
impl RowBatch
