### 2024-02-25 371f55c
impl to_dense for sparse matrices
