+++ src/sparse/csc/matmut.rs
@@ -86,7 +86,7 @@ impl<'a, I: Index, E: Entity> SparseColMatMut<'a, I, E> {
         (*self).rb_mut()
     }
 
-    /// Copies the current matrix into a newly allocated matrix.
+    /// Copies `self` into a newly allocated matrix.
     ///
     /// # Note
     /// Allows unsorted matrices, producing an unsorted output.
@@ -99,7 +99,17 @@ impl<'a, I: Index, E: Entity> SparseColMatMut<'a, I, E> {
         self.rb().to_owned()
     }
 
-    /// Copies the current matrix into a newly allocated matrix, with row-major order.
+    /// Copies `self` into a newly allocated dense matrix
+    #[inline]
+    pub fn to_dense(&self) -> Mat<E::Canonical>
+    where
+        E: Conjugate,
+        E::Canonical: ComplexField,
+    {
+        self.as_ref().to_dense()
+    }
+
+    /// Copies `self` into a newly allocated matrix, with row-major order.
     ///
     /// # Note
     /// Allows unsorted matrices, producing a sorted output. Duplicate entries are kept, however.
