### 2024-02-25 ac8f758
impl sort+dedup indices
