### 2024-02-25 00698a9
support duplicate entries in sparse matrices
