+++ src/sparse/csr/symbolic_ref.rs
@@ -17,11 +17,10 @@ use crate::{assert, sparse::csc::*};
 /// * if `nnz_per_row` is `Some(_)`, elements of `col_indices[row_ptrs[i]..][..nnz_per_row[i]]` are
 ///   less than `ncols`
 ///
-/// * Within each row, column indices are unique and sorted in increasing order.
+/// * Within each row, column indices are sorted in non-decreasing order.
 ///
 /// # Note
-/// Some algorithms allow working with matrices containing duplicate and/or unsorted column
-/// indicers per row.
+/// Some algorithms allow working with matrices containing unsorted row indices per column.
 ///
 /// Passing such a matrix to an algorithm that does not explicitly permit this is unspecified
 /// (though not undefined) behavior.
@@ -100,7 +99,7 @@ impl<'a, I: Index> SymbolicSparseRowMatRef<'a, I> {
                 if !col_indices.is_empty() {
                     let mut j_prev = col_indices[0];
                     for &j in &col_indices[1..] {
-                        assert!(j_prev < j);
+                        assert!(j_prev <= j);
                         j_prev = j;
                     }
                     let ncols = I::truncate(ncols);
@@ -113,7 +112,7 @@ impl<'a, I: Index> SymbolicSparseRowMatRef<'a, I> {
                 if !col_indices.is_empty() {
                     let mut j_prev = col_indices[0];
                     for &j in &col_indices[1..] {
-                        assert!(j_prev < j);
+                        assert!(j_prev <= j);
                         j_prev = j;
                     }
                     let ncols = I::truncate(ncols);
@@ -355,7 +354,7 @@ impl<I: Index> core::fmt::Debug for SymbolicSparseRowMatRef<'_, I> {
                 fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
                     let row = self.0;
                     let col = self.1;
-                    write!(f, "({row}, {col}")
+                    write!(f, "({row}, {col})")
                 }
             }
 
