+++ src/sparse/csc/matref.rs
@@ -378,23 +378,61 @@ impl<'a, I: Index, E: Entity> SparseColMatRef<'a, I, E> {
     }
 
     /// Returns a reference to the value at the given index, or None if the symbolic structure
-    /// doesn't contain it
+    /// doesn't contain it, or contains multiple indices with the given index.
     ///
     /// # Panics
-    /// Panics if `row >= self.nrows()`  
-    /// Panics if `col >= self.ncols()`  
+    /// Panics if `row >= self.nrows()`.  
+    /// Panics if `col >= self.ncols()`.  
     #[track_caller]
     pub fn get(self, row: usize, col: usize) -> Option<GroupFor<E, &'a E::Unit>> {
+        let values = self.get_all(row, col);
+        if E::faer_first(E::faer_as_ref(&values)).len() == 1 {
+            Some(E::faer_map(values, |slice| &slice[0]))
+        } else {
+            None
+        }
+    }
+
+    /// Returns a reference to a slice containing the values at the given index using a binary
+    /// search.
+    ///
+    /// # Panics
+    /// Panics if `row >= self.nrows()`.  
+    /// Panics if `col >= self.ncols()`.  
+    #[track_caller]
+    pub fn get_all(self, row: usize, col: usize) -> GroupFor<E, &'a [E::Unit]> {
         assert!(row < self.nrows());
         assert!(col < self.ncols());
 
-        let Ok(pos) = self
+        let row = I::truncate(row);
+        let start = self
             .row_indices_of_col_raw(col)
-            .binary_search(&I::truncate(row))
-        else {
-            return None;
-        };
+            .partition_point(|&p| p < row);
+        let end = start + self.row_indices_of_col_raw(col)[start..].partition_point(|&p| p <= row);
 
-        Some(E::faer_map(self.values_of_col(col), |slice| &slice[pos]))
+        E::faer_map(self.values_of_col(col), |slice| &slice[start..end])
+    }
+}
+
+impl<I: Index, E: Entity> core::fmt::Debug for SparseColMatRef<'_, I, E> {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        let mat = *self;
+        let mut iter = (0..mat.ncols()).flat_map(move |j| {
+            struct Wrapper<E: Entity>(usize, usize, E);
+            impl<E: Entity> core::fmt::Debug for Wrapper<E> {
+                fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+                    let row = self.0;
+                    let col = self.1;
+                    let val = self.2;
+                    write!(f, "({row}, {col}, {val:?})")
+                }
+            }
+
+            mat.row_indices_of_col(j)
+                .zip(SliceGroup::<'_, E>::new(mat.values_of_col(j)).into_ref_iter())
+                .map(move |(i, val)| Wrapper(i, j, val.read()))
+        });
+
+        f.debug_list().entries(&mut iter).finish()
     }
 }
