+++ src/sparse/csr/symbolic_own.rs
@@ -17,11 +17,10 @@ use crate::sparse::csc::*;
 /// * if `nnz_per_row` is `Some(_)`, elements of `col_indices[row_ptrs[i]..][..nnz_per_row[i]]` are
 ///   less than `ncols`
 ///
-/// * Within each row, column indices are unique and sorted in increasing order.
+/// * Within each row, column indices are sorted in non-decreasing order.
 ///
 /// # Note
-/// Some algorithms allow working with matrices containing duplicate and/or unsorted column
-/// indicers per row.
+/// Some algorithms allow working with matrices containing unsorted row indices per column.
 ///
 /// Passing such a matrix to an algorithm that does not explicitly permit this is unspecified
 /// (though not undefined) behavior.

[…]
