### 2024-02-26 a0f8871
impl to_sorted for sparse matrices
