+++ src/linalg/svd/mod.rs
@@ -702,8 +702,8 @@ pub fn compute_svd<E: ComplexField>(
         s,
         u,
         v,
-        E::Real::faer_epsilon().unwrap(),
-        E::Real::faer_zero_threshold().unwrap(),
+        E::Real::faer_epsilon(),
+        E::Real::faer_zero_threshold(),
         parallelism,
         stack,
         params,
