+++ src/linalg/evd/mod.rs
@@ -137,8 +137,8 @@ pub fn compute_hermitian_evd<E: ComplexField>(
         matrix,
         s,
         u,
-        E::Real::faer_epsilon().unwrap(),
-        E::Real::faer_zero_threshold().unwrap(),
+        E::Real::faer_epsilon(),
+        E::Real::faer_zero_threshold(),
         parallelism,
         stack,
         params,
@@ -371,8 +371,8 @@ pub fn compute_evd_real<E: RealField>(
         s_re,
         s_im,
         u,
-        E::faer_epsilon().unwrap(),
-        E::faer_zero_threshold().unwrap(),
+        E::faer_epsilon(),
+        E::faer_zero_threshold(),
         parallelism,
         stack,
         params,
@@ -1254,8 +1254,8 @@ pub fn compute_evd_complex<E: ComplexField>(
         matrix,
         s,
         u,
-        E::Real::faer_epsilon().unwrap(),
-        E::Real::faer_zero_threshold().unwrap(),
+        E::Real::faer_epsilon(),
+        E::Real::faer_zero_threshold(),
         parallelism,
         stack,
         params,
