### 2024-02-26 7022af4
impl try_as_slice for dense row/col
