+++ src/row/rowref.rs
@@ -471,6 +471,24 @@ impl<'a, E: Entity> RowRef<'a, E> {
         self.as_2d_ref().kron(rhs)
     }
 
+    /// Returns the row as a contiguous slice if its column stride is equal to `1`.
+    ///
+    /// # Note
+    /// The values pointed to by the references are expected to be initialized, even if the
+    /// pointed-to value is not read, otherwise the behavior is undefined.
+    pub fn try_as_slice(self) -> Option<GroupFor<E, &'a [E::Unit]>> {
+        if self.col_stride() == 1 {
+            let len = self.ncols();
+            Some(E::faer_map(
+                self.as_ptr(),
+                #[inline(always)]
+                |ptr| unsafe { core::slice::from_raw_parts(ptr, len) },
+            ))
+        } else {
+            None
+        }
+    }
+
     /// Returns a view over the matrix.
     #[inline]
     pub fn as_ref(&self) -> RowRef<'_, E> {
