### 2024-02-26 23f8924
impl try_as_slice for owned dense row/col
