+++ src/row/rowref.rs
@@ -476,6 +476,7 @@ impl<'a, E: Entity> RowRef<'a, E> {
     /// # Note
     /// The values pointed to by the references are expected to be initialized, even if the
     /// pointed-to value is not read, otherwise the behavior is undefined.
+    #[inline]
     pub fn try_as_slice(self) -> Option<GroupFor<E, &'a [E::Unit]>> {
         if self.col_stride() == 1 {
             let len = self.ncols();
