+++ src/row/rowmut.rs
@@ -472,6 +472,7 @@ impl<'a, E: Entity> RowMut<'a, E> {
     /// # Note
     /// The values pointed to by the references are expected to be initialized, even if the
     /// pointed-to value is not read, otherwise the behavior is undefined.
+    #[inline]
     pub fn try_as_slice(self) -> Option<GroupFor<E, &'a [E::Unit]>> {
         self.into_const().try_as_slice()
     }
@@ -481,6 +482,7 @@ impl<'a, E: Entity> RowMut<'a, E> {
     /// # Note
     /// The values pointed to by the references are expected to be initialized, even if the
     /// pointed-to value is not read, otherwise the behavior is undefined.
+    #[inline]
     pub fn try_as_slice_mut(self) -> Option<GroupFor<E, &'a mut [E::Unit]>> {
         if self.col_stride() == 1 {
             let len = self.ncols();
