### 2024-02-26 15b6968
clarify docs for try_as_uninit_slice_mut
