### 2024-02-26 fe18869
impl {for_each,map}_with_index
