+++ src/linalg/zip.rs
@@ -1107,6 +1107,51 @@ fn annotate_noalias_mat<Z: for<'a> MatIndex<'a>>(
 }
 
 #[inline(always)]
+fn annotate_noalias_mat_with_index<Z: for<'a> MatIndex<'a>>(
+    f: &mut impl for<'a> FnMut(usize, usize, <Z as MatIndex<'a>>::Item),
+    mut slice: Z::Slice,
+    i_begin: usize,
+    i_end: usize,
+    j: usize,
+    transpose: bool,
+    reverse_rows: bool,
+) {
+    if !transpose {
+        if !reverse_rows {
+            for i in i_begin..i_end {
+                unsafe { f(i, j, Z::get_from_slice_unchecked(&mut slice, i - i_begin)) };
+            }
+        } else {
+            for i in i_begin..i_end {
+                unsafe {
+                    f(
+                        i_begin + (i_end - i - 1),
+                        j,
+                        Z::get_from_slice_unchecked(&mut slice, i - i_begin),
+                    )
+                };
+            }
+        }
+    } else {
+        if !reverse_rows {
+            for i in i_begin..i_end {
+                unsafe { f(j, i, Z::get_from_slice_unchecked(&mut slice, i - i_begin)) };
+            }
+        } else {
+            for i in i_begin..i_end {
+                unsafe {
+                    f(
+                        j,
+                        i_begin + (i_end - i - 1),
+                        Z::get_from_slice_unchecked(&mut slice, i - i_begin),
+                    )
+                };
+            }
+        }
+    }
+}
+
+#[inline(always)]
 fn annotate_noalias_col<Z: for<'a> MatIndex<'a>>(
     f: &mut impl for<'a> FnMut(<Z as MatIndex<'a>>::Item),
     mut slice: Z::Slice,
@@ -1115,6 +1160,30 @@ fn annotate_noalias_col<Z: for<'a> MatIndex<'a>>(
 ) {
     for i in i_begin..i_end {
         unsafe { f(Z::get_from_slice_unchecked(&mut slice, i - i_begin)) };
+    }
+}
+
+#[inline(always)]
+fn annotate_noalias_col_with_index<Z: for<'a> MatIndex<'a>>(
+    f: &mut impl for<'a> FnMut(usize, <Z as MatIndex<'a>>::Item),
+    mut slice: Z::Slice,
+    i_begin: usize,
+    i_end: usize,
+    reverse: bool,
+) {
+    if !reverse {
+        for i in i_begin..i_end {
+            unsafe { f(i, Z::get_from_slice_unchecked(&mut slice, i - i_begin)) };
+        }
+    } else {
+        for i in i_begin..i_end {
+            unsafe {
+                f(
+                    i_begin + (i_end - i - 1),
+                    Z::get_from_slice_unchecked(&mut slice, i - i_begin),
+                )
+            };
+        }
     }
 }
 
@@ -1144,6 +1213,386 @@ fn for_each_mat<Z: for<'a> MatIndex<'a, Rows = usize, Cols = usize, Index = (usi
                 }
             }
         }
+    }
+}
+
+// TODO:
+// - for_each_vec_with_index
+
+#[inline(always)]
+fn for_each_mat_with_index<
+    Z: for<'a> MatIndex<
+        'a,
+        Rows = usize,
+        Cols = usize,
+        Index = (usize, usize),
+        LayoutTransform = MatLayoutTransform,
+    >,
+>(
+    z: Z,
+    mut f: impl for<'a> FnMut(usize, usize, <Z as MatIndex<'a>>::Item),
+) {
+    let layout = z.preferred_layout();
+    let mut z = z.with_layout(layout);
+
+    let m = z.nrows();
[…]
