+++ src/sparse/linalg/matmul.rs
@@ -1,0 +1,442 @@
+// PERF: optimize matmul
+// - parallelization
+// - simd(?)
+
+use super::*;
+use crate::{
+    assert,
+    utils::constrained::{self, Size},
+};
+use core::cell::UnsafeCell;
+
+/// Info about the matrix multiplication operation to help split the workload between multiple
+/// threads.
+#[derive(Clone, Debug)]
+pub struct SparseMatmulInfo {
+    flops_prefix_sum: Vec<f64>,
+}
+
+/// Performs a symbolic matrix multiplication of a sparse matrix `lhs` by a sparse matrix `rhs`,
+/// and returns the result.
+///
+/// # Note
+/// Allows unsorted matrices, and produces a sorted output.
+#[track_caller]
+pub fn sparse_sparse_matmul_symbolic<I: Index>(
+    lhs: SymbolicSparseColMatRef<'_, I>,
+    rhs: SymbolicSparseColMatRef<'_, I>,
+) -> Result<(SymbolicSparseColMat<I>, SparseMatmulInfo), FaerError> {
+    assert!(lhs.ncols() == rhs.nrows());
+
+    let m = lhs.nrows();
+    let n = rhs.ncols();
+    let mut col_ptrs = try_zeroed::<I>(n + 1)?;
+    let mut row_ind = Vec::new();
+
+    let mut work = try_collect((0..m).into_iter().map(|_| I::truncate(usize::MAX)))?;
+    let mut info = SparseMatmulInfo {
+        flops_prefix_sum: try_zeroed::<f64>(n + 1)?,
+    };
+
+    for j in 0..n {
+        let mut count = 0usize;
+        let mut flops = 0.0f64;
+        for k in rhs.row_indices_of_col(j) {
+            for i in lhs.row_indices_of_col(k) {
+                if work[i] != I::truncate(j) {
+                    row_ind.try_reserve(1).map_err(|_| FaerError::OutOfMemory)?;
+                    row_ind.push(I::truncate(i));
+                    work[i] = I::truncate(j);
+                    count += 1;
+                }
+            }
+            flops += lhs.row_indices_of_col_raw(k).len() as f64;
+        }
+        info.flops_prefix_sum[j + 1] = info.flops_prefix_sum[j] + flops;
+
+        col_ptrs[j + 1] = col_ptrs[j] + I::truncate(count);
+        if col_ptrs[j + 1] > I::from_signed(I::Signed::MAX) {
+            return Err(FaerError::IndexOverflow);
+        }
+        row_ind[col_ptrs[j].zx()..col_ptrs[j + 1].zx()].sort_unstable();
+    }
+    row_ind.shrink_to_fit();
+
+    unsafe {
+        Ok((
+            SymbolicSparseColMat::new_unchecked(m, n, col_ptrs, None, row_ind),
+            info,
+        ))
+    }
+}
+
+/// Computes the size and alignment of the workspace required to perform the numeric matrix
+/// multiplication into `dst`.
+pub fn sparse_sparse_matmul_numeric_req<I: Index, E: ComplexField>(
+    dst: SymbolicSparseColMatRef<'_, I>,
+    parallelism: Parallelism,
+) -> Result<StackReq, SizeOverflow> {
+    crate::linalg::temp_mat_req::<E>(
+        dst.nrows(),
+        crate::utils::thread::parallelism_degree(parallelism),
+    )
+}
+
+/// Performs a numeric matrix multiplication of a sparse matrix `lhs` by a sparse matrix `rhs`
+/// multiplied by `k`, and stores the result in `dst`.
+///
+/// # Note
[…]
