+++ src/lib.rs
@@ -181,6 +181,7 @@ pub mod linalg;
 
 pub mod complex_native;
 
+/// Similar to the [`dbg`] macro, but takes a format spec as a first parameter.
 pub use dbgf::dbgf;
 pub use dyn_stack;
 #[cfg(feature = "std")]
@@ -191,15 +192,15 @@ pub use reborrow;
 /// Various utilities for low level implementations in generic code.
 pub mod utils;
 
-/// Contiguous resizable column vector type.
+/// Column vector type.
 pub mod col;
-/// Contiguous resizable diagonal matrix type.
+/// Diagonal matrix type.
 pub mod diag;
-/// Contiguous resizable matrix type.
+/// Matrix type.
 pub mod mat;
 /// Permutation matrices.
 pub mod perm;
-/// Contiguous resizable row vector type.
+/// Row vector type.
 pub mod row;
 /// Sparse data structures and algorithms.
 pub mod sparse;
@@ -211,7 +212,7 @@ pub use row::{Row, RowMut, RowRef};
 mod seal;
 mod sort;
 
-pub use faer_entity::{ComplexField, Conjugate, Entity, RealField};
+pub use faer_entity::{ComplexField, Conjugate, Entity, RealField, SimpleEntity};
 
 /// Specifies whether the triangular lower or upper part of a matrix should be accessed.
 #[derive(Copy, Clone, Debug, PartialEq)]
@@ -897,8 +898,185 @@ pub mod prelude {
     pub use crate::{
         complex_native::{c32, c64},
         linalg::solvers::{Solver, SolverLstsq, SpSolver, SpSolverLstsq},
-        Col, ColMut, ColRef, Mat, MatMut, MatRef, Row, RowMut, RowRef,
+        unzipped, zipped, Col, ColMut, ColRef, Mat, MatMut, MatRef, Row, RowMut, RowRef,
     };
+}
+
+/// Matrix solvers and decompositions.
+#[deprecated = "moved to faer::linalg::solvers"]
+pub mod solvers {
+    pub use crate::linalg::solvers::*;
+}
+
+/// Re-exports.
+#[deprecated]
+pub mod modules {
+    /// Emulation layer for `faer_core`
+    #[deprecated]
+    pub mod core {
+        #[deprecated = "moved to faer::col"]
+        pub use crate::col;
+
+        #[deprecated = "moved to faer::complex_native"]
+        pub use crate::complex_native;
+
+        #[deprecated = "moved to faer::utils::constrained"]
+        pub use crate::utils::constrained;
+
+        #[deprecated = "moved to faer::utils::{slice, simd, vec}"]
+        pub use crate::utils as group_helpers;
+
+        #[deprecated = "moved to faer::linalg::householder"]
+        pub use crate::linalg::householder;
+
+        #[deprecated = "moved to faer::linalg::triangular_inverse"]
+        pub use crate::linalg::triangular_inverse as inverse;
+
+        #[deprecated = "moved to faer::linalg::triangular_solve"]
+        pub use crate::linalg::triangular_inverse as solve;
+
+        #[deprecated = "moved to faer::mat"]
+        pub use crate::mat;
+
+        #[deprecated = "moved to faer::linalg::matmul"]
+        pub use crate::linalg::matmul as mul;
+
+        #[deprecated = "moved to faer::perm"]
+        pub use crate::perm as permutation;
+
+        #[deprecated = "moved to faer::row"]
+        pub use crate::row;
+
+        #[deprecated = "moved to faer::sparse"]
+        pub use crate::sparse;
+
+        #[deprecated = "moved to faer::linalg::zip"]
+        pub use crate::linalg::zip;
+
+        #[deprecated = "moved to faer::zipped"]
[…]
