### 2024-02-28 b62d42f
add rand support
