+++ src/stats/mod.rs
@@ -1,0 +1,227 @@
+use crate::{Col, ComplexField, Mat, Row};
+use rand::distributions::Distribution;
+use rand_distr::{Standard, StandardNormal};
+
+/// The normal distribution, `N(mean, std_dev**2)`.
+pub struct Normal<E: ComplexField> {
+    mean: E,
+    std_dev: E::Real,
+}
+
+/// The normal distribution, `N(mean, std_dev**2)` for `0 <= i < nrows`, `0 <= j < ncols`.
+pub struct NormalMat<E: ComplexField> {
+    /// Number of rows of the sampled matrix.
+    pub nrows: usize,
+    /// Number of columns of the sampled matrix.
+    pub ncols: usize,
+    /// Normal distribution parameters for a single scalar.
+    pub normal: Normal<E>,
+}
+
+/// The standard normal distribution, `N(0, 1)` for `0 <= i < nrows`, `0 <= j < ncols`.
+pub struct StandardNormalMat {
+    /// Number of rows of the sampled matrix.
+    pub nrows: usize,
+    /// Number of columns of the sampled matrix.
+    pub ncols: usize,
+}
+
+/// The standard distribution. Samples uniformly distributed values for `0 <= i < nrows`, `0 <= j
+/// < ncols`.
+pub struct StandardMat {
+    /// Number of rows of the sampled matrix.
+    pub nrows: usize,
+    /// Number of columns of the sampled matrix.
+    pub ncols: usize,
+}
+
+/// The normal distribution, `N(mean, std_dev**2)` for `0 <= j < ncols`.
+pub struct NormalRow<E: ComplexField> {
+    /// Number of columns of the sampled row.
+    pub ncols: usize,
+    /// Normal distribution parameters for a single scalar.
+    pub normal: Normal<E>,
+}
+
+/// The standard normal distribution, `N(0, 1)` for `0 <= j < ncols`.
+pub struct StandardNormalRow {
+    /// Number of columns of the sampled row.
+    pub ncols: usize,
+}
+
+/// The standard distribution. Samples uniformly distributed values for `0 <= j < ncols`.
+pub struct StandardRow {
+    /// Number of columns of the sampled row.
+    pub ncols: usize,
+}
+
+/// The normal distribution, `N(mean, std_dev**2)` for `0 <= i < nrows`.
+pub struct NormalCol<E: ComplexField> {
+    /// Number of rows of the sampled column.
+    pub nrows: usize,
+    /// Normal distribution parameters for a single scalar.
+    pub normal: Normal<E>,
+}
+
+/// The standard normal distribution, `N(0, 1)` for `0 <= i < nrows`.
+pub struct StandardNormalCol {
+    /// Number of rows of the sampled column.
+    pub nrows: usize,
+}
+
+/// The standard distribution. Samples uniformly distributed values for `0 <= i < nrows`.
+pub struct StandardCol {
+    /// Number of rows of the sampled column.
+    pub nrows: usize,
+}
+
+/// Uniformly samples a unitary matrix from the unitary group, in the sense of the [Haar measure](https://en.wikipedia.org/wiki/Haar_measure).
+pub struct UnitaryMat {
+    /// Dimension of the sampled matrix.
+    pub dimension: usize,
+}
+
+impl<E: ComplexField> Normal<E> {
+    /// Construct, from dimensions, mean and standard deviation.
+    ///
+    /// Parameters:
+    /// - mean (`μ`, unrestricted)
+    /// - standard deviation (`σ`, must be finite)
[…]
