+++ Cargo.toml
@@ -25,6 +25,8 @@ rayon = { version = "1.8.1", optional = true }
 serde = { version = "1", optional = true,  features = ["derive"] }
 log = { version = "0.4", optional = true, default-features = false }
 npyz = { version = "0.8", optional = true }
+rand = { version = "0.8.5", default-features = false, optional = true }
+rand_distr = { version = "0.4.3", default-features = false, optional = true }
 
 [features]
 default = ["std", "rayon", "serde"]
@@ -36,6 +38,7 @@ std = [
   "num-traits/std",
   "num-complex/std",
 ]
+rand = ["dep:rand", "rand_distr"]
 rayon = ["std", "gemm/rayon", "dep:rayon"]
 nightly = ["faer-entity/nightly", "gemm/nightly"]
 perf-warn = ["log"]
