### 2024-02-28 6f0006e
document sparse cholesky module
