+++ src/sparse/linalg/cholesky.rs
@@ -6,7 +6,6 @@
 //! # Note
 //! The functions in this module accept unsorted input, producing a sorted decomposition factor
 //! (simplicial).
-#![allow(missing_docs)]
 
 // implementation inspired by https://gitlab.com/hodge_star/catamari
 
@@ -49,12 +48,30 @@ enum Ordering<'a, I: Index> {
     ),
 }
 
+/// Simplicial factorization module.
+///
+/// A simplicial factorization is one that processes the elements of the Cholesky factor of the
+/// input matrix one by one, rather than by blocks. This is more efficient if the Cholesky factor is
+/// very sparse.
 pub mod simplicial {
     use super::*;
     use crate::assert;
 
+    /// Computes the size and alignment of the workspace required to compute the elimination tree
+    /// and column counts of a matrix of size `n` with `nnz` non-zero entries.
+    pub fn prefactorize_symbolic_cholesky_req<I: Index>(
+        n: usize,
+        nnz: usize,
+    ) -> Result<StackReq, SizeOverflow> {
+        _ = nnz;
+        StackReq::try_new::<I>(n)
+    }
+
     /// Computes the elimination tree and column counts of the Cholesky factorization of the matrix
     /// `A`.
+    ///
+    /// # Note
+    /// Only the upper triangular part of `A` is analyzed.
     pub fn prefactorize_symbolic_cholesky<'out, I: Index>(
         etree: &'out mut [I::Signed],
         col_counts: &mut [I],
@@ -126,12 +143,23 @@ pub mod simplicial {
         unsafe { Idx::from_slice_ref_unchecked(stack) }
     }
 
+    /// Computes the size and alignment of the workspace required to compute the symbolic
+    /// Cholesky factorization of a square matrix with size `n`.
     pub fn factorize_simplicial_symbolic_req<I: Index>(n: usize) -> Result<StackReq, SizeOverflow> {
         let n_req = StackReq::try_new::<I>(n)?;
         StackReq::try_all_of([n_req, n_req, n_req])
     }
 
-    pub fn factorize_simplicial_symbolic_cholesky<I: Index>(
+    /// Computes the symbolic structure of the Cholesky factor of the matrix `A`.
+    ///
+    /// # Note
+    /// Only the upper triangular part of `A` is analyzed.
+    ///
+    /// # Panics
+    /// The elimination tree and column counts must be computed by calling
+    /// [`prefactorize_symbolic_cholesky`] with the same matrix. Otherwise, the behavior is
+    /// unspecified and panics may occur.
+    pub fn factorize_simplicial_symbolic<I: Index>(
         A: SymbolicSparseColMatRef<'_, I>,
         etree: EliminationTreeRef<'_, I>,
         col_counts: &[I],
@@ -571,6 +599,16 @@ pub mod simplicial {
         )
     }
 
+    /// Computes the numeric values of the Cholesky LLT factor of the matrix `A`, and stores them in
+    /// `L_values`.
+    ///
+    /// # Note
+    /// Only the upper triangular part of `A` is accessed.
+    ///
+    /// # Panics
+    /// The symbolic structure must be computed by calling
+    /// [`factorize_simplicial_symbolic`] on a matrix with the same symbolic structure.
+    /// Otherwise, the behavior is unspecified and panics may occur.
     pub fn factorize_simplicial_numeric_llt<I: Index, E: ComplexField>(
         L_values: GroupFor<E, &mut [E::Unit]>,
         A: SparseColMatRef<'_, I, E>,
@@ -592,6 +630,17 @@ pub mod simplicial {
         )
     }
 
+    /// Computes the row indices and  numeric values of the Cholesky LLT factor of the matrix `A`,
+    /// and stores them in `L_row_indices` and `L_values`.
+    ///
+    /// # Note
[…]
