### 2024-02-28 c708108
document sparse qr module
