+++ src/sparse/linalg/qr.rs
@@ -75,7 +75,20 @@ pub(crate) fn ghost_col_etree<'n, I: Index>(
     }
 }
 
-/// workspace: I×(A.nrows() + A.ncols())
+/// Computes the size and alignment of the workspace required to compute the column elimination tree
+/// of a matrix $A$ with dimensions `(nrows, ncols)`.
+#[inline]
+pub fn col_etree_req<I: Index>(nrows: usize, ncols: usize) -> Result<StackReq, SizeOverflow> {
+    StackReq::try_all_of([
+        StackReq::try_new::<I>(nrows)?,
+        StackReq::try_new::<I>(ncols)?,
+    ])
+}
+
+/// Computes the column elimination tree of $A$, which is the same as the elimination tree of $A^H
+/// A$.
+///
+/// `etree` has length `A.ncols()`
 #[inline]
 pub fn col_etree<'out, I: Index>(
     A: SymbolicSparseColMatRef<'_, I>,
@@ -263,7 +276,30 @@ pub(crate) fn ghost_column_counts_aat<'m, 'n, I: Index>(
     }
 }
 
-/// workspace: I×(5×A.nrows() + A.ncols())
+/// Computes the size and alignment of the workspace required to compute the column counts
+/// of the Cholesky factor of the matrix $A A^H$, where $A$ has dimensions `(nrows, ncols)`.
+#[inline]
+pub fn column_counts_aat_req<I: Index>(
+    nrows: usize,
+    ncols: usize,
+) -> Result<StackReq, SizeOverflow> {
+    StackReq::try_all_of([
+        StackReq::try_new::<I>(nrows)?,
+        StackReq::try_new::<I>(nrows)?,
+        StackReq::try_new::<I>(nrows)?,
+        StackReq::try_new::<I>(nrows)?,
+        StackReq::try_new::<I>(nrows)?,
+        StackReq::try_new::<I>(ncols)?,
+    ])
+}
+
+/// Computes the column counts of the Cholesky factor of $A A^H$.
+///
+/// - `col_counts` has length `A.nrows()`.
+/// - `min_row` has length `A.ncols()`.
+/// - `row_perm` has length `A.nrows()`: fill reducing permutation.
+/// - `etree` has length `A.nrows()`: column elimination tree of $A A^H$.
+/// - `post` has length `A.nrows()`: postordering of `etree`.
 pub fn column_counts_aat<'m, 'n, I: Index>(
     col_counts: &mut [I],
     min_row: &mut [I],
@@ -287,17 +323,35 @@ pub fn column_counts_aat<'m, 'n, I: Index>(
     })
 }
 
-/// workspace: I×(3×n)
+/// Computes the size and alignment of the workspace required to compute the postordering of an
+/// elimination tree of size `n`.
+#[inline]
+pub fn postorder_req<I: Index>(n: usize) -> Result<StackReq, SizeOverflow> {
+    StackReq::try_all_of([StackReq::try_new::<I>(n)?; 3])
+}
+
+/// Computes a postordering of the elimination tree of size `n`.
+#[inline]
 pub fn postorder<I: Index>(post: &mut [I], etree: EliminationTreeRef<'_, I>, stack: PodStack<'_>) {
     Size::with(etree.inner.len(), |N| {
         ghost_postorder(Array::from_mut(post, N), etree.ghost_inner(N), stack)
     })
 }
 
+/// Supernodal factorization module.
+///
+/// A supernodal factorization is one that processes the elements of the QR factors of the
+/// input matrix by blocks, rather than one by one. This is more efficient if the QR factors are
+/// somewhat dense.
 pub mod supernodal {
[…]
