### 2024-02-29 ec32eb2
document sparse lu module
