+++ src/sparse/linalg/lu.rs
@@ -6,7 +6,6 @@
 //! # Warning
 //! The functions in this module accept unsorted input, and always produce unsorted decomposition
 //! factors.
-#![allow(missing_docs)]
 
 use super::{
     cholesky::simplicial::EliminationTreeRef,
@@ -93,10 +92,16 @@ fn resize_index<I: Index>(
     Ok(())
 }
 
+/// Supernodal factorization module.
+///
+/// A supernodal factorization is one that processes the elements of the LU factors of the
+/// input matrix by blocks, rather than by single elements. This is more efficient if the LU
+/// factors are somewhat dense.
 pub mod supernodal {
     use super::*;
     use crate::{assert, sparse::linalg::try_collect};
 
+    /// LU factor structure containing the symbolic structure.
     #[derive(Debug, Clone)]
     pub struct SymbolicSupernodalLu<I> {
         pub(super) supernode_ptr: alloc::vec::Vec<I>,
@@ -108,6 +113,7 @@ pub mod supernodal {
         pub(super) ncols: usize,
     }
 
+    /// LU factor structure containing the symbolic and numerical representations.
     #[derive(Debug, Clone)]
     pub struct SupernodalLu<I, E: Entity> {
         nrows: usize,
@@ -134,6 +140,7 @@ pub mod supernodal {
     }
 
     impl<I: Index, E: Entity> SupernodalLu<I, E> {
+        /// Creates a new Supernodal LU of a $0\times 0$ matrix.
         #[inline]
         pub fn new() -> Self {
             Self {
@@ -157,21 +164,30 @@ pub mod supernodal {
             }
         }
 
+        /// Returns the number of rows of $A$.
         #[inline]
         pub fn nrows(&self) -> usize {
             self.nrows
         }
 
+        /// Returns the number of columns of $A$.
         #[inline]
         pub fn ncols(&self) -> usize {
             self.ncols
         }
 
+        /// Returns the number of supernodes.
         #[inline]
         pub fn n_supernodes(&self) -> usize {
             self.nsupernodes
         }
 
+        /// Solves the equation $\text{Op}(A) x = \text{rhs}$ and stores the result in `rhs`, where
+        /// $\text{Op}$ is either the identity or the conjugate, depending on the value of `conj`.
+        ///
+        /// # Panics
+        /// - Panics if `self.nrows() != self.ncols()`.
+        /// - Panics if `rhs.nrows() != self.nrows()`.
         #[track_caller]
         pub fn solve_in_place_with_conj(
             &self,
@@ -197,6 +213,13 @@ pub mod supernodal {
             crate::perm::permute_rows(X.rb_mut(), temp.rb(), col_perm.inverse());
         }
 
+        /// Solves the equation $\text{Op}(A^\top) x = \text{rhs}$ and stores the result in `rhs`,
+        /// where $\text{Op}$ is either the identity or the conjugate, depending on the
+        /// value of `conj`.
+        ///
+        /// # Panics
+        /// - Panics if `self.nrows() != self.ncols()`.
+        /// - Panics if `rhs.nrows() != self.nrows()`.
         #[track_caller]
         pub fn solve_transpose_in_place_with_conj(
             &self,
@@ -530,6 +553,8 @@ pub mod supernodal {
         }
     }
 
+    /// Computes the size and alignment of the workspace required to compute the symbolic
+    /// LU factorization of a square matrix with size `n`.
     pub fn factorize_supernodal_symbolic_lu_req<I: Index>(
[…]
