### 2024-02-29 3db95f8
support duplicate entries in amd
