### 2024-02-29 0c0cfad
add FIXME note for sparse triangular solve
