+++ src/sparse/linalg/cholesky.rs
@@ -6,7 +6,521 @@
 //! # Note
 //! The functions in this module accept unsorted input, producing a sorted decomposition factor
 //! (simplicial).
-
+//!
+//! # Example (low level API)
+//! Simplicial:
+//! ```
+//! fn simplicial_cholesky() -> Result<(), faer::sparse::FaerError> {
+//!     use faer::{
+//!         assert_matrix_eq,
+//!         dyn_stack::{GlobalPodBuffer, PodStack, StackReq},
+//!         reborrow::*,
+//!         sparse::{
+//!             linalg::{
+//!                 amd,
+//!                 cholesky::{simplicial, LdltRegularization, LltRegularization},
+//!             },
+//!             CreationError, SparseColMat, SymbolicSparseColMat,
+//!         },
+//!     };
+//!     use rand::prelude::*;
+//!
+//!     // The simplicial Cholesky api takes an upper triangular matrix as input, to be
+//!     // interpreted as self-adjoint.
+//!     let dim = 4;
+//!     let A_upper = match SparseColMat::<usize, f64>::try_new_from_triplets(
+//!         dim,
+//!         dim,
+//!         &[
+//!             // diagonal entries
+//!             (0, 0, 10.0),
+//!             (1, 1, 11.0),
+//!             (2, 2, 12.0),
+//!             (3, 3, 13.0),
+//!             // non diagonal entries
+//!             (0, 1, 1.0),
+//!             (0, 3, 1.5),
+//!             (1, 3, -3.2),
+//!         ],
+//!     ) {
+//!         Ok(A) => Ok(A),
+//!         Err(CreationError::Generic(err)) => Err(err),
+//!         Err(CreationError::OutOfBounds { .. }) => panic!(),
+//!     }?;
+//!
+//!     let mut A = faer::sparse::ops::add(
+//!         A_upper.as_ref(),
+//!         A_upper.to_row_major()?.as_ref().transpose(),
+//!     )?;
+//!     for i in 0..dim {
+//!         A[(i, i)] /= 2.0;
+//!     }
+//!
+//!     let A_nnz = A_upper.compute_nnz();
+//!     let mut rhs = Vec::new();
+//!     let mut rng = StdRng::seed_from_u64(0);
+//!     rhs.try_reserve_exact(dim)?;
+//!     rhs.resize_with(dim, || rng.gen::<f64>());
+//!
+//!     let mut sol = Vec::new();
+//!     sol.try_reserve_exact(dim)?;
+//!     sol.resize(dim, 0.0f64);
+//!
+//!     let rhs = faer::mat::from_column_major_slice::<f64>(&rhs, dim, 1);
+//!     let mut sol = faer::mat::from_column_major_slice_mut::<f64>(&mut sol, dim, 1);
+//!
+//!     // Optional: fill reducing permutation
+//!     let (perm, perm_inv) = {
+//!         let mut perm = Vec::new();
+//!         let mut perm_inv = Vec::new();
+//!         perm.try_reserve_exact(dim)?;
+//!         perm_inv.try_reserve_exact(dim)?;
+//!         perm.resize(dim, 0usize);
+//!         perm_inv.resize(dim, 0usize);
+//!
+//!         let mut mem = GlobalPodBuffer::try_new(amd::order_req::<usize>(dim, A_nnz)?)?;
+//!         amd::order(
+//!             &mut perm,
+//!             &mut perm_inv,
+//!             A_upper.symbolic(),
+//!             amd::Control::default(),
+//!             PodStack::new(&mut mem),
+//!         )?;
+//!
+//!         (perm, perm_inv)
+//!     };
+//!
+//!     let perm = unsafe { faer::perm::PermRef::new_unchecked(&perm, &perm_inv) };
+//!
+//!     let A_perm_upper = {
[…]
