### 2024-03-01 d3cb2f6
svd: add pseudo_inverse computation (#106)
