### 2024-03-01 1676da9
optimize pseudoinverse
