### 2024-03-01 d5705cb
thin svd pseudoinverse
