### 2024-03-02 854e164
only read the lower triangular part of A_lower in sparse cholesky, impl l_[transpose_]solve
