+++ src/linalg/solvers.rs
@@ -48,24 +48,24 @@ pub struct Cholesky<E: Entity> {
 pub struct Lblt<E: Entity> {
     factors: Mat<E>,
     subdiag: Mat<E>,
-    perm: Vec<usize>,
-    perm_inv: Vec<usize>,
+    perm: alloc::vec::Vec<usize>,
+    perm_inv: alloc::vec::Vec<usize>,
 }
 
 /// LU decomposition with partial pivoting.
 pub struct PartialPivLu<E: Entity> {
     pub(crate) factors: Mat<E>,
-    row_perm: Vec<usize>,
-    row_perm_inv: Vec<usize>,
+    row_perm: alloc::vec::Vec<usize>,
+    row_perm_inv: alloc::vec::Vec<usize>,
     n_transpositions: usize,
 }
 /// LU decomposition with full pivoting.
 pub struct FullPivLu<E: Entity> {
     factors: Mat<E>,
-    row_perm: Vec<usize>,
-    row_perm_inv: Vec<usize>,
-    col_perm: Vec<usize>,
-    col_perm_inv: Vec<usize>,
+    row_perm: alloc::vec::Vec<usize>,
+    row_perm_inv: alloc::vec::Vec<usize>,
+    col_perm: alloc::vec::Vec<usize>,
+    col_perm_inv: alloc::vec::Vec<usize>,
     n_transpositions: usize,
 }
 
@@ -78,8 +78,8 @@ pub struct Qr<E: Entity> {
 pub struct ColPivQr<E: Entity> {
     factors: Mat<E>,
     householder: Mat<E>,
-    col_perm: Vec<usize>,
-    col_perm_inv: Vec<usize>,
+    col_perm: alloc::vec::Vec<usize>,
+    col_perm_inv: alloc::vec::Vec<usize>,
 }
 
 /// Singular value decomposition.
@@ -270,8 +270,8 @@ impl<E: ComplexField> Lblt<E> {
 
         let mut factors = Mat::<E>::zeros(dim, dim);
         let mut subdiag = Mat::<E>::zeros(dim, 1);
-        let mut perm = vec![0; dim];
-        let mut perm_inv = vec![0; dim];
+        let mut perm = alloc::vec![0; dim];
+        let mut perm_inv = alloc::vec![0; dim];
 
         match side {
             Side::Lower => {
@@ -471,8 +471,8 @@ impl<E: ComplexField> PartialPivLu<E> {
 
         let params = Default::default();
 
-        let mut row_perm = vec![0usize; dim];
-        let mut row_perm_inv = vec![0usize; dim];
+        let mut row_perm = alloc::vec![0usize; dim];
+        let mut row_perm_inv = alloc::vec![0usize; dim];
 
         let (n_transpositions, _) = crate::linalg::lu::partial_pivoting::compute::lu_in_place(
             factors.as_mut(),
@@ -654,10 +654,10 @@ impl<E: ComplexField> FullPivLu<E> {
 
         let params = Default::default();
 
-        let mut row_perm = vec![0usize; m];
-        let mut row_perm_inv = vec![0usize; m];
-        let mut col_perm = vec![0usize; n];
-        let mut col_perm_inv = vec![0usize; n];
+        let mut row_perm = alloc::vec![0usize; m];
+        let mut row_perm_inv = alloc::vec![0usize; m];
+        let mut col_perm = alloc::vec![0usize; n];
+        let mut col_perm_inv = alloc::vec![0usize; n];
 
         let (n_transpositions, _, _) = crate::linalg::lu::full_pivoting::compute::lu_in_place(
             factors.as_mut(),
@@ -1092,8 +1092,8 @@ impl<E: ComplexField> ColPivQr<E> {
 
[…]
