+++ src/linalg/evd/mod.rs
@@ -13,7 +13,6 @@
 
 #![allow(clippy::type_complexity)]
 #![allow(clippy::too_many_arguments)]
-#![cfg_attr(not(feature = "std"), no_std)]
 
 use crate::{
     assert,
