### 2024-03-04 4c3db18
use panic instead of assert for better codegen
