### 2024-03-06 add1479
fix parallelism index splitting bug
