+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "faer"
-version = "0.18.1"
+version = "0.18.2"
 edition = "2021"
 authors = ["sarah <>"]
 description = "Linear algebra routines"
