### 2024-03-16 7ae6eb0
better conjugate handling
