+++ src/linop/lsmr.rs
@@ -1,0 +1,864 @@
+use crate::{
+    linalg::{householder, matmul::matmul, qr, temp_mat_req, temp_mat_uninit},
+    linop::{BiLinOp, BiPrecond, InitialGuessStatus},
+    prelude::*,
+    ComplexField, Conj, Parallelism, RealField,
+};
+use dyn_stack::{PodStack, SizeOverflow, StackReq};
+use equator::{assert, debug_assert};
+use reborrow::*;
+
+#[derive(Copy, Clone, Debug)]
+#[non_exhaustive]
+pub struct LsmrParams<E: ComplexField> {
+    pub initial_guess: InitialGuessStatus,
+    pub abs_tolerance: E::Real,
+    pub rel_tolerance: E::Real,
+    pub max_iters: usize,
+}
+
+impl<E: ComplexField> Default for LsmrParams<E> {
+    #[inline]
+    fn default() -> Self {
+        Self {
+            initial_guess: InitialGuessStatus::MaybeNonZero,
+            abs_tolerance: E::Real::faer_zero(),
+            rel_tolerance: E::Real::faer_epsilon().faer_mul(E::Real::faer_from_f64(128.0)),
+            max_iters: usize::MAX,
+        }
+    }
+}
+
+#[derive(Copy, Clone, Debug)]
+#[non_exhaustive]
+pub struct LsmrInfo<E: ComplexField> {
+    pub abs_residual: E::Real,
+    pub rel_residual: E::Real,
+    pub iter_count: usize,
+}
+
+#[derive(Copy, Clone, Debug)]
+pub enum LsmrError<E: ComplexField> {
+    NoConvergence {
+        abs_residual: E::Real,
+        rel_residual: E::Real,
+    },
+}
+
+#[allow(dead_code)]
+#[cfg(test)]
+fn lsmr_prototype_v2<E: ComplexField>(
+    mut x: MatMut<'_, E>,
+    _M: MatRef<'_, E>,
+    A: MatRef<'_, E>,
+    b: MatRef<'_, E>,
+    params: LsmrParams<E>,
+) {
+    let mut u = vec![];
+    let mut v = vec![];
+    let mut alpha = vec![];
+    let mut beta = vec![];
+
+    // let m = A.nrows();
+    // let n = A.ncols();
+    let s = b.ncols();
+
+    let qr = (b - A * &x).qr();
+    u.push(qr.compute_thin_q());
+    beta.push(qr.compute_thin_r());
+
+    let qr = (A.adjoint() * &u[0]).qr();
+    v.push(qr.compute_thin_q());
+    alpha.push(qr.compute_thin_r());
+
+    let mut L_kn = Mat::<E>::zeros((params.max_iters + 1) * s, (params.max_iters + 1) * s);
+    L_kn.as_mut()
+        .submatrix_mut(0, 0, s, s)
+        .copy_from(alpha[0].adjoint().to_owned());
+
+    for k in 0..params.max_iters {
+        let qr = (A * &v[k] - &u[k] * alpha[k].adjoint()).qr();
+        u.push(qr.compute_thin_q());
+        beta.push(qr.compute_thin_r());
+
+        let qr = (A.adjoint() * &u[k + 1] - &v[k] * beta[k + 1].adjoint()).qr();
+        v.push(qr.compute_thin_q());
+        alpha.push(qr.compute_thin_r());
+
+        L_kn.as_mut()
+            .submatrix_mut((k + 1) * s, (k + 1) * s, s, s)
+            .copy_from(alpha[k + 1].adjoint().to_owned());
+        L_kn.as_mut()
+            .submatrix_mut((k + 1) * s, k * s, s, s)
[…]
