+++ src/linop/linop_impl/mod.rs
@@ -1,0 +1,3 @@
+mod matmut;
+mod matown;
+mod matref;
