+++ Cargo.toml
@@ -17,7 +17,7 @@ dbgf = "0.1.2"
 paste = "1.0.14"
 reborrow = "0.5.5"
 
-dyn-stack = "0.10.0"
+dyn-stack = { version = "0.10.0", default-features = false }
 equator = "0.1.10"
 faer-entity = { version ="0.18.0", default-features = false, path = "./faer-entity" }
 
