### 2024-03-30 #6099 @FCLC
Reviewing the paper itself, I have a few comments, organized by line number (L{number})

L11: Worth clarifying that faer can operate on non symmetric Mats, as well as other categories it can and cannot natively handle? (Triangular? Hermitian?)

L14: Which standard do these floats adhere to? (presumably that's IEEE754 in either the 1985 or the 2008 flavour?). Complex types are mentioned, are they equivalent to the complex types defined in the reference BLAS? (they are later discussed at L28-31)

L15: The idea of universal type support is interesting, what can this be extended to? Integer maths? Fixed point? Posits? Fundamentally, could a user map any type that Rust/LLVM can use and have it be decently performant? Why or why not?

L18: SIMD is decently well documented online, but is not very well known outside of low level optimization circles; it may be helpful for higher level devs to provide the acronym if not a short description?

L22-26: I'd agree that a Rust native matrix library is useful, but can this rust library be actively built and used from C native projects? If so, does it share a common API/ABI (LAPACK, BLAS etc.?) to be useable from those other languages?

L24-26: Does the lack of a Rust ABI present portability challenges to faer? Is there an expectation that any users of faer will always have to rebuild faer from source between versions? If so, being as linear algebra libraries tend to be heavily leveraged in HPC contexts/environments, does the usage of rust mandate a change in approach to (sub)versioning?

L29: To help those not intimately familiar with performance engineering, is it worth adding a brief sentence explaining why contiguous memory is valuable? Even something as simple as a mention of strided/aligned memory accesses?

L38: How is CPU feature detection handled? AOT compiled multi dispatch for a list of targets, local target only compilation, etc.? Is this detection reliant on cargo or does it make use of other/homegrown methods?

L41-45: it might be worth having a glossary or spelling out what these terms mean?

L47: Brief explanation of what is meant by "Optimized fuzed kernels"?

L57-61: since most performance metric for BLAS and BLAS adjacent libraries tend to express performance in terms of floating point operations per second, is it worth providing a table comparing faer vs various other implementations in terms of (giga) floating point operations per second?

L61: {Graphs} F128 is mentioned, but not "which" f128 is this? is this IEEE FP128? IBM/PPC FP128? 2xFP64 packed FP128 (where the leading FP64 is IEEE754 Binary_64 with an extended 64 bits treated as a uint softfloat)?

Beyond that: is it worth having a statement/line indicating what platforms are supported? For example, is faer a linux x86-64 only library? what about Arm64? MacOS? Windows? FreeBSD? Does rust/cargo being limited in HPC enviroments cause concerns for long term support of versions? or are users expected to use cargo/spack/EasyBuild to stay on top of releases?
