### 2024-03-31 #6099 @sarah-ek
i'll reply to the comments here for now, but let me know if i should also add these clarifications to the paper itself and rebuild

L11: could you explain what you mean by natively handle? there's algorithms for general, triangular, symmetric, hermitian matrices. (in the past year sparse matrix algorithms have also advanced quite a bit)
the main ones that are missing afaik are banded matrices, and packed triangular/symmetric/hermitian storage, but that's not an inherent limitation of faer

L14: rust uses IEEE 754-2008 floating types for f32 and f64, complex types c32 and c64 are the same as `float _Complex`, `double _Complex` in C, but there's also a generic implementation for `Complex<T>` that can be used with different float types

L15: integer maths is unlikely, but the most interesting scenarios i can think of would be stuff like extended/multiprecision floating point types, similar to what mpfr implements. or things like interval arithmetic, forward autodiff types (dual/hyperdual numbers), and any possible combinations of such types (`Complex<Dual<BigFloat>>` should also work in most situations)

L18: i can add in a short description

L22-26: rust allows for interoperability with C by using the C abi on functions that are marked as `extern "C"` (commonly also used with `#[no_mangle]` so the symbol can be found by the linker as you'd expect)
i've used faer from C and C++ in the past and it works quite well

L24-26: users are expected to rebuild when upgrading versions, yes. but there's nothing preventing us or other people from distributing precompiled static/dynamic library files exposing a C api.

L29: i can add in an explanation

L38: cpu detection is done for predetermined targets, but each type can specify which targets it cares about. for example if a user wants to work with f16 (half) matrices, they can specify that they only want to try to detect something like avx512fp16 on x86

L41-L45: sure thing. i'll add that to the paper

L47: in this context it refers to computing things like `A x` and `A.T y` simultaneously, only traversing `A` once, which can significantly speed up some memory-bound algorithms

L57-61: gflops/s works for matrix multiplication. but for more complex iterative algorithms like the eigenvalue decomposition, the flop count can depend on several factors and there's no closed form formula for it, so im not sure how to communicate that

L61: that's my bad, i should have clarified that it's f64x2 (commonly referred to as double double)

faer supports any platforms that rust itself supports. so it can run on the common operating systems, and the target architectures that rust supports

    x86
    x86_64
    mips
    powerpc
    powerpc64
    arm
    aarch64

although vectorization is currently only supported on x86/x86_64 and aarch64 (no sve support at the moment), due to other platforms' intrinsics being currently unstable in rust

other than that, users are expected to use cargo to stay up to date, but they can build and distribute faer as a C api (still not currently exposed by faer itself, but users can easily expose the parts they care about) + static/dynamic library if the environment doesn't support building rust projects from source
