### 2024-03-31 ead917e
small improvement for pseudoinverse
