+++ src/row/rowmut.rs
@@ -635,3 +635,9 @@ impl<E: Conjugate> RowBatch<E> for RowMut<'_, E> {
 }
 
 impl<E: Conjugate> RowBatchMut<E> for RowMut<'_, E> {}
+
+/// Returns a view over a row with 1 column containing value as its only element, pointing to
+/// `value`.
+pub fn from_mut<E: Entity>(value: GroupFor<E, &mut E::Unit>) -> RowMut<'_, E> {
+    unsafe { from_raw_parts_mut(E::faer_map(value, |ptr| ptr as *mut E::Unit), 1, 1) }
+}
