+++ src/col/colmut.rs
@@ -655,3 +655,9 @@ impl<E: Conjugate> ColBatch<E> for ColMut<'_, E> {
     }
 }
 impl<E: Conjugate> ColBatchMut<E> for ColMut<'_, E> {}
+
+/// Returns a view over a column with 1 row containing value as its only element, pointing to
+/// `value`.
+pub fn from_mut<E: Entity>(value: GroupFor<E, &mut E::Unit>) -> ColMut<'_, E> {
+    unsafe { from_raw_parts_mut(E::faer_map(value, |ptr| ptr as *mut E::Unit), 1, 1) }
+}
