### 2024-04-07 81124e5
fix diagonal selfadjoint eigendecomp case
