+++ src/linalg/evd/tridiag_real_evd.rs
@@ -540,8 +540,10 @@ fn compute_tridiag_real_evd_impl<E: RealField>(
     let n = diag.len();
 
     if n <= 1 {
-        zipped!(u.rb_mut().diagonal_mut().column_vector_mut().as_2d_mut())
-            .for_each(|unzipped!(mut x)| x.write(E::faer_one()));
+        u.rb_mut()
+            .diagonal_mut()
+            .column_vector_mut()
+            .fill(E::faer_one());
         return;
     }
 
@@ -722,14 +724,13 @@ fn compute_tridiag_real_evd_impl<E: RealField>(
     // diag([D0 D1]) + rho×z×z.T = Pl_before^-1×H^-1×Pl_after^-1×Q×Pr × E × ...
 
     let (mut z0, mut z1) = z.rb_mut().split_at_row_mut(n1);
-    zipped!(z0.rb_mut(), u0.rb().row(n1 - 1).transpose().as_2d())
-        .for_each(|unzipped!(mut z, u)| z.write(u.read()));
+    z0.rb_mut()
+        .copy_from(u0.rb().row(n1 - 1).transpose().as_2d());
     if rho < E::faer_zero() {
         zipped!(z1.rb_mut(), u1.rb().row(0).transpose().as_2d())
             .for_each(|unzipped!(mut z, u)| z.write(u.read().faer_neg()));
     } else {
-        zipped!(z1.rb_mut(), u1.rb().row(0).transpose().as_2d())
-            .for_each(|unzipped!(mut z, u)| z.write(u.read()));
+        z1.rb_mut().copy_from(u1.rb().row(0).transpose().as_2d());
     }
 
     let inv_sqrt2 = E::faer_from_f64(2.0).faer_sqrt().faer_inv();
@@ -793,17 +794,13 @@ fn compute_tridiag_real_evd_impl<E: RealField>(
         // return
 
         let (mut tmp_tl, mut tmp_tr, mut tmp_bl, mut tmp_br) = tmp.rb_mut().split_at_mut(n1, n1);
-        zipped!(tmp_tl.rb_mut(), u0.rb()).for_each(|unzipped!(mut dst, src)| dst.write(src.read()));
-        zipped!(tmp_br.rb_mut(), u1.rb()).for_each(|unzipped!(mut dst, src)| dst.write(src.read()));
-        zipped!(tmp_tr.rb_mut()).for_each(|unzipped!(mut dst)| dst.write(E::faer_zero()));
-        zipped!(tmp_bl.rb_mut()).for_each(|unzipped!(mut dst)| dst.write(E::faer_zero()));
+        tmp_tl.rb_mut().copy_from(u0.rb());
+        tmp_br.rb_mut().copy_from(u1.rb());
+        tmp_tr.fill_zero();
+        tmp_bl.fill_zero();
 
-        for (j, &dst_j) in pl_before.iter().enumerate() {
-            zipped!(
-                u.rb_mut().col_mut(dst_j).as_2d_mut(),
-                tmp.rb().col(j).as_2d()
-            )
-            .for_each(|unzipped!(mut dst, src)| dst.write(src.read()));
+        for (j, &pl_before) in pl_before.iter().enumerate() {
+            u.rb_mut().col_mut(j).copy_from(tmp.rb().col(pl_before));
         }
 
         for (j, diag) in diag.iter_mut().enumerate() {
@@ -846,8 +843,7 @@ fn compute_tridiag_real_evd_impl<E: RealField>(
             let mut householder = householder.rb_mut().subrows_mut(idx, run_len);
             let mut permuted_z = permuted_z.rb_mut().subrows_mut(idx, run_len);
 
-            zipped!(householder.rb_mut(), permuted_z.rb())
[…]
