### 2024-04-09 c1020eb
fix plu compute_l and compute_u
