+++ src/linalg/solvers.rs
@@ -521,6 +521,11 @@ impl<E: ComplexField> PartialPivLu<E> {
                 dst.write(E::faer_zero())
             });
         factor
+            .as_mut()
+            .diagonal_mut()
+            .column_vector_mut()
+            .fill(E::faer_one());
+        factor
     }
     /// Returns the factor $U$ of the LU decomposition.
     pub fn compute_u(&self) -> Mat<E> {
@@ -529,11 +534,6 @@ impl<E: ComplexField> PartialPivLu<E> {
             .for_each_triangular_lower(crate::linalg::zip::Diag::Skip, |unzipped!(mut dst)| {
                 dst.write(E::faer_zero())
             });
-        factor
-            .as_mut()
-            .diagonal_mut()
-            .column_vector_mut()
-            .fill(E::faer_one());
         factor
     }
 }
@@ -3049,5 +3049,21 @@ mod tests {
             (lambda_1 - correct_lamba_1).abs() < 1e-10,
             "lambda_1 = {lambda_1}, correct_lamba_1 = {correct_lamba_1}",
         );
+    }
+
+    #[test]
+    fn test_lu() {
+        let a = mat![
+            [0.75026225, 0.35005635, -0.55833477],
+            [0.57985423, -0.75391293, 0.30216142],
+            [0.31665369, 0.54900739, 0.76136962],
+        ];
+        let plu = a.partial_piv_lu();
+        let p = plu.row_permutation();
+        let l = plu.compute_l();
+        let u = plu.compute_u();
+
+        let diff = (p * a) - (l * u);
+        assert!(diff.norm_max() < 1e-12);
     }
 }
