### 2024-04-09 30ccd8e
fix flu compute_l and compute_u
