### 2024-04-09 97fc816
impl col/row mean and variance
