+++ src/stats/meanvar.rs
@@ -1,0 +1,323 @@
+use crate::{
+    linalg::entity::{pulp, SimdCtx, SimdGroupFor},
+    prelude::*,
+    utils::slice::{RefGroup, SliceGroup},
+    ComplexField,
+};
+use equator::assert;
+use pulp::Read;
+
+/// Computes the mean of the columns of `mat` and stores the result in `out`.
+#[track_caller]
+pub fn col_mean<E: ComplexField>(out: ColMut<'_, E>, mat: MatRef<'_, E>) {
+    assert!(all(out.nrows() == mat.nrows(), mat.ncols() > 0));
+
+    fn col_mean_row_major<E: ComplexField>(out: ColMut<'_, E>, mat: MatRef<'_, E>) {
+        struct Impl<'a, E: ComplexField> {
+            out: ColMut<'a, E>,
+            mat: MatRef<'a, E>,
+        }
+
+        impl<E: ComplexField> pulp::WithSimd for Impl<'_, E> {
+            type Output = ();
+
+            #[inline(always)]
+            fn with_simd<S: pulp::Simd>(self, simd: S) -> Self::Output {
+                let Self { mut out, mat } = self;
+                let simd = crate::utils::simd::SimdFor::<E, S>::new(simd);
+
+                let m = mat.nrows();
+                let n = mat.ncols();
+                let one_n = E::Real::faer_from_f64(n as f64).faer_inv();
+
+                let offset = simd.align_offset_ptr(mat.as_ptr(), mat.ncols());
+                for i in 0..m {
+                    let row = SliceGroup::<'_, E>::new(mat.row(i).try_as_slice().unwrap());
+                    let (head, body, tail) = simd.as_aligned_simd(row, offset);
+                    let mut sum0 = head.read_or(simd.splat(E::faer_zero()));
+                    let mut sum1 = simd.splat(E::faer_zero());
+                    let mut sum2 = simd.splat(E::faer_zero());
+                    let mut sum3 = simd.splat(E::faer_zero());
+
+                    let (body4, body1) = body.as_arrays::<4>();
+                    for [x0, x1, x2, x3] in body4.into_ref_iter().map(RefGroup::unzip) {
+                        sum0 = simd.add(sum0, x0.get());
+                        sum1 = simd.add(sum1, x1.get());
+                        sum2 = simd.add(sum2, x2.get());
+                        sum3 = simd.add(sum3, x3.get());
+                    }
+                    for x0 in body1.into_ref_iter() {
+                        sum0 = simd.add(sum0, x0.get());
+                    }
+                    sum0 = simd.add(sum0, tail.read_or(simd.splat(E::faer_zero())));
+
+                    sum0 = simd.add(sum0, sum1);
+                    sum2 = simd.add(sum2, sum3);
+                    sum0 = simd.add(sum0, sum2);
+
+                    sum0 = simd.rotate_left(sum0, offset.rotate_left_amount());
+                    let sum = simd.reduce_add(sum0);
+
+                    out.write(i, sum.faer_scale_real(one_n));
+                }
+            }
+        }
+
+        E::Simd::default().dispatch(Impl { out, mat });
+    }
+
+    let mat = if mat.col_stride() >= 0 {
+        mat
+    } else {
+        mat.reverse_cols()
+    };
+    if mat.col_stride() == 1 {
+        col_mean_row_major(out, mat)
+    } else {
+        let mut out = out;
+
+        let n = mat.ncols();
+        let one_n = E::Real::faer_from_f64(n as f64).faer_inv();
+
+        out.fill_zero();
+        for j in 0..n {
+            out += mat.col(j);
+        }
+        zipped!(out).for_each(|unzipped!(mut x)| x.write(x.read().faer_scale_real(one_n)));
+    }
+}
+
[…]
