### 2024-04-12 113cc8e
row col mean variance with nan handling
