+++ src/stats/meanvar.rs
@@ -1,17 +1,1189 @@
 use crate::{
-    linalg::entity::{pulp, SimdCtx, SimdGroupFor},
+    linalg::entity::{pulp, SimdCtx, SimdGroupFor, SimdIndexFor},
     prelude::*,
-    utils::slice::{RefGroup, SliceGroup},
-    ComplexField,
+    utils::{
+        simd::SimdFor,
+        slice::{RefGroup, SliceGroup},
+    },
+    ComplexField, RealField,
 };
+use coe::Coerce;
 use equator::assert;
+use num_complex::Complex;
 use pulp::Read;
 
-/// Computes the mean of the columns of `mat` and stores the result in `out`.
-#[track_caller]
-pub fn col_mean<E: ComplexField>(out: ColMut<'_, E>, mat: MatRef<'_, E>) {
-    assert!(all(out.nrows() == mat.nrows(), mat.ncols() > 0));
+/// Specifies how missing values should be handled in mean and variance computations.
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
+pub enum NanHandling {
+    /// NaNs are passed as-is to arithmetic operators.
+    Propagate,
+    /// NaNs are skipped, and they're not included in the total count of entries.
+    Ignore,
+}
 
+#[inline(always)]
+fn from_usize<E: RealField>(n: usize) -> E {
+    E::faer_from_f64(n as u32 as f64)
+        .faer_add(E::faer_from_f64((n as u64 - (n as u32 as u64)) as f64))
+}
+
+#[inline(always)]
+fn reduce<E: RealField, S: pulp::Simd>(non_nan_count: SimdIndexFor<E, S>) -> usize {
+    let slice: &[E::Index] = bytemuck::cast_slice(core::slice::from_ref(&non_nan_count));
+
+    let mut acc = 0usize;
+    for &count in slice {
+        acc += E::faer_index_to_usize(count);
+    }
+    acc
+}
+
+fn col_mean_row_major_ignore_nan_real<E: RealField>(out: ColMut<'_, E>, mat: MatRef<'_, E>) {
+    struct Impl<'a, E: RealField> {
+        out: ColMut<'a, E>,
+        mat: MatRef<'a, E>,
+    }
+
+    impl<E: RealField> pulp::WithSimd for Impl<'_, E> {
+        type Output = ();
+
+        #[inline(always)]
+        fn with_simd<S: pulp::Simd>(self, simd: S) -> Self::Output {
+            let Self { mut out, mat } = self;
+            let simd = SimdFor::<E, S>::new(simd);
+
+            let m = mat.nrows();
+            let chunk_size = if core::mem::size_of::<E::Index>() < core::mem::size_of::<usize>() {
+                1usize << (core::mem::size_of::<E::Index>() * 8)
+            } else {
+                usize::MAX
+            } / 4;
+
+            let offset = simd.align_offset_ptr(mat.as_ptr(), mat.ncols());
+            for i in 0..m {
+                let row = SliceGroup::<'_, E>::new(mat.row(i).try_as_slice().unwrap());
+                let (head, body, tail) = simd.as_aligned_simd(row, offset);
+
+                let mut non_nan_count_total = 0usize;
+
+                #[inline(always)]
+                fn process<E: RealField, S: pulp::Simd>(
+                    simd: SimdFor<E, S>,
+                    acc: SimdGroupFor<E, S>,
+                    non_nan_count: SimdIndexFor<E, S>,
+                    val: impl Read<Output = SimdGroupFor<E, S>>,
+                ) -> (SimdGroupFor<E, S>, SimdIndexFor<E, S>) {
+                    let val = val.read_or(simd.splat(E::faer_nan()));
+                    let is_not_nan = simd.less_than_or_equal(val, val);
+
+                    (
[…]
