### 2024-04-13 6d9f737
optimize col meanvar
