+++ src/stats/meanvar.rs
@@ -3,14 +3,16 @@ use crate::{
     prelude::*,
     utils::{
         simd::SimdFor,
-        slice::{RefGroup, SliceGroup},
+        slice::{RefGroup, SliceGroup, SliceGroupMut},
     },
     ComplexField, RealField,
 };
 use coe::Coerce;
+use core::iter::zip;
 use equator::assert;
 use num_complex::Complex;
-use pulp::Read;
+use pulp::{Read, Write};
+use reborrow::*;
 
 /// Specifies how missing values should be handled in mean and variance computations.
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
@@ -1238,6 +1240,75 @@ fn col_mean_propagate<E: ComplexField>(out: ColMut<'_, E>, mat: MatRef<'_, E>) {
         E::Simd::default().dispatch(Impl { out, mat });
     }
 
+    fn col_mean_col_major<E: ComplexField>(out: ColMut<'_, E>, mat: MatRef<'_, E>) {
+        struct Impl<'a, E: ComplexField> {
+            out: ColMut<'a, E>,
+            mat: MatRef<'a, E>,
+        }
+
+        impl<E: ComplexField> pulp::WithSimd for Impl<'_, E> {
+            type Output = ();
+
+            #[inline(always)]
+            fn with_simd<S: pulp::Simd>(self, simd: S) -> Self::Output {
+                let Self { out, mat } = self;
+                let simd_real = SimdFor::<E::Real, S>::new(simd);
+                let simd = SimdFor::<E, S>::new(simd);
+
+                let m = mat.nrows();
+                let n = mat.ncols();
+                let one_n = simd_real.splat(from_usize::<E::Real>(n).faer_inv());
+
+                let offset = simd.align_offset_ptr(mat.as_ptr(), m);
+                let mut out = SliceGroupMut::<'_, E>::new(out.try_as_slice_mut().unwrap());
+
+                out.fill_zero();
+                for j in 0..n {
+                    let col = SliceGroup::<'_, E>::new(mat.col(j).try_as_slice().unwrap());
+                    let (head, body, tail) = simd.as_aligned_simd(col, offset);
+                    let (out_head, out_body, out_tail) =
+                        simd.as_aligned_simd_mut(out.rb_mut(), offset);
+
+                    #[inline(always)]
+                    fn process<E: ComplexField, S: pulp::Simd>(
+                        simd: SimdFor<E, S>,
+                        mut out: impl Write<Output = SimdGroupFor<E, S>>,
+                        val: impl Read<Output = SimdGroupFor<E, S>>,
+                    ) {
+                        out.write(simd.add(
+                            out.read_or(simd.splat(E::faer_zero())),
+                            val.read_or(simd.splat(E::faer_zero())),
+                        ))
+                    }
+
+                    process(simd, out_head, head);
+                    for (out, x) in zip(out_body.into_mut_iter(), body.into_ref_iter()) {
+                        process(simd, out, x)
+                    }
+                    process(simd, out_tail, tail);
+                }
+
+                #[inline(always)]
+                fn process<E: ComplexField, S: pulp::Simd>(
+                    simd: SimdFor<E, S>,
+                    mut out: impl Write<Output = SimdGroupFor<E, S>>,
+                    one_n: SimdGroupFor<E::Real, S>,
+                ) {
+                    out.write(simd.scale_real(one_n, out.read_or(simd.splat(E::faer_zero()))))
+                }
+                let (out_head, out_body, out_tail) = simd.as_aligned_simd_mut(out.rb_mut(), offset);
+                process(simd, out_head, one_n);
+                for out in out_body.into_mut_iter() {
+                    process(simd, out, one_n);
+                }
[…]
