+++ Cargo.toml
@@ -56,6 +56,7 @@ npy = ["std", "dep:npyz"]
 [dev-dependencies]
 amd = "0.2.2"
 assert_approx_eq = "1.1.0"
+divan = "0.1.14"
 matrix-market-rs = "0.1.3"
 matrixcompare = "0.3.0"
 rand = "0.8.5"
@@ -67,3 +68,7 @@ opt-level = 3
 [package.metadata.docs.rs]
 all-features = true
 rustdoc-args = ["--cfg", "docsrs", "--html-in-header", "katex-header.html"]
+
+[[bench]]
+name = "meanvar"
+harness = false
