### 2024-04-13 614cc66
delegate a bunch of functions using as_ref/as_mut
