+++ benches/meanvar.rs
@@ -1,15 +1,15 @@
+use diol::prelude::*;
 use faer::{prelude::*, ComplexField};
 
-fn args() -> Vec<[usize; 2]> {
-    vec![[4096; 2]]
+fn args() -> Vec<List![usize, usize]> {
+    (5..12).map(|i| 1 << i).map(|n| list![n, n]).collect()
 }
 
-#[divan::bench(types = [f32, c32, f64, c64], args = args())]
-fn col_mean_propagate<E: ComplexField>(bencher: divan::Bencher, [m, n]: [usize; 2]) {
+fn col_mean_propagate<E: ComplexField>(bencher: Bencher, unlist![m, n]: List![usize, usize]) {
     let a = Mat::from_fn(m, n, |_, _| E::faer_from_f64(1.0_f64));
     let mut out = Col::zeros(m);
 
-    bencher.bench_local(|| {
+    bencher.bench(|| {
         faer::stats::col_mean(
             out.as_mut(),
             a.as_ref(),
@@ -18,22 +18,20 @@ fn col_mean_propagate<E: ComplexField>(bencher: divan::Bencher, [m, n]: [usize; 
     })
 }
 
-#[divan::bench(types = [f32, c32, f64, c64], args = args())]
-fn col_mean_ignore<E: ComplexField>(bencher: divan::Bencher, [m, n]: [usize; 2]) {
+fn col_mean_ignore<E: ComplexField>(bencher: Bencher, unlist![m, n]: List![usize, usize]) {
     let a = Mat::from_fn(m, n, |_, _| E::faer_from_f64(1.0_f64));
     let mut out = Col::zeros(m);
 
-    bencher.bench_local(|| {
+    bencher.bench(|| {
         faer::stats::col_mean(out.as_mut(), a.as_ref(), faer::stats::NanHandling::Ignore);
     })
 }
 
-#[divan::bench(types = [f32, c32, f64, c64], args = args())]
-fn row_mean_propagate<E: ComplexField>(bencher: divan::Bencher, [m, n]: [usize; 2]) {
+fn row_mean_propagate<E: ComplexField>(bencher: Bencher, unlist![m, n]: List![usize, usize]) {
     let a = Mat::from_fn(m, n, |_, _| E::faer_from_f64(1.0_f64));
     let mut out = Row::zeros(n);
 
-    bencher.bench_local(|| {
+    bencher.bench(|| {
         faer::stats::row_mean(
             out.as_mut(),
             a.as_ref(),
@@ -42,23 +40,21 @@ fn row_mean_propagate<E: ComplexField>(bencher: divan::Bencher, [m, n]: [usize; 
     })
 }
 
-#[divan::bench(types = [f32, c32, f64, c64], args = args())]
-fn row_mean_ignore<E: ComplexField>(bencher: divan::Bencher, [m, n]: [usize; 2]) {
+fn row_mean_ignore<E: ComplexField>(bencher: Bencher, unlist![m, n]: List![usize, usize]) {
     let a = Mat::from_fn(m, n, |_, _| E::faer_from_f64(1.0_f64));
     let mut out = Row::zeros(n);
 
-    bencher.bench_local(|| {
+    bencher.bench(|| {
         faer::stats::row_mean(out.as_mut(), a.as_ref(), faer::stats::NanHandling::Ignore);
     })
 }
 
-#[divan::bench(types = [f32, c32, f64, c64], args = args())]
-fn col_varm_propagate<E: ComplexField>(bencher: divan::Bencher, [m, n]: [usize; 2]) {
+fn col_varm_propagate<E: ComplexField>(bencher: Bencher, unlist![m, n]: List![usize, usize]) {
     let a = Mat::from_fn(m, n, |_, _| E::faer_from_f64(1.0_f64));
[…]
