+++ Cargo.toml
@@ -56,7 +56,7 @@ npy = ["std", "dep:npyz"]
 [dev-dependencies]
 amd = "0.2.2"
 assert_approx_eq = "1.1.0"
-divan = "0.1.14"
+diol = "0.2.0"
 matrix-market-rs = "0.1.3"
 matrixcompare = "0.3.0"
 rand = "0.8.5"
