+++ Cargo.toml
@@ -56,10 +56,15 @@ npy = ["std", "dep:npyz"]
 [dev-dependencies]
 amd = "0.2.2"
 assert_approx_eq = "1.1.0"
-diol = "0.2.0"
+diol = { version = "0.6.1", path = "../diol" }
 matrix-market-rs = "0.1.3"
 matrixcompare = "0.3.0"
+nalgebra = "0.32.5"
+blas-src = { version = "0.8", features = ["openblas"] }
+ndarray = { version = "0.15.6", features = ["blas"] }
+ndarray-linalg = { version = "0.16.0", features = ["openblas-system"], git = "https://github.com/mike-kfed/ndarray-linalg.git", branch="arm-cross-compile"}
 rand = "0.8.5"
+rand_distr = "0.4.3"
 serde_test = "1.0.176"
 
 [profile.dev]
@@ -71,4 +76,8 @@ rustdoc-args = ["--cfg", "docsrs", "--html-in-header", "katex-header.html"]
 
 [[bench]]
 name = "meanvar"
+harness = false
+
+[[bench]]
+name = "bench"
 harness = false

[…]
