### 2024-04-23 c746d3e
Add 'ones' and 'full' constructors for owned Col, Row, and Matrix.
