### 2024-04-24 ee71247
impl faer_mul_add and faer_abs2_add
