### 2024-04-26 3489353
fixes crash when (pme1,pme2) = (0,-1)
