+++ lapack.cpp
@@ -1,0 +1,402 @@
+#include <diol.hpp>
+#include <eigen3/Eigen/Core>
+
+#include <mkl/mkl_lapack.h>
+
+#define CAT_(A, B) A##B
+#define CAT(A, B) CAT_(A, B)
+#define MAKE_LAPACK(prefix, func)                                              \
+  inline static constexpr auto func = CAT(CAT(prefix, func), _);
+
+using namespace diol;
+template <typename T> T const *addr(T const &ptr) { return &ptr; }
+template <typename T> T *addr(T &&ptr) { return &ptr; }
+MKL_INT max(MKL_INT a, MKL_INT b) { return a > b ? a : b; }
+
+using f32 = float;
+using f64 = double;
+using c32 = std::complex<f32>;
+using c64 = std::complex<f64>;
+
+template <typename T> using Mat = Eigen::Matrix<T, -1, -1>;
+
+#define MAKE_MKL(ty, mkl_ty, prefix, sym)                                      \
+  template <> struct Mkl<ty> {                                                 \
+    using T = mkl_ty;                                                          \
+    MAKE_LAPACK(prefix, potrf);                                                \
+    MAKE_LAPACK(prefix, geqrf);                                                \
+    MAKE_LAPACK(prefix, geqp3);                                                \
+    MAKE_LAPACK(prefix, getrf);                                                \
+    MAKE_LAPACK(prefix, getc2);                                                \
+    MAKE_LAPACK(prefix, gesdd);                                                \
+    MAKE_LAPACK(CAT(prefix, sym), evd);                                        \
+    MAKE_LAPACK(prefix, geev);                                                 \
+  };
+
+template <typename E> struct Mkl;
+
+MAKE_MKL(f32, f32, s, sy);
+MAKE_MKL(f64, f64, d, sy);
+MAKE_MKL(c32, MKL_Complex8, c, he);
+MAKE_MKL(c64, MKL_Complex16, z, he);
+
+template <typename T> using Data = typename Mkl<T>::T *;
+
+template <typename T> MKL_INT stride(Mat<T> const &mat) {
+  return mat.outerStride();
+}
+template <typename T> auto data(Mat<T> &mat) { return Data<T>(mat.data()); }
+
+template <typename T> Mat<T> rand(MKL_INT m, MKL_INT n) {
+  Mat<T> A(m, n);
+  A.setRandom();
+  return A;
+}
+
+template <typename T> Mat<T> rand_pos_def(MKL_INT n) {
+  Mat<T> A(n, n);
+  A.setRandom();
+  Mat<T> H = T(0.5) * (A + A.adjoint());
+  H += T(n) * Mat<T>::Identity(n, n);
+  return H;
+}
+
+template <typename T> Mat<T> randh(MKL_INT n) {
+  Mat<T> A(n, n);
+  A.setRandom();
+  Mat<T> H = T(0.5) * (A + A.adjoint());
+  return H;
+}
+
+template <typename T> void cholesky(Bencher bencher, PlotArg arg) {
+  MKL_INT n = arg.n;
+  std::srand(0);
+  Mat<T> H_orig = rand_pos_def<T>(n);
+  Mat<T> H = H_orig;
+
+  std::move(bencher).bench([&] {
+    H = H_orig;
+
[…]
