+++ Cargo.toml
@@ -21,7 +21,7 @@ dyn-stack = { version = "0.10.0", default-features = false }
 equator = "0.1.10"
 faer-entity = { version ="0.18.0", default-features = false, path = "./faer-entity" }
 
-gemm = { version = "0.17.1", default-features = false }
+gemm = { version = "0.17.1", default-features = false, path = "../gemm/gemm" }
 num-complex = { version = "0.4.5", default-features = false }
 num-traits = { version = "0.2.18", default-features = false }
 
@@ -46,7 +46,7 @@ std = [
   "num-traits/std",
   "num-complex/std",
 ]
-rand = ["dep:rand", "rand_distr"]
+rand = ["dep:rand", "rand_distr", "num-complex/rand"]
 rayon = ["std", "gemm/rayon", "dep:rayon"]
 nightly = ["faer-entity/nightly", "gemm/nightly"]
 perf-warn = ["log"]
@@ -56,15 +56,13 @@ npy = ["std", "dep:npyz"]
 [dev-dependencies]
 amd = "0.2.2"
 assert_approx_eq = "1.1.0"
-diol = { version = "0.6.1" }
+diol = { version = "0.8.2", path = "../diol" }
 matrix-market-rs = "0.1.3"
 matrixcompare = "0.3.0"
 nalgebra = "0.32.5"
-blas-src = { version = "0.8", features = ["openblas"] }
-ndarray = { version = "0.15.6", features = ["blas"] }
-ndarray-linalg = { version = "0.16.0", features = ["openblas-system"], git = "https://github.com/mike-kfed/ndarray-linalg.git", branch="arm-cross-compile"}
 rand = "0.8.5"
 rand_distr = "0.4.3"
+serde_json = "1.0.116"
 serde_test = "1.0.176"
 
 [profile.dev]
@@ -80,4 +78,8 @@ harness = false
 
 [[bench]]
 name = "bench"
+harness = false
+
+[[bench]]
+name = "bench_aggregate"
 harness = false

[…]
