+++ Cargo.toml
@@ -21,7 +21,7 @@ dyn-stack = { version = "0.10.0", default-features = false }
 equator = "0.1.10"
 faer-entity = { version ="0.18.0", default-features = false, path = "./faer-entity" }
 
-gemm = { version = "0.17.1", default-features = false, path = "../gemm/gemm" }
+gemm = { version = "0.17.1", default-features = false }
 num-complex = { version = "0.4.5", default-features = false }
 num-traits = { version = "0.2.18", default-features = false }
 
