+++ Cargo.toml
@@ -56,7 +56,7 @@ npy = ["std", "dep:npyz"]
 [dev-dependencies]
 amd = "0.2.2"
 assert_approx_eq = "1.1.0"
-diol = { version = "0.8.2", path = "../diol" }
+diol = { version = "0.8.3" }
 matrix-market-rs = "0.1.3"
 matrixcompare = "0.3.0"
 nalgebra = "0.32.5"
