### 2024-05-05 7b92c0c
Corrected some typos in the documentation
Minor typos only
