+++ src/mat/matown.rs
@@ -1605,7 +1605,7 @@ impl<E: Entity> Mat<E> {
     }
 
     /// Given a matrix with a single column, returns an object that interprets
-    /// the column as a diagonal matrix, whoes diagonal elements are values in the column.
+    /// the column as a diagonal matrix, whose diagonal elements are values in the column.
     #[track_caller]
     #[inline(always)]
     pub fn column_vector_as_diagonal(&self) -> DiagRef<'_, E> {
@@ -1613,7 +1613,7 @@ impl<E: Entity> Mat<E> {
     }
 
     /// Given a matrix with a single column, returns an object that interprets
-    /// the column as a diagonal matrix, whoes diagonal elements are values in the column.
+    /// the column as a diagonal matrix, whose diagonal elements are values in the column.
     #[track_caller]
     #[inline(always)]
     pub fn column_vector_as_diagonal_mut(&mut self) -> DiagMut<'_, E> {
