+++ src/mat/matref.rs
@@ -834,7 +834,7 @@ impl<'a, E: Entity> MatRef<'a, E> {
     }
 
     /// Given a matrix with a single column, returns an object that interprets
-    /// the column as a diagonal matrix, whoes diagonal elements are values in the column.
+    /// the column as a diagonal matrix, whose diagonal elements are values in the column.
     #[track_caller]
     #[inline(always)]
     pub fn column_vector_as_diagonal(self) -> DiagRef<'a, E> {
