+++ src/linalg/evd/hessenberg_real_evd.rs
@@ -1602,7 +1602,7 @@ fn aggressive_early_deflation<E: RealField>(
     }
 
     if ns == jw {
-        // Agressive early deflation didn't deflate any eigenvalues
+        // Aggressive early deflation didn't deflate any eigenvalues
         // We don't need to apply the update to the rest of the matrix
         nd = jw - ns;
         ns -= infqr;
@@ -1714,7 +1714,7 @@ fn aggressive_early_deflation<E: RealField>(
         i += n1;
     }
 
-    // Reduce A back to Hessenberg form (if neccesary)
+    // Reduce A back to Hessenberg form (if necessary)
     if s_spike != E::faer_zero() {
         // Reflect spike back
         {
@@ -2356,7 +2356,7 @@ fn multishift_qr_sweep<E: RealField>(
     }
 
     //
-    // The following code block moves the bulges down untill they are low enough
+    // The following code block moves the bulges down until they are low enough
     // to be removed
     //
     while i_pos_block + n_block_desired < ihi {
@@ -3178,7 +3178,7 @@ pub fn multishift_qr<E: RealField>(
         }
 
         //
-        // Agressive early deflation
+        // Aggressive early deflation
         //
         let nh = istop - istart;
         let nwupbd = Ord::min(nh, nw_max);
