+++ src/linalg/qr/no_pivoting/mod.rs
@@ -9,5 +9,5 @@ pub mod compute;
 pub mod inverse;
 /// Reconstructing the original matrix from the decomposition.
 pub mod reconstruct;
-/// Solving a linear system usin the decomposition.
+/// Solving a linear system using the decomposition.
 pub mod solve;
