### 2024-05-05 afe82d1
Corrected some typos in the documentation
Minor typos only
