+++ src/linalg/cholesky/llt/mod.rs
@@ -8,7 +8,7 @@ pub mod compute;
 pub mod inverse;
 /// Reconstructing the original matrix from the decomposition.
 pub mod reconstruct;
-/// Solving a linear system usin the decomposition.
+/// Solving a linear system using the decomposition.
 pub mod solve;
 /// Updating the decomposition.
 pub mod update;
