### 2024-05-05 f80f844
add Hermitian matrix pseudo inverse
