+++ src/linalg/evd/mod.rs
@@ -335,6 +335,118 @@ pub fn compute_hermitian_evd_custom_epsilon<E: ComplexField>(
     );
 }
 
+/// Computes the size and alignment of required workspace for performing a decomposed
+/// Hermitian matrix pseudoi inverse
+pub fn compute_hermitian_pseudoinverse_req<E: ComplexField>(
+    n: usize,
+    parallelism: Parallelism,
+) -> Result<StackReq, SizeOverflow> {
+    let _ = parallelism;
+    StackReq::try_all_of([
+        temp_mat_req::<E>(n, n)?,
+        temp_mat_req::<E>(n, n)?,
+        temp_mat_req::<E::Real>(1, n)?,
+        StackReq::try_new::<usize>(n)?,
+    ])
+}
+
+/// Computes the pseudo inverse of ai decomposed square Hermitian matrix.
+///
+/// `s` represents the diagonal of the matrix $S$, and must have size equal to the dimension of the
+/// matrix.
+///
+/// `u` represents the eigenvectors.
+///
+/// `pinv` represents the pseudo inversion of decomposed matrix.
+///
+/// This can also panic if the provided memory in `stack` is insufficient (see
+/// [`compute_hermitian_pseudoinverse_req`]).
+pub fn compute_hermitian_pseudoinverse<E: ComplexField>(
+    s: MatRef<'_, E>,
+    u: MatRef<'_, E>,
+    pinv: MatMut<'_, E>,
+    parallelism: Parallelism,
+    stack: PodStack<'_>,
+) {
+    compute_hermitian_pseudoinverse_custom_epsilon(s, u, pinv, None, None, parallelism, stack);
+}
+
+/// See [`compute_hermitian_pseudoinverse`].
+///
+/// This function takes an additional `atol` and `rtol` parameters. `atol`
+/// represents the absolute threshold term, if `None` provided value is 0, and `rtol` represents the
+/// relative threshold term, if `None` provided value is `N * eps` where
+/// `eps` is the machine precision value of the `E::Real`
+///
+/// # Panics
+/// Panics if `atol` or `rtol` equal or less than zero
+pub fn compute_hermitian_pseudoinverse_custom_epsilon<E: ComplexField>(
+    s: MatRef<'_, E>,
+    u: MatRef<'_, E>,
+    mut pinv: MatMut<'_, E>,
+    atol: Option<E::Real>,
+    rtol: Option<E::Real>,
+    parallelism: Parallelism,
+    stack: PodStack<'_>,
+) {
+    let n = u.ncols();
+    let mut max_s = E::Real::faer_zero();
+    for i in 0..n {
+        let v = s.read(i, 0).faer_abs();
+        if v > max_s {
+            max_s = v;
+        }
+    }
+
+    let atol = atol.unwrap_or(E::Real::faer_zero());
+    let rtol = rtol.unwrap_or(E::Real::faer_epsilon() * E::Real::faer_from_f64(n as f64));
+    assert!(
+        rtol >= E::Real::faer_zero() && atol >= E::Real::faer_zero(),
+        "atol and rtol values must be positive.",
+    );
+    let (above_cutoff, stack) = stack.make_raw::<usize>(n);
+    let val = atol + max_s * rtol;
+
+    let mut r_take = 0;
+    for i in 0..n {
+        if s.read(i, 0).faer_abs() > val {
+            above_cutoff[r_take] = i;
+            r_take += 1;
+        }
+    }
+
+    let (mut psigma_diag, stack) = temp_mat_uninit::<E::Real>(1, r_take, stack);
+    for i in 0..r_take {
+        psigma_diag.write(
+            0,
+            i,
[…]
