### 2024-05-08 e0fbbe8
add shape function
