### 2024-05-13 74c7326
store B instead of inv(B) in lblt
