+++ src/linalg/cholesky/bunch_kaufman/mod.rs
@@ -239,7 +239,6 @@ pub mod compute {
                         }
                     }
                 }
-                let d11 = d11.faer_inv();
                 a.write(k, k, E::faer_from_real(d11));
             } else {
                 if abs_akk >= colmax.faer_mul(alpha) {
@@ -334,8 +333,8 @@ pub mod compute {
                             dynamic_regularization_count += 1;
                         }
                     }
-                    let d11 = d11.faer_inv();
                     a.write(k, k, E::faer_from_real(d11));
+                    let d11 = d11.faer_inv();
 
                     let x = a.rb_mut().subrows_mut(k + 1, n - k - 1).col_mut(k);
                     zipped!(x.as_2d_mut())
@@ -347,28 +346,32 @@ pub mod compute {
                         .as_2d_mut())
                     .for_each(|unzipped!(mut x)| x.write(x.read().faer_conj()));
                 } else {
-                    let d21 = w.read(k + 1, k).faer_abs();
-                    let d21_inv = d21.faer_inv();
-                    let mut d11 = d21_inv.faer_scale_real(w.read(k + 1, k + 1).faer_real());
-                    let mut d22 = d21_inv.faer_scale_real(w.read(k, k).faer_real());
+                    let dd = w.read(k + 1, k).faer_abs();
+                    let dd_inv = dd.faer_inv();
+                    let mut d11 = dd_inv.faer_scale_real(w.read(k + 1, k + 1).faer_real());
+                    let mut d22 = dd_inv.faer_scale_real(w.read(k, k).faer_real());
 
-                    let eps = eps.faer_mul(d21_inv);
-                    let delta = delta.faer_mul(d21_inv);
+                    let eps = eps.faer_mul(dd_inv);
+                    let delta = delta.faer_mul(dd_inv);
                     if has_eps {
                         if let Some(signs) = signs.rb_mut() {
                             if signs[k] > 0 && signs[k + 1] > 0 {
                                 if d11 <= eps {
                                     d11 = delta;
+                                    dynamic_regularization_count += 1;
                                 }
                                 if d22 <= eps {
                                     d22 = delta;
+                                    dynamic_regularization_count += 1;
                                 }
                             } else if signs[k] < 0 && signs[k + 1] < 0 {
                                 if d11 >= eps.faer_neg() {
                                     d11 = delta.faer_neg();
+                                    dynamic_regularization_count += 1;
                                 }
                                 if d22 >= eps.faer_neg() {
                                     d22 = delta.faer_neg();
+                                    dynamic_regularization_count += 1;
                                 }
                             }
                         }
@@ -393,12 +396,14 @@ pub mod compute {
                     }
 
                     let t = t.faer_inv();
-                    let d21 = w.read(k + 1, k).faer_scale_real(d21_inv);
-                    let d = t.faer_mul(d21_inv);
+                    let d21 = w.read(k + 1, k).faer_scale_real(dd_inv);
+                    let d = t.faer_mul(dd_inv);
 
-                    a.write(k, k, E::faer_from_real(d11.faer_mul(d)));
-                    a.write(k + 1, k, d21.faer_scale_real(d.faer_neg()));
-                    a.write(k + 1, k + 1, E::faer_from_real(d22.faer_mul(d)));
+                    a.write(k, k, w.read(k, k));
+                    a.write(k + 1, k, w.read(k + 1, k));
+                    a.write(k + 1, k + 1, w.read(k + 1, k + 1));
+                    make_real(a.rb_mut(), k, k);
+                    make_real(a.rb_mut(), k + 1, k + 1);
 
                     for j in k + 2..n {
                         let wk = (w
@@ -416,18 +421,10 @@ pub mod compute {
[…]
