+++ Cargo.toml
@@ -56,10 +56,12 @@ npy = ["std", "dep:npyz"]
 [dev-dependencies]
 amd = "0.2.2"
 assert_approx_eq = "1.1.0"
+core_affinity = "0.8.1"
 diol = { version = "0.8.3" }
 matrix-market-rs = "0.1.3"
 matrixcompare = "0.3.0"
 nalgebra = "0.32.5"
+num_cpus = "1.16.0"
 rand = "0.8.5"
 rand_distr = "0.4.3"
 serde_json = "1.0.116"
