+++ src/sparse/linalg/cholesky.rs
@@ -1787,46 +1787,30 @@ pub mod supernodal {
                     let subdiag = subdiag.read(idx);
                     let i = idx + s.start();
                     if subdiag == E::faer_zero() {
-                        let d = Bs.read(idx, idx).faer_real();
+                        let d = Bs.read(idx, idx).faer_real().faer_inv();
                         for j in 0..k {
                             x.write(i, j, x.read(i, j).faer_scale_real(d))
                         }
                         idx += 1;
                     } else {
-                        let d11 = Bs.read(idx, idx).faer_real();
-                        let d22 = Bs.read(idx + 1, idx + 1).faer_real();
-                        let d21 = subdiag;
-
+                        let mut d21 = subdiag;
                         if conj == Conj::Yes {
-                            for j in 0..k {
-                                let xi = x.read(i, j);
-                                let xip1 = x.read(i + 1, j);
+                            d21 = d21.faer_conj();
+                        }
+                        d21 = d21.faer_inv();
+                        let d11 = d21
+                            .faer_conj()
+                            .faer_scale_real(Bs.read(idx, idx).faer_real());
+                        let d22 = d21.faer_scale_real(Bs.read(idx + 1, idx + 1).faer_real());
 
-                                x.write(i, j, xi.faer_scale_real(d11).faer_add(xip1.faer_mul(d21)));
-                                x.write(
-                                    i + 1,
-                                    j,
-                                    xip1.faer_scale_real(d22)
-                                        .faer_add(xi.faer_mul(d21.faer_conj())),
-                                );
-                            }
-                        } else {
-                            for j in 0..k {
-                                let xi = x.read(i, j);
-                                let xip1 = x.read(i + 1, j);
+                        let denom = d11.faer_mul(d22).faer_sub(E::faer_one()).faer_inv();
 
-                                x.write(
-                                    i,
-                                    j,
-                                    xi.faer_scale_real(d11)
-                                        .faer_add(xip1.faer_mul(d21.faer_conj())),
-                                );
-                                x.write(
-                                    i + 1,
-                                    j,
-                                    xip1.faer_scale_real(d22).faer_add(xi.faer_mul(d21)),
-                                );
-                            }
+                        for j in 0..k {
+                            let xk = x.read(i, j).faer_mul(d21.faer_conj());
+                            let xkp1 = x.read(i + 1, j).faer_mul(d21);
+
+                            x.write(i, j, (d22.faer_mul(xk).faer_sub(xkp1)).faer_mul(denom));
+                            x.write(i + 1, j, (d11.faer_mul(xkp1).faer_sub(xk)).faer_mul(denom));
                         }
                         idx += 2;
                     }
@@ -3843,48 +3827,30 @@ pub mod supernodal {
                 while j < d_ncols {
                     let subdiag = d_subdiag.read(j);
                     if subdiag == E::faer_zero() {
-                        let d = Ld_top.read(j, j).faer_real().faer_inv();
+                        let d = Ld_top.read(j, j).faer_real();
                         for i in 0..d_pattern_mid_len {
                             Ld_mid_x_D.write(i, j, Ld_mid.read(i, j).faer_scale_real(d));
                         }
                         j += 1;
                     } else {
-                        // 1/d21
-                        let akp1k = subdiag.faer_inv();
-                        // d11/d21
-                        let ak = akp1k.faer_scale_real(Ld_top.read(j, j).faer_real());
[…]
