+++ src/linalg/evd/tridiag_real_evd.rs
@@ -4,14 +4,10 @@ use crate::{
         matmul::{inner_prod::inner_prod_with_conj, matmul},
         temp_mat_req, temp_mat_uninit,
     },
-    unzipped, zipped, ComplexField, Conj, Entity, MatMut, MatRef, Parallelism, RealField,
+    unzipped, zipped, ColMut, ColRef, Conj, Entity, MatMut, Parallelism, RealField,
 };
 use dyn_stack::{PodStack, SizeOverflow, StackReq};
 use reborrow::*;
-
-pub fn norm2<E: ComplexField>(v: MatRef<'_, E>) -> E::Real {
-    inner_prod_with_conj(v, Conj::Yes, v, Conj::No).faer_real()
-}
 
 fn max<T: PartialOrd>(x: T, y: T) -> T {
     if x > y {
@@ -24,8 +20,8 @@ fn max<T: PartialOrd>(x: T, y: T) -> T {
 /// the argument of the secular eq is mu + shift
 #[inline(always)]
 fn secular_eq<E: RealField>(
-    d: MatRef<'_, E>,
-    z: MatRef<'_, E>,
+    d: ColRef<'_, E>,
+    z: ColRef<'_, E>,
     rho: E,
     mu: E,
     shift: E,
@@ -50,22 +46,22 @@ fn secular_eq<E: RealField>(
         let mut i = 0;
 
         while i < n / 8 * 8 {
-            let z0 = z.read_unchecked(i, 0);
-            let d0 = d.read_unchecked(i, 0);
-            let z1 = z.read_unchecked(i + 1, 0);
-            let d1 = d.read_unchecked(i + 1, 0);
-            let z2 = z.read_unchecked(i + 2, 0);
-            let d2 = d.read_unchecked(i + 2, 0);
-            let z3 = z.read_unchecked(i + 3, 0);
-            let d3 = d.read_unchecked(i + 3, 0);
-            let z4 = z.read_unchecked(i + 4, 0);
-            let d4 = d.read_unchecked(i + 4, 0);
-            let z5 = z.read_unchecked(i + 5, 0);
-            let d5 = d.read_unchecked(i + 5, 0);
-            let z6 = z.read_unchecked(i + 6, 0);
-            let d6 = d.read_unchecked(i + 6, 0);
-            let z7 = z.read_unchecked(i + 7, 0);
-            let d7 = d.read_unchecked(i + 7, 0);
+            let z0 = z.read_unchecked(i);
+            let d0 = d.read_unchecked(i);
+            let z1 = z.read_unchecked(i + 1);
+            let d1 = d.read_unchecked(i + 1);
+            let z2 = z.read_unchecked(i + 2);
+            let d2 = d.read_unchecked(i + 2);
+            let z3 = z.read_unchecked(i + 3);
+            let d3 = d.read_unchecked(i + 3);
+            let z4 = z.read_unchecked(i + 4);
+            let d4 = d.read_unchecked(i + 4);
+            let z5 = z.read_unchecked(i + 5);
+            let d5 = d.read_unchecked(i + 5);
+            let z6 = z.read_unchecked(i + 6);
+            let d6 = d.read_unchecked(i + 6);
+            let z7 = z.read_unchecked(i + 7);
+            let d7 = d.read_unchecked(i + 7);
 
             // res = res + z * (z / (d - arg))
             res0 = res0.faer_add(z0.faer_mul(z0.faer_div(d0.faer_sub(shift).faer_sub(mu))));
@@ -81,8 +77,8 @@ fn secular_eq<E: RealField>(
         }
 
         while i < n {
-            let z0 = z.read_unchecked(i, 0);
-            let d0 = d.read_unchecked(i, 0);
[…]
