+++ src/linalg/evd/tridiag.rs
@@ -3,7 +3,7 @@ use crate::{
     linalg::{matmul::inner_prod::inner_prod_with_conj, temp_mat_req, temp_mat_zeroed},
     unzipped,
     utils::thread::parallelism_degree,
-    zipped, Conj, MatMut, MatRef, Parallelism,
+    zipped, ColMut, ColRef, Conj, MatMut, MatRef, Parallelism,
 };
 use core::iter::zip;
 use dyn_stack::{PodStack, SizeOverflow, StackReq};
@@ -24,20 +24,20 @@ pub fn tridiagonalize_in_place_req<E: Entity>(
 }
 
 struct SymMatVecWithLhsUpdate<'a, E: Entity> {
-    acc: MatMut<'a, E>,
-    acct: MatMut<'a, E>,
+    acc: ColMut<'a, E>,
+    acct: ColMut<'a, E>,
     lhs: MatMut<'a, E>,
-    rhs: MatRef<'a, E>,
-    u: MatRef<'a, E>,
-    y: MatRef<'a, E>,
+    rhs: ColRef<'a, E>,
+    u: ColRef<'a, E>,
+    y: ColRef<'a, E>,
     first_col: usize,
     last_col: usize,
 }
 
 struct SymMatVec<'a, E: Entity> {
-    acc: MatMut<'a, E>,
+    acc: ColMut<'a, E>,
     lhs: MatRef<'a, E>,
-    rhs: MatRef<'a, E>,
+    rhs: ColRef<'a, E>,
     first_col: usize,
     last_col: usize,
 }
@@ -988,21 +988,22 @@ pub fn tridiagonalize_in_place<E: ComplexField>(
         return;
     }
 
-    let (mut u, stack) = temp_mat_zeroed::<E>(n, 1, stack);
-    let (mut y, stack) = temp_mat_zeroed::<E>(n, 1, stack);
+    let (u, stack) = temp_mat_zeroed::<E>(n, 1, stack);
+    let (y, stack) = temp_mat_zeroed::<E>(n, 1, stack);
 
     let (mut v, stack) = temp_mat_zeroed::<E>(n, parallelism_degree(parallelism), stack);
     let (mut w, _) = temp_mat_zeroed::<E>(n, parallelism_degree(parallelism), stack);
 
-    let mut u = u.as_mut();
-    let mut y = y.as_mut();
+    let mut u = u.col_mut(0);
+    let mut y = y.col_mut(0);
     let mut v = v.as_mut();
     let mut w = w.as_mut();
 
     let arch = E::Simd::default();
     for k in 0..n - 1 {
         let a_cur = a.rb_mut().submatrix_mut(k, k, n - k, n - k);
-        let (mut a11, _, mut a21, a22) = a_cur.split_at_mut(1, 1);
+        let (mut a11, _, a21, a22) = a_cur.split_at_mut(1, 1);
+        let mut a21 = a21.col_mut(0);
 
         let parallelism = if n - k <= 256 {
             Parallelism::None
@@ -1010,10 +1011,10 @@ pub fn tridiagonalize_in_place<E: ComplexField>(
             parallelism
         };
 
-        let (_, u) = u.rb_mut().split_at_row_mut(k);
-        let (nu, mut u21) = u.split_at_row_mut(1);
-        let (_, y) = y.rb_mut().split_at_row_mut(k);
-        let (psi, mut y21) = y.split_at_row_mut(1);
+        let (_, u) = u.rb_mut().split_at_mut(k);
+        let (nu, mut u21) = u.split_at_mut(1);
+        let (_, y) = y.rb_mut().split_at_mut(k);
+        let (psi, mut y21) = y.split_at_mut(1);
 
[…]
