### 2024-05-18 a302ea8
allow col/row to be used in matmul
