+++ src/linalg/cholesky/llt/mod.rs
@@ -35,13 +35,13 @@ impl std::error::Error for CholeskyError {}
 
 #[cfg(test)]
 mod tests {
-    use assert_approx_eq::assert_approx_eq;
-    use dyn_stack::{GlobalPodBuffer, PodStack};
-
     use super::{compute::*, inverse::*, reconstruct::*, solve::*, update::*};
     use crate::{
-        complex_native::c64, linalg::matmul as mul, ComplexField, Conj, Mat, MatRef, Parallelism,
+        complex_native::c64, linalg::matmul as mul, Col, ComplexField, Conj, Mat, MatRef,
+        Parallelism,
     };
+    use assert_approx_eq::assert_approx_eq;
+    use dyn_stack::{GlobalPodBuffer, PodStack};
 
     type E = c64;
 
@@ -211,13 +211,13 @@ mod tests {
             let mut a = random_positive_definite(n);
             let mut a_updated = a.clone();
             let mut w = Mat::from_fn(n, k, |_, _| random());
-            let mut alpha = Mat::from_fn(k, 1, |_, _| E::faer_from_real(rand::random()));
-            let alpha = alpha.as_mut().col_mut(0).as_2d_mut();
+            let mut alpha = Col::from_fn(k, |_| E::faer_from_real(rand::random()));
+            let alpha = alpha.as_mut();
 
             let mut w_alpha = Mat::zeros(n, k);
             for j in 0..k {
                 for i in 0..n {
-                    w_alpha.write(i, j, alpha.read(j, 0).faer_mul(w.read(i, j)));
+                    w_alpha.write(i, j, alpha.read(j).faer_mul(w.read(i, j)));
                 }
             }
 
@@ -242,7 +242,7 @@ mod tests {
             )
             .unwrap();
 
-            rank_r_update_clobber(a.as_mut(), w.as_mut(), alpha.col_mut(0)).unwrap();
+            rank_r_update_clobber(a.as_mut(), w.as_mut(), alpha).unwrap();
 
             let a_reconstructed = reconstruct_matrix(a.as_ref());
 
