### 2024-05-18 d5a62f4
document iterative solvers, simplify constructors
