### 2024-05-20 f1dab94
more linear operators
