+++ src/linop/linop_impl/mod.rs
@@ -1,3 +1,15 @@
 mod matmut;
 mod matown;
 mod matref;
+
+mod csc_matmut;
+mod csc_matown;
+mod csc_matref;
+
+mod csr_matmut;
+mod csr_matown;
+mod csr_matref;
+
+mod diagmut;
+mod diagown;
+mod diagref;
