+++ src/diag/diagref.rs
@@ -2,6 +2,7 @@ use super::*;
 use crate::col::ColRef;
 
 /// Diagonal matrix view.
+#[derive(Debug)]
 pub struct DiagRef<'a, E: Entity> {
     pub(crate) inner: ColRef<'a, E>,
 }

[…]
