### 2024-05-20 3449993
fix flaky pseudoinverse test
