+++ faer-no-std-test/src/main.rs
@@ -10,15 +10,32 @@ use core::alloc::{GlobalAlloc, Layout};
 #[derive(Default)]
 pub struct Allocator;
 
+#[cfg(windows)]
+extern "C" {
+    fn _aligned_malloc(size: libc::size_t, align: libc::size_t) -> *mut libc::c_void;
+    fn _aligned_free(memblock: *mut libc::c_void);
+}
+
 unsafe impl GlobalAlloc for Allocator {
     unsafe fn alloc(&self, layout: Layout) -> *mut u8 {
-        let mut mem: *mut libc::c_void = core::ptr::null_mut();
-        libc::posix_memalign(&mut mem, layout.align(), layout.size());
-        assert!(!mem.is_null());
-        mem as *mut u8
+        #[cfg(not(windows))]
+        {
+            let mut mem: *mut libc::c_void = core::ptr::null_mut();
+            libc::posix_memalign(&mut mem, layout.align(), layout.size());
+            mem as *mut u8
+        }
+
+        #[cfg(windows)]
+        {
+            _aligned_malloc(layout.size(), layout.align()) as *mut u8
+        }
     }
     unsafe fn dealloc(&self, ptr: *mut u8, _layout: Layout) {
+        #[cfg(not(windows))]
         libc::free(ptr as *mut libc::c_void);
+
+        #[cfg(windows)]
+        _aligned_free(ptr as *mut libc::c_void)
     }
 }
 
@@ -37,7 +54,6 @@ fn my_allocator_error(_layout: Layout) -> ! {
     panic!("out of memory");
 }
 
-#[cfg(not(tests))]
 #[lang = "eh_personality"]
 #[no_mangle]
 pub extern "C" fn rust_eh_personality() {}
