### 2024-05-21 0e8cbd4
allow using custom permutations with sparse cholesky
