+++ src/sparse/linalg/cholesky.rs
@@ -524,13 +524,14 @@
 // implementation inspired by https://gitlab.com/hodge_star/catamari
 
 use super::{
-    amd::{self, Control},
+    amd::{self, Control, FlopCount},
     ghost::{self, Array, Idx, MaybeIdx},
     ghost_permute_hermitian_unsorted, ghost_permute_hermitian_unsorted_symbolic, make_raw_req, mem,
     mem::NONE,
-    nomem, triangular_solve, try_collect, try_zeroed, windows2, FaerError, Index, PermRef, Side,
-    SliceGroup, SliceGroupMut, SparseColMatRef, SupernodalThreshold, SymbolicSparseColMatRef,
-    SymbolicSupernodalParams,
+    nomem, triangular_solve, try_collect, try_zeroed,
+    utils::ghost_adjoint,
+    windows2, FaerError, Index, PermRef, Side, SliceGroup, SliceGroupMut, SparseColMatRef,
+    SupernodalThreshold, SymbolicSparseColMatRef, SymbolicSupernodalParams,
 };
 pub use crate::linalg::cholesky::{
     bunch_kaufman::compute::BunchKaufmanRegularization,
@@ -547,17 +548,20 @@ use dyn_stack::{PodStack, SizeOverflow, StackReq};
 use faer_entity::GroupFor;
 use reborrow::*;
 
-#[derive(Copy, Clone)]
-#[allow(dead_code)]
-enum Ordering<'a, I: Index> {
+/// Fill reducing ordering to use for the Cholesky factorization.
+#[derive(Copy, Clone, Default)]
+pub enum SymmetricOrdering<'a, I: Index> {
+    /// Approximate minimum degree ordering. Default option.
+    #[default]
+    Amd,
+    /// No reordering.
     Identity,
-    Custom(&'a [I]),
+    /// Custom reordering.
     Algorithm(
         &'a dyn Fn(
             &mut [I],                       // perm
             &mut [I],                       // perm_inv
             SymbolicSparseColMatRef<'_, I>, // A
-            PodStack<'_>,
         ) -> Result<(), FaerError>,
     ),
 }
@@ -4160,8 +4164,8 @@ pub enum SymbolicCholeskyRaw<I: Index> {
 #[derive(Debug)]
 pub struct SymbolicCholesky<I: Index> {
     raw: SymbolicCholeskyRaw<I>,
-    perm_fwd: alloc::vec::Vec<I>,
-    perm_inv: alloc::vec::Vec<I>,
+    perm_fwd: Option<alloc::vec::Vec<I>>,
+    perm_inv: Option<alloc::vec::Vec<I>>,
     A_nnz: usize,
 }
 
@@ -4189,8 +4193,13 @@ impl<I: Index> SymbolicCholesky<I> {
 
     /// Returns the permutation that was computed during symbolic analysis.
     #[inline]
-    pub fn perm(&self) -> PermRef<'_, I> {
-        unsafe { PermRef::new_unchecked(&self.perm_fwd, &self.perm_inv) }
+    pub fn perm(&self) -> Option<PermRef<'_, I>> {
+        match (&self.perm_fwd, &self.perm_inv) {
+            (Some(perm_fwd), Some(perm_inv)) => unsafe {
+                Some(PermRef::new_unchecked(perm_fwd, perm_inv))
+            },
+            _ => None,
+        }
     }
 
     /// Returns the length of the slice needed to store the numerical values of the Cholesky
@@ -4336,8 +4345,6 @@ impl<I: Index> SymbolicCholesky<I> {
             let A_nnz = self.A_nnz;
             let A = ghost::SparseColMatRef::new(A, N, N);
 
[…]
