+++ src/mat/matown.rs
@@ -2,6 +2,7 @@ use super::*;
 use crate::{
     assert, debug_assert,
     diag::{DiagMut, DiagRef},
+    iter,
     mat::matalloc::{align_for, is_vectorizable, MatUnit, RawMat, RawMatUnit},
     utils::DivCeil,
 };
@@ -533,6 +534,86 @@ impl<E: Entity> Mat<E> {
                 self.col_stride(),
             )
         }
+    }
+
+    /// Returns a reference to the first column and a view over the remaining ones if the matrix has
+    /// at least one column, otherwise `None`.
+    #[inline]
+    pub fn split_first_col(&self) -> Option<(ColRef<'_, E>, MatRef<'_, E>)> {
+        self.as_ref().split_first_col()
+    }
+
+    /// Returns a reference to the last column and a view over the remaining ones if the matrix has
+    /// at least one column,  otherwise `None`.
+    #[inline]
+    pub fn split_last_col(&self) -> Option<(ColRef<'_, E>, MatRef<'_, E>)> {
+        self.as_ref().split_last_col()
+    }
+
+    /// Returns a reference to the first row and a view over the remaining ones if the matrix has
+    /// at least one row, otherwise `None`.
+    #[inline]
+    pub fn split_first_row(&self) -> Option<(RowRef<'_, E>, MatRef<'_, E>)> {
+        self.as_ref().split_first_row()
+    }
+
+    /// Returns a reference to the last row and a view over the remaining ones if the matrix has
+    /// at least one row,  otherwise `None`.
+    #[inline]
+    pub fn split_last_row(&self) -> Option<(RowRef<'_, E>, MatRef<'_, E>)> {
+        self.as_ref().split_last_row()
+    }
+
+    /// Returns a reference to the first column and a view over the remaining ones if the matrix has
+    /// at least one column, otherwise `None`.
+    #[inline]
+    pub fn split_first_col_mut(&mut self) -> Option<(ColMut<'_, E>, MatMut<'_, E>)> {
+        self.as_mut().split_first_col_mut()
+    }
+
+    /// Returns a reference to the last column and a view over the remaining ones if the matrix has
+    /// at least one column,  otherwise `None`.
+    #[inline]
+    pub fn split_last_col_mut(&mut self) -> Option<(ColMut<'_, E>, MatMut<'_, E>)> {
+        self.as_mut().split_last_col_mut()
+    }
+
+    /// Returns a reference to the first row and a view over the remaining ones if the matrix has
+    /// at least one row, otherwise `None`.
+    #[inline]
+    pub fn split_first_row_mut(&mut self) -> Option<(RowMut<'_, E>, MatMut<'_, E>)> {
+        self.as_mut().split_first_row_mut()
+    }
+
+    /// Returns a reference to the last row and a view over the remaining ones if the matrix has
+    /// at least one row,  otherwise `None`.
+    #[inline]
+    pub fn split_last_row_mut(&mut self) -> Option<(RowMut<'_, E>, MatMut<'_, E>)> {
+        self.as_mut().split_last_row_mut()
+    }
+
+    /// Returns an iterator over the columns of the matrix.
+    #[inline]
+    pub fn col_iter(&self) -> iter::ColIter<'_, E> {
+        self.as_ref().col_iter()
+    }
+
+    /// Returns an iterator over the rows of the matrix.
+    #[inline]
+    pub fn row_iter(&self) -> iter::RowIter<'_, E> {
+        self.as_ref().row_iter()
+    }
+
+    /// Returns an iterator over the columns of the matrix.
+    #[inline]
+    pub fn col_iter_mut(&mut self) -> iter::ColIterMut<'_, E> {
+        self.as_mut().col_iter_mut()
+    }
+
[…]
