+++ src/iter/mod.rs
@@ -1,0 +1,535 @@
+use crate::{linalg::entity::GroupFor, ColMut, ColRef, Entity, MatMut, MatRef, RowMut, RowRef};
+
+use self::chunks::ChunkPolicy;
+
+/// Fixed-size chunked column iterator over the elements.
+#[derive(Debug, Clone)]
+pub struct ColElemPartition<'a, E: Entity> {
+    pub(crate) inner: ColRef<'a, E>,
+    pub(crate) policy: chunks::PartitionCountPolicy,
+}
+/// Fixed-size chunked column iterator over the elements.
+#[derive(Debug)]
+pub struct ColElemPartitionMut<'a, E: Entity> {
+    pub(crate) inner: ColMut<'a, E>,
+    pub(crate) policy: chunks::PartitionCountPolicy,
+}
+/// Fixed-size chunked row iterator over the elements.
+#[derive(Debug, Clone)]
+pub struct RowElemPartition<'a, E: Entity> {
+    pub(crate) inner: RowRef<'a, E>,
+    pub(crate) policy: chunks::PartitionCountPolicy,
+}
+/// Fixed-size chunked row iterator over the elements.
+#[derive(Debug)]
+pub struct RowElemPartitionMut<'a, E: Entity> {
+    pub(crate) inner: RowMut<'a, E>,
+    pub(crate) policy: chunks::PartitionCountPolicy,
+}
+
+/// Fixed-size chunked matrix iterator over the columns.
+#[derive(Debug, Clone)]
+pub struct ColPartition<'a, E: Entity> {
+    pub(crate) inner: MatRef<'a, E>,
+    pub(crate) policy: chunks::PartitionCountPolicy,
+}
+/// Fixed-size chunked matrix iterator over the columns.
+#[derive(Debug)]
+pub struct ColPartitionMut<'a, E: Entity> {
+    pub(crate) inner: MatMut<'a, E>,
+    pub(crate) policy: chunks::PartitionCountPolicy,
+}
+/// Fixed-size chunked matrix iterator over the rows.
+#[derive(Debug, Clone)]
+pub struct RowPartition<'a, E: Entity> {
+    pub(crate) inner: MatRef<'a, E>,
+    pub(crate) policy: chunks::PartitionCountPolicy,
+}
+/// Fixed-size chunked matrix iterator over the rows.
+#[derive(Debug)]
+pub struct RowPartitionMut<'a, E: Entity> {
+    pub(crate) inner: MatMut<'a, E>,
+    pub(crate) policy: chunks::PartitionCountPolicy,
+}
+
+/// Chunked column iterator over the elements.
+#[derive(Debug, Clone)]
+pub struct ColElemChunks<'a, E: Entity> {
+    pub(crate) inner: ColRef<'a, E>,
+    pub(crate) policy: chunks::ChunkSizePolicy,
+}
+/// Chunked column iterator over the elements.
+#[derive(Debug)]
+pub struct ColElemChunksMut<'a, E: Entity> {
+    pub(crate) inner: ColMut<'a, E>,
+    pub(crate) policy: chunks::ChunkSizePolicy,
+}
+/// Chunked row iterator over the elements.
+#[derive(Debug, Clone)]
+pub struct RowElemChunks<'a, E: Entity> {
+    pub(crate) inner: RowRef<'a, E>,
+    pub(crate) policy: chunks::ChunkSizePolicy,
+}
+/// Chunked row iterator over the elements.
+#[derive(Debug)]
+pub struct RowElemChunksMut<'a, E: Entity> {
+    pub(crate) inner: RowMut<'a, E>,
+    pub(crate) policy: chunks::ChunkSizePolicy,
+}
[…]
