+++ src/row/rowmut.rs
@@ -811,6 +811,10 @@ impl<'a, E: Entity> RowMut<'a, E> {
 
     /// Returns an iterator that provides successive chunks of the elements of this row, with
     /// each having at most `chunk_size` elements.
+    ///
+    /// Only available with the `rayon` feature.
+    #[cfg(feature = "rayon")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "rayon")))]
     #[inline]
     #[track_caller]
     pub fn par_chunks(
@@ -822,6 +826,10 @@ impl<'a, E: Entity> RowMut<'a, E> {
 
     /// Returns an iterator that provides exactly `count` successive chunks of the elements of this
     /// row.
+    ///
+    /// Only available with the `rayon` feature.
+    #[cfg(feature = "rayon")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "rayon")))]
     #[inline]
     #[track_caller]
     pub fn par_partition(
@@ -862,6 +870,10 @@ impl<'a, E: Entity> RowMut<'a, E> {
 
     /// Returns an iterator that provides successive chunks of the elements of this row, with
     /// each having at most `chunk_size` elements.
+    ///
+    /// Only available with the `rayon` feature.
+    #[cfg(feature = "rayon")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "rayon")))]
     #[inline]
     #[track_caller]
     pub fn par_chunks_mut(
@@ -876,6 +888,10 @@ impl<'a, E: Entity> RowMut<'a, E> {
 
     /// Returns an iterator that provides exactly `count` successive chunks of the elements of this
     /// row.
+    ///
+    /// Only available with the `rayon` feature.
+    #[cfg(feature = "rayon")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "rayon")))]
     #[inline]
     #[track_caller]
     pub fn par_partition_mut(
