+++ src/row/rowown.rs
@@ -934,6 +934,10 @@ impl<E: Entity> Row<E> {
 
     /// Returns an iterator that provides successive chunks of the elements of this row, with
     /// each having at most `chunk_size` elements.
+    ///
+    /// Only available with the `rayon` feature.
+    #[cfg(feature = "rayon")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "rayon")))]
     #[inline]
     #[track_caller]
     pub fn par_chunks(
@@ -945,6 +949,10 @@ impl<E: Entity> Row<E> {
 
     /// Returns an iterator that provides exactly `count` successive chunks of the elements of this
     /// row.
+    ///
+    /// Only available with the `rayon` feature.
+    #[cfg(feature = "rayon")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "rayon")))]
     #[inline]
     #[track_caller]
     pub fn par_partition(
@@ -972,6 +980,10 @@ impl<E: Entity> Row<E> {
 
     /// Returns an iterator that provides successive chunks of the elements of this row, with
     /// each having at most `chunk_size` elements.
+    ///
+    /// Only available with the `rayon` feature.
+    #[cfg(feature = "rayon")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "rayon")))]
     #[inline]
     #[track_caller]
     pub fn par_chunks_mut(
@@ -983,6 +995,10 @@ impl<E: Entity> Row<E> {
 
     /// Returns an iterator that provides exactly `count` successive chunks of the elements of this
     /// row.
+    ///
+    /// Only available with the `rayon` feature.
+    #[cfg(feature = "rayon")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "rayon")))]
     #[inline]
     #[track_caller]
     pub fn par_partition_mut(
