+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 - Implemented conjugate gradient, BiCGSTAB, and LSMR iterative solvers (currently gated by the `unstable` feature).
 - Implemented Hermitian matrix pseudoinverse implementation. Thanks @lishen_ for the contribution.
 - Implemented column and row mean and variance in `faer::stats`.
+- Added more iterator and parallel iterator functions (`MatRef::[col|row]_partition`, `MatRef::par_[col|row]_partition`, etc.).
 - Added `full` and `zeros` constructors to owned Col, Row, and Matrix ([issue-125](https://github.com/sarah-ek/faer-rs/issues/125)).
 - Added `shape` function to return both the row and the column count of a matrix.
 - Added several missing associated functions from the mut and owning variants of matrices.
+- Sparse Cholesky can now be used with user-provided permutations.
 - Simplified matrix constructors, adding a variant with a `_generic` prefix for the old behavior.
 - Bunch-Kaufman decomposition now stores the diagonal blocks instead of their inverses. This helps avoid infinities and NaNs when dealing with singular matrices.
 
