+++ Cargo.toml
@@ -58,6 +58,7 @@ npy = ["std", "dep:npyz"]
 amd = "0.2.2"
 assert_approx_eq = "1.1.0"
 core_affinity = "0.8.1"
+csv = "1.3.0"
 diol = { version = "0.8.3" }
 matrix-market-rs = "0.1.3"
 matrixcompare = "0.3.0"
