+++ README.md
@@ -24,308 +24,4 @@ The current MSRV is Rust 1.67.0.
 
 # Benchmarks
 
-The benchmarks were run on an `11th Gen Intel(R) Core(TM) i5-11400 @ 2.60GHz` with 12 threads.  
-- `nalgebra` is used with the `matrixmultiply` backend
-- `ndarray` is used with the `openblas` backend
-- `eigen` is compiled with `-march=native -O3 -fopenmp`
-
-All computations are done on column major matrices containing `f64` values.
-
-## Matrix multiplication
-
-Multiplication of two square matrices of dimension `n`.
-
-```
-    n       faer  faer(par)    ndarray   nalgebra      eigen
-    4       40ns       41ns      139ns       29ns       17ns
-    8       77ns       80ns       63ns      161ns       85ns
-   16      189ns      193ns      201ns      363ns      219ns
-   32      1.1µs      1.1µs      1.1µs      1.5µs      1.2µs
-   64      7.9µs      7.9µs      7.9µs     10.5µs      5.1µs
-   96     27.5µs     11.2µs     26.2µs     34.9µs     10.1µs
-  128     65.5µs     17.1µs     35.7µs     78.3µs     32.9µs
-  192    216.6µs     54.4µs     57.3µs    260.7µs     51.7µs
-  256    510.8µs    117.8µs    183.2µs    602.6µs    142.9µs
-  384      1.7ms    339.1µs    575.8µs        2ms    327.9µs
-  512        4ms    785.6µs      1.3ms      4.7ms        1ms
-  640      7.9ms      1.6ms      2.3ms      9.2ms      1.9ms
-  768     13.8ms      2.9ms      3.6ms       16ms      3.2ms
-  896     22.2ms      4.6ms      6.5ms     25.7ms      5.9ms
- 1024     33.9ms      6.6ms      9.7ms     39.1ms      8.3ms
-```
-
-## Triangular solve
-
-Solving `AX = B` in place where `A` and `B` are two square matrices of dimension `n`, and `A` is a triangular matrix.
-
-```
-    n       faer  faer(par)    ndarray   nalgebra      eigen
-    4       20ns       19ns      755ns       39ns       65ns
-    8      118ns      118ns      1.5µs      308ns      156ns
-   16      498ns      502ns      3.3µs      1.5µs      671ns
-   32      2.1µs      2.1µs      8.6µs      6.6µs      2.9µs
-   64      9.7µs      9.8µs     25.9µs     34.2µs     13.8µs
-   96     27.7µs     24.5µs     55.2µs    101.4µs     36.9µs
-  128     56.4µs     39.9µs    145.2µs      232µs     81.7µs
-  192    167.8µs       92µs    263.6µs    815.5µs    213.6µs
-  256    367.7µs      163µs      660µs      1.9ms    488.1µs
-  384      1.1ms    317.5µs      1.4ms      7.4ms      1.4ms
-  512      2.6ms    662.7µs      3.5ms     17.2ms      3.3ms
-  640      4.7ms      1.2ms      5.7ms     33.6ms      5.5ms
-  768        8ms      2.3ms      9.4ms     56.2ms      9.3ms
-  896     12.3ms      3.6ms     13.6ms     89.3ms       14ms
- 1024     18.7ms      5.2ms     20.1ms    131.9ms     22.9ms
-```
-
-## Triangular inverse
-
-Computing `A^-1` where `A` is a square triangular matrix with dimension `n`.
-
-```
-    n       faer  faer(par)    ndarray   nalgebra      eigen
-    4      162ns      5.2µs      771ns       38ns       65ns
[…]
