### 2024-05-22 eca1244
update bench script
