### 2024-05-22 5fb1090
impl Sum and Product for c32 and c64
