+++ Cargo.toml
@@ -22,7 +22,7 @@ equator = "0.1.10"
 faer-entity = { version ="0.18.0", default-features = false, path = "./faer-entity" }
 
 gemm = { version = "0.17.1", default-features = false }
-nano-gemm = { version = "0.1.0", default-features = false }
+nano-gemm = { version = "0.1.1", default-features = false }
 num-complex = { version = "0.4.5", default-features = false }
 num-traits = { version = "0.2.18", default-features = false }
 
