+++ Cargo.toml
@@ -18,11 +18,11 @@ paste = "1.0.14"
 reborrow = "0.5.5"
 
 dyn-stack = { version = "0.10.0", default-features = false }
-equator = "0.1.10"
+equator = "0.2.2"
 faer-entity = { version ="0.18.0", default-features = false, path = "./faer-entity" }
 
 gemm = { version = "0.17.1", default-features = false }
-nano-gemm = { version = "0.1.1", default-features = false }
+nano-gemm = { version = "0.1.2", default-features = false }
 num-complex = { version = "0.4.5", default-features = false }
 num-traits = { version = "0.2.18", default-features = false }
 
