### 2024-05-23 749d353
slight optimizations for small matrices
