+++ src/linalg/cholesky/llt/compute.rs
@@ -20,7 +20,8 @@ fn cholesky_in_place_left_looking_impl<E: ComplexField>(
     let mut matrix = matrix;
     let _ = params;
     let _ = parallelism;
-    assert_eq!(matrix.ncols(), matrix.nrows());
+    assert!(matrix.ncols() == matrix.nrows());
+    assert!(matrix.row_stride() == 1);
 
     let n = matrix.nrows();
 
@@ -29,7 +30,7 @@ fn cholesky_in_place_left_looking_impl<E: ComplexField>(
     }
 
     let mut idx = 0;
-    let arch = E::Simd::default();
+    let arch = E::ScalarSimd::default();
     let eps = regularization
         .dynamic_regularization_epsilon
         .faer_real()
@@ -40,6 +41,7 @@ fn cholesky_in_place_left_looking_impl<E: ComplexField>(
         .faer_abs();
     let has_eps = delta > E::Real::faer_zero();
     let mut dynamic_regularization_count = 0usize;
+
     loop {
         let block_size = 1;
 
@@ -68,15 +70,23 @@ fn cholesky_in_place_left_looking_impl<E: ComplexField>(
         // L20×L00^H  L20×L10^H + L21×L11^H  L20×L20^H + L21×L21^H + L22×L22^H
 
         // A11 -= L10 × L10^H
-        let mut dot = E::Real::faer_zero();
-        for j in 0..idx {
-            dot = dot.faer_add(l10.read(j).faer_abs2());
+        let mut dot0 = E::Real::faer_zero();
+        let mut dot1 = E::Real::faer_zero();
+        let mut dot2 = E::Real::faer_zero();
+        let mut dot3 = E::Real::faer_zero();
+        for j in 0..idx / 4 {
+            let j = j * 4;
+            dot0 += l10.read(j + 0).faer_abs2();
+            dot1 += l10.read(j + 1).faer_abs2();
+            dot2 += l10.read(j + 2).faer_abs2();
+            dot3 += l10.read(j + 3).faer_abs2();
         }
-        a11.write(
-            0,
-            0,
-            E::faer_from_real(a11.read(0, 0).faer_real().faer_sub(dot)),
-        );
+        for j in idx / 4 * 4..idx / 4 {
+            dot0 += l10.read(j + 0).faer_abs2();
+        }
+
+        let dot = (dot0 + dot1) + (dot2 + dot3);
+        a11.write(0, 0, E::faer_from_real(a11.read(0, 0).faer_real() - dot));
 
         let mut real = a11.read(0, 0).faer_real();
         if has_eps && real >= E::Real::faer_zero() && real <= eps {
@@ -110,11 +120,9 @@ fn cholesky_in_place_left_looking_impl<E: ComplexField>(
                 let l20_col = l20.col(j);
                 let l10_conj = l10.read(j).faer_conj();
 
-                zipped!(a21.rb_mut().as_2d_mut(), l20_col.as_2d()).for_each(
-                    |unzipped!(mut dst, src)| {
-                        dst.write(dst.read().faer_sub(src.read().faer_mul(l10_conj)))
-                    },
-                );
+                zipped!(a21.rb_mut(), l20_col).for_each(|unzipped!(mut dst, src)| {
+                    dst.write(dst.read().faer_sub(src.read().faer_mul(l10_conj)))
+                });
             }
         }
 
@@ -124,11 +132,11 @@ fn cholesky_in_place_left_looking_impl<E: ComplexField>(
         // conj(L11) L21^T = A21^T
 
         let r = l11.faer_real().faer_inv();
-        zipped!(a21.rb_mut().as_2d_mut())
[…]
