### 2024-05-24 ff40fa9
optimize cholesky
