+++ src/linalg/cholesky/ldlt_diagonal/compute.rs
@@ -1,67 +1,628 @@
 use crate::{
     assert, debug_assert,
     linalg::{
-        matmul::triangular::BlockStructure, temp_mat_req, temp_mat_uninit,
-        triangular_solve as solve,
+        entity,
+        entity::{from_copy, GroupFor, SimdCtx, SimdGroupFor, SimdUnitFor},
+        matmul::triangular::BlockStructure,
+        temp_mat_req, temp_mat_uninit, triangular_solve as solve,
     },
+    solvers::CholeskyError,
+    sparse::linalg::cholesky::LltRegularization,
     unzipped,
     utils::{simd::*, slice::*},
-    zipped, ColMut, Conj, MatMut, MatRef, Parallelism, RowRef,
+    zipped, ComplexField, MatMut, Parallelism,
 };
+use core::{convert::Infallible, marker::PhantomData};
 use dyn_stack::{PodStack, SizeOverflow, StackReq};
 use faer_entity::*;
+use pulp::Simd;
 use reborrow::*;
 
-pub(crate) struct RankUpdate<'a, E: ComplexField> {
-    pub a21: ColMut<'a, E>,
-    pub l20: MatRef<'a, E>,
-    pub l10: RowRef<'a, E>,
+#[inline(always)]
+fn first_elem<E: ComplexField, S: Simd>(values: SimdGroupFor<E, S>) -> E {
+    E::faer_from_units(E::faer_map(
+        from_copy::<E, _>(values),
+        #[inline(always)]
+        |x| pulp::cast_lossy::<SimdUnitFor<E, S>, E::Unit>(x),
+    ))
 }
 
-impl<E: ComplexField> pulp::WithSimd for RankUpdate<'_, E> {
-    type Output = ();
+#[inline(always)]
+fn as_simd<E: ComplexField, S: Simd>(
+    slice: GroupFor<E, &[E::Unit]>,
+) -> (GroupFor<E, &[SimdUnitFor<E, S>]>, GroupFor<E, &[E::Unit]>) {
+    use crate::linalg::entity;
 
-    #[inline(always)]
-    fn with_simd<S: pulp::Simd>(self, simd: S) -> Self::Output {
-        let Self { a21, l20, l10 } = self;
+    let slice = SliceGroup::<E>::new(slice);
 
-        assert_eq!(a21.row_stride(), 1);
-        assert_eq!(l20.row_stride(), 1);
-        assert_eq!(l20.nrows(), a21.nrows());
-        assert_eq!(l20.ncols(), l10.ncols());
+    let lanes = core::mem::size_of::<entity::SimdUnitFor<E, S>>() / core::mem::size_of::<E::Unit>();
+    let len = slice.len();
 
-        let m = l20.nrows();
-        let n = l20.ncols();
+    let mut prefix = len / lanes * lanes;
+    if len % lanes == 0 {
+        prefix -= lanes;
+    }
 
-        if m == 0 {
-            return;
-        }
+    let (head, tail) = slice.split_at(prefix);
+    let head = entity::slice_as_simd::<E, S>(head.into_inner()).0;
+    (head, tail.into_inner())
+}
 
-        let simd = SimdFor::<E, S>::new(simd);
-        let mut acc = SliceGroupMut::<'_, E>::new(a21.try_get_contiguous_col_mut());
+#[inline(always)]
+fn as_simd_mut<E: ComplexField, S: Simd>(
+    slice: GroupFor<E, &mut [E::Unit]>,
+) -> (
+    GroupFor<E, &mut [SimdUnitFor<E, S>]>,
+    GroupFor<E, &mut [E::Unit]>,
+) {
+    use crate::linalg::entity;
 
-        for j in 0..n {
[…]
