+++ benches/cholesky.rs
@@ -86,11 +86,11 @@ fn main() -> std::io::Result<()> {
     let mut bench = Bench::new(BenchConfig::from_args()?);
     bench.register_many(
         list![faer::<f32>, nalgebra::<f32>],
-        [1, 2, 4, 8, 16, 32, 64, 128].map(PlotArg),
+        [1, 2, 4, 8, 16, 32, 64, 128, 256].map(PlotArg),
     );
     bench.register_many(
         list![faer::<f64>, nalgebra::<f64>],
-        [1, 2, 4, 8, 16, 32, 64, 128].map(PlotArg),
+        [1, 2, 4, 8, 16, 32, 64, 128, 256].map(PlotArg),
     );
     bench.run()?;
 

[…]
