### 2024-05-24 2170736
handle non contiguous matrices in new cholesky
