### 2024-05-24 9706431
msrv 1.67
