+++ src/linalg/matmul/mod.rs
@@ -149,14 +149,246 @@ pub mod inner_prod {
             acc6 = simd.conditional_conj_mul_add_e(conj, a6.read_or(zero), b6.read_or(zero), acc6);
             acc7 = simd.conditional_conj_mul_add_e(conj, a7.read_or(zero), b7.read_or(zero), acc7);
         }
-        for (a, b) in zip(a_body1.into_ref_iter(), b_body1.into_ref_iter()) {
-            acc0 = simd.conditional_conj_mul_add_e(conj, a.read_or(zero), b.read_or(zero), acc0);
+        unsafe {
+            match a_body1.len() {
+                0 => {
+                    acc7 = simd.conditional_conj_mul_add_e(
+                        conj,
+                        a_tail.read_or(zero),
+                        b_tail.read_or(zero),
+                        acc7,
+                    );
+                }
+                1 => {
+                    acc0 = simd.conditional_conj_mul_add_e(
+                        conj,
+                        a_body1.get_unchecked(0).get(),
+                        b_body1.get_unchecked(0).get(),
+                        acc0,
+                    );
+                    acc1 = simd.conditional_conj_mul_add_e(
+                        conj,
+                        a_tail.read_or(zero),
+                        b_tail.read_or(zero),
+                        acc1,
+                    );
+                }
+                2 => {
+                    acc0 = simd.conditional_conj_mul_add_e(
+                        conj,
+                        a_body1.get_unchecked(0).get(),
+                        b_body1.get_unchecked(0).get(),
+                        acc0,
+                    );
+                    acc1 = simd.conditional_conj_mul_add_e(
+                        conj,
+                        a_body1.get_unchecked(1).get(),
+                        b_body1.get_unchecked(1).get(),
+                        acc1,
+                    );
+                    acc2 = simd.conditional_conj_mul_add_e(
+                        conj,
+                        a_tail.read_or(zero),
+                        b_tail.read_or(zero),
+                        acc2,
+                    );
+                }
+                3 => {
+                    acc0 = simd.conditional_conj_mul_add_e(
+                        conj,
+                        a_body1.get_unchecked(0).get(),
+                        b_body1.get_unchecked(0).get(),
+                        acc0,
+                    );
+                    acc1 = simd.conditional_conj_mul_add_e(
+                        conj,
+                        a_body1.get_unchecked(1).get(),
+                        b_body1.get_unchecked(1).get(),
+                        acc1,
+                    );
+                    acc2 = simd.conditional_conj_mul_add_e(
+                        conj,
+                        a_body1.get_unchecked(2).get(),
+                        b_body1.get_unchecked(2).get(),
+                        acc2,
+                    );
+                    acc3 = simd.conditional_conj_mul_add_e(
+                        conj,
+                        a_tail.read_or(zero),
+                        b_tail.read_or(zero),
+                        acc3,
+                    );
+                }
+                4 => {
+                    acc0 = simd.conditional_conj_mul_add_e(
+                        conj,
+                        a_body1.get_unchecked(0).get(),
+                        b_body1.get_unchecked(0).get(),
+                        acc0,
+                    );
+                    acc1 = simd.conditional_conj_mul_add_e(
+                        conj,
+                        a_body1.get_unchecked(1).get(),
+                        b_body1.get_unchecked(1).get(),
+                        acc1,
+                    );
+                    acc2 = simd.conditional_conj_mul_add_e(
+                        conj,
+                        a_body1.get_unchecked(2).get(),
+                        b_body1.get_unchecked(2).get(),
+                        acc2,
+                    );
+                    acc3 = simd.conditional_conj_mul_add_e(
+                        conj,
+                        a_body1.get_unchecked(3).get(),
[…]

[…]
