### 2024-05-24 3872f42
Adds a derive(Clone)
