### 2024-05-24 85223fa
Adds Clone to SparseRowMat
