+++ src/sparse/linalg/triangular_solve.rs
@@ -1,5 +1,3 @@
-// FIXME: support duplicate entries
-
 use crate::{assert, sparse::SparseColMatRef, utils::slice::*, Conj, Index, MatMut, Parallelism};
 use core::iter::zip;
 use faer_entity::ComplexField;
@@ -45,12 +43,17 @@ pub fn solve_lower_triangular_in_place<I: Index, E: ComplexField>(
                             let xj0 = x.read(j, k0).faer_mul(d);
                             x.write(j, k0, xj0);
 
-                            for (i, lij) in zip(
-                                l.row_indices_of_col(j),
-                                slice_group(l.values_of_col(j)).into_ref_iter(),
-                            )
-                            .skip(1)
-                            {
+                            let row_ind = l.row_indices_of_col_raw(j);
+                            let len = row_ind.len();
+                            for (i, lij) in unsafe {
+                                zip(
+                                    row_ind.get_unchecked(1..),
+                                    slice_group(l.values_of_col(j))
+                                        .subslice_unchecked(1..len)
+                                        .into_ref_iter(),
+                                )
+                            } {
+                                let i = i.zx();
                                 let lij = lij.read();
                                 let lij = if conj == Conj::Yes {
                                     lij.faer_conj()
@@ -73,12 +76,17 @@ pub fn solve_lower_triangular_in_place<I: Index, E: ComplexField>(
                             let xj1 = x.read(j, k1).faer_mul(d);
                             x.write(j, k1, xj1);
 
-                            for (i, lij) in zip(
-                                l.row_indices_of_col(j),
-                                slice_group(l.values_of_col(j)).into_ref_iter(),
-                            )
-                            .skip(1)
-                            {
+                            let row_ind = l.row_indices_of_col_raw(j);
+                            let len = row_ind.len();
+                            for (i, lij) in unsafe {
+                                zip(
+                                    row_ind.get_unchecked(1..),
+                                    slice_group(l.values_of_col(j))
+                                        .subslice_unchecked(1..len)
+                                        .into_ref_iter(),
+                                )
+                            } {
+                                let i = i.zx();
                                 let lij = lij.read();
                                 let lij = if conj == Conj::Yes {
                                     lij.faer_conj()
@@ -105,12 +113,17 @@ pub fn solve_lower_triangular_in_place<I: Index, E: ComplexField>(
                             let xj2 = x.read(j, k2).faer_mul(d);
                             x.write(j, k2, xj2);
 
-                            for (i, lij) in zip(
-                                l.row_indices_of_col(j),
-                                slice_group(l.values_of_col(j)).into_ref_iter(),
-                            )
-                            .skip(1)
-                            {
+                            let row_ind = l.row_indices_of_col_raw(j);
+                            let len = row_ind.len();
+                            for (i, lij) in unsafe {
+                                zip(
+                                    row_ind.get_unchecked(1..),
+                                    slice_group(l.values_of_col(j))
+                                        .subslice_unchecked(1..len)
+                                        .into_ref_iter(),
+                                )
+                            } {
+                                let i = i.zx();
                                 let lij = lij.read();
                                 let lij = if conj == Conj::Yes {
                                     lij.faer_conj()
@@ -141,12 +154,17 @@ pub fn solve_lower_triangular_in_place<I: Index, E: ComplexField>(
                             let xj3 = x.read(j, k3).faer_mul(d);
                             x.write(j, k3, xj3);
 
-                            for (i, lij) in zip(
-                                l.row_indices_of_col(j),
-                                slice_group(l.values_of_col(j)).into_ref_iter(),
-                            )
-                            .skip(1)
-                            {
+                            let row_ind = l.row_indices_of_col_raw(j);
+                            let len = row_ind.len();
+                            for (i, lij) in unsafe {
[…]
