+++ benches/qr.rs
@@ -51,7 +51,10 @@ where
 
 fn main() -> std::io::Result<()> {
     let mut bench = Bench::new(BenchConfig::from_args()?);
-    let args = [1, 2, 4, 8, 16, 24, 32, 48, 64, 128].map(PlotArg);
+    let args = [
+        1, 2, 4, 8, 16, 24, 32, 48, 64, 128, 256, 512, 1024, 2048, 4096,
+    ]
+    .map(PlotArg);
     bench.register_many(list![faer::<f32>, nalgebra::<f32>], args);
     bench.register_many(list![faer::<f64>, nalgebra::<f64>], args);
     bench.run()?;
