+++ Cargo.toml
@@ -21,7 +21,7 @@ dyn-stack = { version = "0.10.0", default-features = false }
 equator = "0.2.2"
 faer-entity = { version ="0.18.0", default-features = false, path = "./faer-entity" }
 
-gemm = { version = "0.17.1", default-features = false }
+gemm = { version = "0.18.0", default-features = false }
 nano-gemm = { version = "0.1.2", default-features = false }
 num-complex = { version = "0.4.5", default-features = false }
 num-traits = { version = "0.2.18", default-features = false }
@@ -36,7 +36,6 @@ npyz = { version = "0.8", optional = true }
 rand = { version = "0.8.5", default-features = false, optional = true }
 rand_distr = { version = "0.4.3", default-features = false, optional = true }
 libm = "0.2.8"
-likely_stable = "0.1.2"
 
 [features]
 default = ["std", "rayon", "serde", "rand", "npy"]
@@ -93,10 +92,6 @@ harness = false
 
 [[bench]]
 name = "meanvar"
-harness = false
-
-[[bench]]
-name = "sparse_solve"
 harness = false
 
 [[bench]]
