### 2024-05-27 876eebc
fix determinism in most reductions
