+++ src/linalg/matmul/mod.rs
@@ -68,14 +68,34 @@ pub mod inner_prod {
             a_body2.into_ref_iter().map(RefGroup::unzip),
             b_body2.into_ref_iter().map(RefGroup::unzip),
         ) {
-            acc0 = simd.conditional_conj_mul_add_e(conj, a0.read_or(zero), b0.read_or(zero), acc0);
             acc1 = simd.conditional_conj_mul_add_e(conj, a1.read_or(zero), b1.read_or(zero), acc1);
+            acc0 = simd.conditional_conj_mul_add_e(conj, a0.read_or(zero), b0.read_or(zero), acc0);
         }
-        for (a, b) in zip(a_body1.into_ref_iter(), b_body1.into_ref_iter()) {
-            acc0 = simd.conditional_conj_mul_add_e(conj, a.read_or(zero), b.read_or(zero), acc0);
+        match a_body1.len() {
+            0 => {
+                acc1 = simd.conditional_conj_mul_add_e(
+                    conj,
+                    a_tail.read_or(zero),
+                    b_tail.read_or(zero),
+                    acc1,
+                );
+            }
+            1 => {
+                acc1 = simd.conditional_conj_mul_add_e(
+                    conj,
+                    a_body1.get_unchecked(0).get(),
+                    b_body1.get_unchecked(0).get(),
+                    acc1,
+                );
+                acc0 = simd.conditional_conj_mul_add_e(
+                    conj,
+                    a_tail.read_or(zero),
+                    b_tail.read_or(zero),
+                    acc0,
+                );
+            }
+            _ => unreachable!(),
         }
-        acc0 =
-            simd.conditional_conj_mul_add_e(conj, a_tail.read_or(zero), b_tail.read_or(zero), acc0);
         simd.add(acc0, acc1)
     }
 
@@ -101,17 +121,86 @@ pub mod inner_prod {
             a_body4.into_ref_iter().map(RefGroup::unzip),
             b_body4.into_ref_iter().map(RefGroup::unzip),
         ) {
-            acc0 = simd.conditional_conj_mul_add_e(conj, a0.read_or(zero), b0.read_or(zero), acc0);
-            acc1 = simd.conditional_conj_mul_add_e(conj, a1.read_or(zero), b1.read_or(zero), acc1);
-            acc2 = simd.conditional_conj_mul_add_e(conj, a2.read_or(zero), b2.read_or(zero), acc2);
-            acc3 = simd.conditional_conj_mul_add_e(conj, a3.read_or(zero), b3.read_or(zero), acc3);
+            acc1 = simd.conditional_conj_mul_add_e(conj, a0.read_or(zero), b0.read_or(zero), acc1);
+            acc2 = simd.conditional_conj_mul_add_e(conj, a1.read_or(zero), b1.read_or(zero), acc2);
+            acc3 = simd.conditional_conj_mul_add_e(conj, a2.read_or(zero), b2.read_or(zero), acc3);
+            acc0 = simd.conditional_conj_mul_add_e(conj, a3.read_or(zero), b3.read_or(zero), acc0);
         }
-        for (a, b) in zip(a_body1.into_ref_iter(), b_body1.into_ref_iter()) {
-            acc0 = simd.conditional_conj_mul_add_e(conj, a.read_or(zero), b.read_or(zero), acc0);
+        unsafe {
+            match a_body1.len() {
+                0 => {
+                    acc1 = simd.conditional_conj_mul_add_e(
+                        conj,
+                        a_tail.read_or(zero),
+                        b_tail.read_or(zero),
+                        acc1,
+                    );
+                }
+                1 => {
+                    acc1 = simd.conditional_conj_mul_add_e(
+                        conj,
+                        a_body1.get_unchecked(0).get(),
+                        b_body1.get_unchecked(0).get(),
+                        acc1,
[…]
