### 2024-06-11 #6099 @mhoemmen
Jed met with me today to go through the library and help me get started downloading it.  I've had some "editor's block" about starting the review.  My apologies to the author, who has waited far too long on it.  I didn't feel ready because I didn't understand Rust well enough, but Jed helped me understand a tiny bit so that I wouldn't be completely lost.  He also emphasized that I should focus on the library's functionality independent of whether the library is idiomatic Rust code.
