### 2024-06-11 #6099 @mhoemmen
I'll add some remarks here from my conversation with Jed today.

Jed explained that Sarah wrote a clean-room reimplementation from the algorithms, rather than porting existing code to Rust or wrapping existing non-Rust libraries.  This is an extraordinary effort.  From an algorithm development perspective, it was comparable to that of developing EISPACK _and_ LINPACK from the Handbook.  From a software library development perspective, it was actually more effort, because the idioms for managing views of subsets of multidimensional arrays in languages like Fortran or C++ do not carry over naturally to Rust.  This means that Sarah couldn't just port (say) LAPACK, because Rust doesn't come with a way to say things like "create a view of rows 10-15 and columns 20-29 of this matrix."  As a result, Sarah had to develop her own data structures, such as a matrix container (a C++ word meaning something like "data structure owning an allocation") and mutable and immutable "reference" (C++ might say "subview") types.  Along with this, Sarah had to come up with new idioms, like what an "in-place factorization" means.  (Instead of "operate in place on a view of the matrix" as in LAPACK, it means "consume this matrix and produce a new matrix that possibly reuses the storage of the consumed matrix.")

Thanks to Jed's help, I was able to build (is that even the right word?) and run the library's tests.  The experience was completely without hassle.  The hardest part by far was convincing my Linux system to install Rust, and that wasn't even so hard.  The test output was extensive and gave a tour of the library's features.  It also showed that one test (the matrix multiply test) didn't run, because it would have taken too long on my system.  Presumably it's possible to enable that test by hand.
