### 2024-06-11 #6099 @mhoemmen
Regarding

> Functionality: Have the functional claims of the software been confirmed?

I have to learn how to navigate the code repository in order to do this, because I don't know a priori how to tell where the tests live.  There's no "test" directory, for example.  `git grep test` has led me to src/lib.rs, which appears to be the main entry point to the library.  I see extensive human-readable documentation comments and some code constructs like `#[cfg(test)]` and `#[test]`.  I'm guessing that the latter are embedded tests.
