### 2024-06-11 #6099 @jedbrown
Yeah, [this](https://doc.rust-lang.org/book/ch11-01-writing-tests.html) is the pattern used for unit tests. It causes this code to be compiled out for regular use (in which case `cfg(test)` is false), but automatically put it in the test harness when running tests.
```rs
#[cfg(test)]
mod tests {
    #[test]
    fn test_name() {
        // ...
    }
}
```
You can also see how things are tested in CI (which includes running the test suite with code coverage) in [`.github/workflows/run-tests.yml`](https://github.com/sarah-ek/faer-rs/blob/main/.github/workflows/run-tests.yml). Note that there is also some [static analysis/linting](https://github.com/sarah-ek/faer-rs/blob/main/.github/workflows/code-quality.yml) in CI.
