### 2024-06-11 #6099 @mhoemmen
src/lib.rs documentation begins with matrix creation, slicing, and overloaded operators.  The functionality looks familiar both to Matlab users and to users of comparable C++ libraries like Eigen.

I do wish the documentation had explained what this `mat!` thing is, though perhaps "`foo!` creates a `Foo`" is a Rust idiom with which I am unfamiliar.  Jed went into some detail about how this is a macro that avoids creating a temporary to hold the literal input.  I've written something just a bit comparable in C++ for the `mdarray` proposal (P1684), namely a nested `initializer_list<initializer_list<value_type>>` constructor.

Expressions like the following in the documentation appear to act like C++ expression templates, in that they are nonowning views of existing matrices that represent matrix arithmetic expressions.  Rust doesn't have the dangling references problem that classical C++ expression templates have.  Rust's syntax also makes it more clear (contrasted with the "`auto` return type problem" that e.g., Eigen has) that the type of something like `add` below is a reference, not a new allocation.

```
let add = &a + &b;
let sub = &a - &b;
let scale = scale(3.0) * &a;
let mul = &a * b.transpose();
```

I'm curious about evaluation of non-elementwise expressions like `&a * b.transpose()`.  Is that an elementwise multiplication or the usual matrix multiply?  The documentation says "`*` for either scalar or matrix multiplication depending on the types of the operands."  If `*` can represent matrix-matrix multiplication, then what happens with expressions like `&a * &b * &c`?  Does the implementation attempt to solve the matrix chain multiplication problem?  Are these expressions always lazily evaluated?  Do they need to allocate intermediate temporary storage, and if so, do they attempt to reuse such storage for the various products in the chain?

Regarding the indexing expression `let a00 = a[(0, 0)];` in the documentation, Jed explained that Rust does not permit array indexing expressions to have more than one argument.  I was happy to be involved in C++23's removal of that restriction, so that `mdspan` and `mdarray` can say e.g., `a[0, 0]`.  This is a language restriction, not a library restriction, though.
