### 2024-06-11 #6099 @mhoemmen
For the factorizations listed in src/lib.rs, I'll separate my discussion into "SVD or eigendecomposition" and "everything else" (the "one-sided" factorizations).

Regarding "everything else," I'm pleased to see some harder-to-implement but more interesting factorizations, like Bunch-Kaufman decomposition, QR with column pivoting (useful for applications like hierarchical matrices), and LU with full pivoting.

Regarding "SVD or eigendecomposition," it's a bold choice for a library to include those.  These are much trickier to implement and test.  Even LAPACK's eigendecompositions have evolved quite a bit in the last 15-20 years, including entirely new algorithms (e.g., MRRR) in the symmetric / Hermitian case.  It's possible that LAPACK is the only source for well-maintained, well-tested implementations of these algorithms.  It looks like faer has functionality comparable to LAPACK's "standard" drivers, either `_geev`, `_syev`, or `_heev` (all of which do QR iteration).  Considering how difficult these algorithms are to implement and test, it's reasonable for a single-developer effort not to cover the full space of problems (e.g., generalized eigenproblem) and algorithms.
