### 2024-06-12 #6099 @mhoemmen
I've been going through the tests.  It's great to see both the usual random matrix tests, and special cases, including matrices of all NaNs.

Something that I miss, particularly for the SVD and eigendecompositions, is a discussion of test provenance.  The experience of the QR algorithm is "we had to do these funny things to avoid known bad cases," so specific test cases matter a lot.  For increasing test coverage, it may pay to go through the LAPACK test suite and the LAPACK Working Notes (LAWNs).
