### 2024-08-15 #6099 @sarah-ek
once again, sorry for the delay. i've been busy dealing with immigration and paperwork hurdles and it hasn't left me with much time or energy to work on faer as much as i would like

i updated the paper taking into consideration the suggestions and remarks from @FCLC, though im not sure how much detail i should go into, considering the paper is already on the upper end of the suggested word count

@mhoemmen the svd, evd and bunch kaufman decomposition (generalized qz/evd and rook pivoted bunch kaufman coming soon:tm: ) are mostly inspired from existing implementations in reference lapack, [tlapack](https://github.com/tlapack/tlapack/) and eigen. im not confident enough to try implementing stuff the multishift qr algorithm on my own :sweat_smile:
i try to keep up to date with newer advancements in lapack, and fix bugs and edge cases that are reported by users. the library's still in its infancy so there have been a few of those so far. i plan to go through the code once again and take a closer look at differences from the original implementation, and fix any bugs that are due to that

simpler decompositions: cholesky, lu, qr (dense + sparse) are mostly my own implementations of published papers (sparse cholesky supernode relaxation is additionally inspired by the one in [catamari](https://gitlab.com/hodge_star/catamari)).
