### 2024-10-02 #6099 @jedbrown
Sorry about my delay. I was unclear if it was considered ready, though I checked with Sarah and it looks good. I submitted a PR with some bib/copy edits and a couple questions. Once resolved, we'll be ready to move forward as far as I'm concerned.
