### 2024-10-10 #6099 @jedbrown
I think the updates (with stronger static guarantees) are a novel and interesting contribution, and am fine with pausing this for a few weeks so you can include it. When ready, if you can point to release notes or otherwise share highlights here, I'd like reviewers to be aware of what changed.
