### 2025-01-14 #6099 @jedbrown
Yay. I saw the release and figured I'd wait a couple days before pinging about the paper. Great to hear it's being wrapped up.
