### 2025-06-11 #6099 @danielskatz
👋 @sarah-quinones - Have you have any success in being able to work on this? The JOSS review system really doesn't work if the author isn't able to respond to reviewers and make changes based on their reviews. Given that this has been in review for over 18 months, I think we might need to withdraw it if you can't work on it.
