### 2025-06-15 #6099 @mhoemmen
This is excellent work -- I hope it sees publication in some form!
