### 2025-10-31 #6099 @sarah-quinones
sorry, i got a bit distracted by the new job. i can probably do it tomorrow morning. the paper is more or less ready if i remember correctly? if not it might take a bit longer but it should still be good by monday morning at the latest
