### 2026-03-12 #6099 @danielskatz
👋 @sarah-quinones - I've proofread the paper, and suggest the changes in https://github.com/sarah-quinones/faer-rs/pull/263 - please check to make sure I haven't changed anything in a misleading way.

Also, I noticed two other possible issues:

1. Did you want to change your name in the paper.md file, or leave it as originally submitted?
2. Did you want to add any acknowledgements?
