### 2026-03-18 #6099 @sarah-quinones
i forgot to mention, but the project is now hosted at https://codeberg.org/sarah-quinones/faer/

the github repo is a mirror of that. i should change the name to "Sarah Q. El Kazdadi"

im not sure how to move a pr from github to codeberg, but i'm happy to merge the changes
