### 2026-05-02 #6099 @danielskatz
@sarah-quinones - I found one more thing in the bib that needed to be changed, and then also made a change for consistency, in https://codeberg.org/sarah-quinones/faer/pulls/289 - please merge this
